### 2026-01-20 #9745 @dguzh
We just noticed the January 2026 JOSS post describing updated scope and requirements. Our paper was submitted in November 2025 and was written under the previous guidelines. We would be grateful for advice on how to proceed in light of the new requirements.

Thank you in advance.
