### 2026-02-15 #9745 @arfon
## Update regarding JOSS scope and paper format changes

### What this means for your submission

1. State of the Field (~200 words): A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear "build vs. contribute" justification explaining your unique scholarly contribution and why existing alternatives are insufficient.a

For full details, please see:
* Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
* An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html
