### 2026-03-07 a2e854b
update paper for new JOSS requirements
