+++ paper/paper.bib
@@ -1,6 +1,14 @@
+@misc{honnibal2020,
+  author = {Honnibal, Matthew and Montani, Ines and Van Landeghem, Sofie and Boyd, Adriane},
+  title = {{spaCy: Industrial-strength Natural Language Processing in Python}},
+  year = {2020},
+  publisher = {Zenodo},
+  doi = {10.5281/zenodo.1212303}
+}
+
 @article{purves2018,
   author = {Purves, Ross S. and Clough, Paul and Jones, Christopher B. and Hall, Mark H. and Murdock, Vanessa},
-  title = {Geographic Information Retrieval: Progress and Challenges in Spatial Search of Text},
+  title = {{Geographic Information Retrieval: Progress and Challenges in Spatial Search of Text}},
   journal = {Foundations and Trends in Information Retrieval},
   volume = {12},
   number = {2-3},
@@ -11,7 +19,7 @@
 
 @inproceedings{teitler2008,
   author = {Teitler, Benjamin E. and Lieberman, Michael D. and Panozzo, Daniele and Sankaranarayanan, Jagan and Samet, Hanan and Sperling, Jon},
-  title = {NewsStand: a new view on news},
+  title = {{NewsStand: a new view on news}},
   booktitle = {Proceedings of the 16th ACM SIGSPATIAL International Conference on Advances in Geographic Information Systems},
   year = {2008},
   doi = {10.1145/1463434.1463458}
@@ -19,7 +27,7 @@
 
 @article{grover2010,
   author = {Grover, Claire and Tobin, Richard and Byrne, Kate and Woollard, Matthew and Reid, James and Dunn, Stuart and Ball, Julian},
-  title = {Use of the Edinburgh geoparser for georeferencing digitized historical collections},
+  title = {{Use of the Edinburgh geoparser for georeferencing digitized historical collections}},
   journal = {Philosophical Transactions of the Royal Society A: Mathematical, Physical and Engineering Sciences},
   volume = {368},
   number = {1925},
@@ -30,7 +38,7 @@
 
 @article{chesnokova2019,
   author = {Chesnokova, Olga and Taylor, Joanna E. and Gregory, Ian N. and Purves, Ross S.},
-  title = {Hearing the silence: finding the middle ground in the spatial humanities? Extracting and comparing perceived silence and tranquillity in the English Lake District},
+  title = {{Hearing the silence: finding the middle ground in the spatial humanities? Extracting and comparing perceived silence and tranquillity in the English Lake District}},
   journal = {International Journal of Geographical Information Science},
   volume = {33},
   number = {12},
@@ -47,7 +55,7 @@
 
 @article{karimzadeh2019,
   author = {Karimzadeh, Morteza and Pezanowski, Scott and MacEachren, Alan M. and Wallgrün, Jan O.},
-  title = {GeoTxt: A scalable geoparsing system for unstructured text geolocation},
+  title = {{GeoTxt: A scalable geoparsing system for unstructured text geolocation}},
   journal = {Transactions in GIS},
   volume = {23},
   number = {1},
@@ -58,7 +66,7 @@
 
 @article{halterman2017,
   author = {Halterman, Andrew},
-  title = {Mordecai: Full Text Geoparsing and Event Geocoding},
+  title = {{Mordecai: Full Text Geoparsing and Event Geocoding}},
   journal = {Journal of Open Source Software},
   volume = {2},
   number = {9},
@@ -69,7 +77,7 @@
 
 @misc{halterman2023,
   author = {Halterman, Andrew},
-  title = {Mordecai 3: A Neural Geoparser and Event Geocoder},
+  title = {{Mordecai 3: A Neural Geoparser and Event Geocoder}},
   year = {2023},
[…]
