+++ paper/paper.md
@@ -19,7 +19,7 @@ affiliations:
   - name: Department of Geography, University of Zurich, Switzerland
     index: 1
     ror: 02crff812
-date: 28 November 2025
+date: 7 March 2026
 bibliography: paper.bib
 ---
 
@@ -29,25 +29,37 @@ The Irchel Geoparser is an open-source Python library for recognising and resolv
 
 # Statement of Need
 
-Geoparsing is an essential step in linking text to geographical locations [@purves2018], enabling diverse applications including mapping news coverage, exploring historical documents, or analysing discourse across regions [@teitler2008; @grover2010; @chesnokova2019]. Although several geoparsers already exist, they have limitations that restrict their practical applicability. Systems such as the Edinburgh Geoparser [@grover2010], CLAVIN [@berico2012], GeoTxt [@karimzadeh2019], and the Mordecai family [@halterman2017; @halterman2023] have contributed significantly to the field, but they generally require substantial effort for deployment and offer only limited flexibility in the customisation of the geoparsing process. Adapting these tools to a specific task, whether by incorporating new models for languages other than English, using specialised gazetteers or aligning the pipeline to a domain typically entails maintaining a separate fork of the software.
+Geoparsing is an essential step in linking text to geographical locations [@purves2018], enabling diverse applications including mapping news coverage, exploring historical documents, or analysing discourse across regions [@teitler2008; @grover2010; @chesnokova2019]. Typically, geoparsing is a two-stage process involving firstly toponym recognition and subsequently toponym resolution. Toponyms, or place names, are first identified in text, a special case of the more general problem of named entity recognition. Toponym recognition must differentiate between instances of words being used in other contexts - for example to describe generic objects (as for instance in the case of bath (a place to wash oneself and an English city)) or to name people and firms (for example (George) Washington or Zurich (Insurance)). Toponym resolution resolves candidate place names to a unique identifier, dealing with the problem of multiple instances of the same place name (for example London, UK vs. London, Ontario, Canada).
+
+The Irchel Geoparser was designed to provide a modular, out of the box functional Python library which could easily be incorporated into text processing chains using default settings. As such, it is well suited to teaching applications introducing for example graduate students to the geoparsing process. At the same time, because of its modular design, the entire processing chain is configurable to particular use cases - for example by swapping gazetteers to use those more appropriate to historical documents or retraining language models underlying toponym resolution modules on a specific corpus. Thus, the Irchel Geoparser is also well suited for use in research. Finally, we also aimed to provide a well documented Python library with an open licence allowing not only experimentation but also reproducibility of results.
+
+
+# State of the Field
+
+Although several geoparsers already exist, they have limitations that restrict their practical applicability. Systems such as the Edinburgh Geoparser [@grover2010], CLAVIN [@berico2012], GeoTxt [@karimzadeh2019], and the Mordecai family [@halterman2017; @halterman2023] have contributed significantly to the field, but they generally require substantial effort for deployment and offer only limited flexibility in the customisation of the geoparsing process. Adapting these tools to a specific task, whether by incorporating new models for languages other than English, using specialised gazetteers or aligning the pipeline to a domain typically involves separate forks of the software, where this is possible, and these are often not well documented.
 
[…]
