### 2026-03-07 #9745 @dguzh
Thank you for your reply.

We have updated the paper to align with the new JOSS requirements and added the requested sections.
