### 2026-03-26 #9745 @samhforbes
In fact, @jGaboardi might this be in your wheelhouse?
