### 2026-03-27 #9745 @jGaboardi
> Thank you for your reply.
>
> We have updated the paper to align with the new JOSS requirements and added the requested sections.
>
> [@editorialbot](https://github.com/editorialbot) generate pdf

@dguzh – EditorialBot commands have to be in their own cell.
