### 2026-03-27 #9745 @jGaboardi
Hi @dguzh thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding reviewers. I'll comment over time as I have a chance to go through them:

- [x] [Required sections](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) in paper
- [x] No Conflicts of Interest

In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers. For reviewers, please suggest 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their GitHub handles are most useful to receive but *please don't use "@" to reference them* since it will prematurely ping them.
