### 2026-03-27 #9745 @jGaboardi
> **License info**:
>
> 🟡 License found: `Other` ([Check here](https://opensource.org/licenses) for OSI approval)

Not sure why this is being flagged as `Other`. The package is clearly [licensed under `MIT`](https://github.com/dguzh/geoparser/blob/main/LICENSE). Perhaps due the [`THIRD_PARTY_LICENSES`](https://github.com/dguzh/geoparser/blob/main/THIRD_PARTY_LICENSES) file?

@samhforbes - Out of curiosity, have you ever dealt with this?
