### 2026-04-01 #9745 @jGaboardi
👋 @caimeng2, @KennethEnevoldsen, @cforgaci - would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
