### 2026-04-02 #9745 @jGaboardi
@cforgaci -- Thanks to quick reply! I'll go ahead and assign you. And no problem and not finished before the end of April. We ask for [4-6 weeks for initial review](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#:~:text=We%20ask%20reviewers%20to%20complete%20their%20reviews%20in%204%2D6%20weeks).

@alex-ip -- We'll go ahead and start the review with @SamHames and @cforgaci. If you are still interested we can add you later!
