### 2026-04-02 #10365 @editorialbot
[REVIEW]: The Irchel Geoparser: A Modular Python Library for Toponym Recognition and Resolution
**Submitting author:** @dguzh (<a href="http://orcid.org/0009-0003-8449-2603">Diego Gomes</a>)
**Repository:** https://github.com/dguzh/geoparser
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.3.1
