### 2026-04-02 #10365 @jGaboardi
👋🏼 @SamHames & @cforgaci –  this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#10365`, so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
