### 2026-04-27 dguzh/geoparser#91 @SamHames
Documentation issue with first example
Note that this is part of my [review for JOSS](https://github.com/openjournals/joss-reviews/issues/10365).

When first trying to get everything working I tried [running the example from the docs](https://docs.geoparser.app/en/latest/quickstart.html), but I'd missed the installation step about importing the geonames dataset. This resulted in no errors and also no results when running the example.

It would make more sense to raise an error if geonames data hasn't been imported yet.

I think it would also be less surprising to have to explicitly specify the default recogniser and resolver - explicitly providing the recogniser and resolver might avoid some of the configuration issues that happen when people forget to specify one and not the other.

Note also that the Python REPL docstring for Geoparser gives an opaque reference to an object which is a little tricky to make sense of what the defaults actually are.

```class Geoparser(builtins.object)
 |  Geoparser(recognizer: Optional[geoparser.modules.recognizers.base.Recognizer] = <object object at 0x7fc30b609d20>, resolver: Optional[geoparser.modules.resolvers.base.Resolver] = <object object at 0x7fc30b609d20>, spacy_model: Optional[str] = None, transformer_model: Optional[str] = None)
```
