### 2026-04-27 dguzh/geoparser#92 @SamHames
Packaging: no source distribution and missing manylinux wheels
Note that this is part of my [review for JOSS](https://github.com/openjournals/joss-reviews/issues/10365).

There are two major problems with the Python packaging - both of which come from the need to package and include spatialite.

1. There is no working source distribution: the source distribution should be fully installable on any system assuming standard compiler availability (and may need documentation to indicate assumed available packages). While I don't expect this project to test and maintain systems beyond windows/linux/mac, not having a properly packaged source distribution means that it will be unlikely to keep this working into the future or on other operating systems or processor architectures.
2. Packaging compiled binaries for linux is complex, and this project is not following best practice. Currently the version of spatialite shipped will only work on linux systems contemporary with or newer than wherever you happened to compile it (try running in Ubuntu 22.04 for example).  If you want to ship compiled objects, use projects like [CIBuildWheel](https://cibuildwheel.pypa.io/en/stable/) to produce appropriately tagged platform specific [manylinux wheels](https://github.com/pypa/manylinux)
