### 2026-04-27 dguzh/geoparser#93 @SamHames
Docs: Small end-to-end example with custom gazetteer
Note that this is part of my [review for JOSS](https://github.com/openjournals/joss-reviews/issues/10365).

I'm raising this from following along with @alex-ip as he was trying to import a small dataset of Australian placenames, he'll raise a separate issue or issues about what he ran into with the import process while doing that.

A central aim of this package is to encourage modularity and customisation, which necessarily requires that people be able to bring in their own placename data. However there's no end-to-end worked example of doing this for a small custom dataset that's easy to follow along with.

The current examples:

1. assume a significant amount of knowledge of geospatial data processing and preparation
2. don't clearly document what the end state of all of this processing should be
3. don't give guidance for common problems that are likely to pop up, or the choices that need to be made
4. only present small parts of the overall problem in isolation (the [complete example](https://docs.geoparser.app/en/latest/guides/gazetteers.html#complete-example) isn't even with a real dataset)

Without a worked example I wouldn't have anyway of knowing if problems were with my data or the process I'm using, or if there's a problem on the geoparser side that could be fixed. This is further exacerbated by using a custom YAML configuration approach to loading which is quite divorced from the original source data.
