### 2026-04-27 dguzh/geoparser#94 @SamHames
Architecture is neither here nor there in terms of modularity and implementation
Note that this is part of my [review for JOSS](https://github.com/openjournals/joss-reviews/issues/10365).

The aims of this project are modularity and configurability, but the default installation combines only partial modularity with extremely heavyweight default installations. For example, I needed to download gigabytes of torch libraries and geodata, and recompile my Python with sqlite extension loading support before I could even try out a basic example from the docs.

I recommend you consider:

1. Committing to your aim of modularity, by making this core package about the high level architecture of distinct concerns with pluggable implementations.  The core components of this architecture are:
    - Recognisers
    - Resolvers
    - Projects (and the associated workflow support for experimentation which is common to any implementations of the above two components)
    - The high level interfaces between them
2. Demonstrating that modularity by moving the current default heavy implementations out to their own optional packages and having only a very lightweight (even toy) implementation available in this package. This creates more potential for innovation by allowing new components to be packaged and shipped separately without being weighed down by a large default in the core.
3. Not building your spatial data processing architecture around spatialite specifically - consider the choice of any particular package or database as an implementation detail of a resolver component (for example, maybe geopandas will be fine for small geographically focused projects, maybe you want postgis for something enormous). There are multiple options for geospatial data processing and reliably trying to ship spatialite looks like it is going to cause you maintenance headaches in the long term.
