### 2026-04-27 dguzh/geoparser#95 @SamHames
How do I join cached document text back to where it came from?
Note that this is part of my [review for JOSS](https://github.com/openjournals/joss-reviews/issues/10365).

If I'm looking to resolve toponyms I'm typically doing this in the context of materials that have significant contextual information, not just text: think provenance metadata, document identifiers, author information and more. Typically I want identified toponyms as augmentations in the context of the original materials.

It would be nice if I could either assign an identifier on adding documents to the system which I can then join back later, or (less work but probably easier to implement) when a document is added, returning the internally assigned identifier so I can maintain my own lookup table.

Otherwise the project interface is quite limiting: I'd have to reprocess everything just so I can keep track of the particular place this piece of text came from. I also couldn't do things like performance metric evaluations on particular subsets of the materials (for example older vs newer materials).
