### 2026-04-27 #10365 @SamHames
I think there are some good ideas in this project, and it certainly aims to address a relevant and important research problem. However, I don't think the implementation holds up to the promise of the documentation and the aims.

I ran into quite a few issues with installation and trying to get it running according to the documentation, and in conjunction with a colleague we struggled to get basic things working with datasets outside the examples provided. On top of that, it looks there are enough packaging related issues that I'm concerned about the long term maintainability of this package.

Note that I herded my geospatial colleague @alex-ip (thanks!) to also have a look at this, because on paper it seemed extremely relevant to a project he's been working on.

To me the most interesting part of this project is:

1. The high level architecture separating recognition and resolution into separate components with a specified interface.
2. The state management to support experimentation and different workflow runs over the same set of data, allowing persistence and comparison of results. This is the kind of thing that gets done in an adhoc manner all the time and having this built in can reduce the effort of systematically working on this.

I think these are either compromised by the actual implementation (1), or missing some key features to make this a project worth directing people to (2).

I've raised the following issues to track my concerns in the geoparser repository:

Major concerns:

- https://github.com/dguzh/geoparser/issues/92
- https://github.com/dguzh/geoparser/issues/93
- https://github.com/dguzh/geoparser/issues/94 (This ones a much bigger issue, and not something I'd expect to be fully addressed in the context of a typical review cycle)

Smaller issues:

- https://github.com/dguzh/geoparser/issues/91
- https://github.com/dguzh/geoparser/issues/95
