### 2026-05-08 dguzh/geoparser#100 @cforgaci
Licenses for SpatiaLite and Spellfix1 not specified in the README
This issue is part of my review in https://github.com/openjournals/joss-reviews/issues/10365.

While the README file specifies the licenses of the library and Python-based third-party licenses, it does not provide information about the licenses of SpatiaLite and Spellfix1 found in `geoparser\db\extensions\`, which are LGPL v2.1 and Public Domain, respectively. This should be added for a complete license specification.
