### 2026-05-30 8f4e455
Pin Black to 25.11.0 and align CI lint with pyproject.toml. Prevents psf/black@stable from installing Black 26.x and failing on code formatted with the project's locked dev dependency version.
