+++ .github/workflows/lint.yml
@@ -11,7 +11,12 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
       - uses: psf/black@stable
+        with:
+          use_pyproject: true
 
   isort:
     name: isort
