### 2026-05-30 206a992
Introduce a new YAML configuration file for a lightweight gazetteer comprising of geonames cities
