+++ geoparser/gazetteer/configs/geonames_cities.yaml
@@ -1,0 +1,645 @@
+name: geonames_cities
+sources:
+  - name: cities500
+    url: https://download.geonames.org/export/dump/cities500.zip
+    file: cities500.txt
+    type: tabular
+    separator: "\t"
+    attributes:
+      original:
+        - name: geonameid
+          type: INTEGER
+          index: true
+        - name: name
+          type: TEXT
+        - name: asciiname
+          type: TEXT
+        - name: alternatenames
+          type: TEXT
+        - name: latitude
+          type: REAL
+        - name: longitude
+          type: REAL
+        - name: feature_class
+          type: TEXT
+        - name: feature_code
+          type: TEXT
+        - name: country_code
+          type: TEXT
+          index: true
+        - name: cc2
+          type: TEXT
+        - name: admin1_code
+          type: TEXT
+        - name: admin2_code
+          type: TEXT
+        - name: admin3_code
+          type: TEXT
+        - name: admin4_code
+          type: TEXT
+        - name: population
+          type: INTEGER
+        - name: elevation
+          type: INTEGER
+        - name: dem
+          type: INTEGER
+        - name: timezone
+          type: TEXT
+        - name: modification_date
+          type: TEXT
+      derived:
+        - name: geometry
+          type: GEOMETRY
+          expression: "'POINT(' || longitude || ' ' || latitude || ')'"
+          index: true
+          srid: 4326
+        - name: feature_code_full
+          type: TEXT
+          expression: "feature_class || '.' || feature_code"
+          index: true
+        - name: admin1_code_full
+          type: TEXT
+          expression: "country_code || '.' || admin1_code"
+          index: true
+        - name: admin2_code_full
+          type: TEXT
+          expression: "country_code || '.' || admin1_code || '.' || admin2_code"
+          index: true
+        - name: name_no_parens
+          type: TEXT
+          expression: "CASE WHEN instr(name, '(') > 0 THEN trim(substr(name, 1, instr(name, '(') - 1)) ELSE name END"
+        - name: asciiname_no_parens
+          type: TEXT
+          expression: "CASE WHEN instr(asciiname, '(') > 0 THEN trim(substr(asciiname, 1, instr(asciiname, '(') - 1)) ELSE asciiname END"
+    view:
+      select:
+        - source: cities500
+          column: geonameid
+        - source: cities500
+          column: name
+        - source: cities500
+          column: name_no_parens
+        - source: cities500
+          column: asciiname
+        - source: cities500
+          column: asciiname_no_parens
+        - source: cities500
+          column: alternatenames
+        - source: cities500
+          column: latitude
+        - source: cities500
+          column: longitude
+        - source: cities500
+          column: feature_class
+        - source: cities500
+          column: feature_code
+        - source: featureCodes
+          column: name
+          alias: feature_name
+        - source: cities500
+          column: country_code
+        - source: countryInfo
+          column: Country
+          alias: country_name
+        - source: cities500
+          column: cc2
+        - source: cities500
+          column: admin1_code
+        - source: admin1CodesASCII
+          column: name
+          alias: admin1_name
+        - source: cities500
+          column: admin2_code
+        - source: admin2Codes
+          column: name
+          alias: admin2_name
+        - source: cities500
+          column: admin3_code
+        - source: cities500
+          column: admin4_code
+        - source: cities500
+          column: population
+        - source: cities500
+          column: elevation
+        - source: cities500
+          column: dem
+        - source: cities500
+          column: timezone
+        - source: cities500
+          column: modification_date
[…]
