### 2026-05-30 fe309b1
Update gazetteer name validation to allow hyphens and rename geonames_cities to geonames-cities
