### 2026-05-30 a0633f1
Add geonames-cities configuration to SentenceTransformerResolver attribute map
