### 2026-06-06 55fe2dc
Add join condition configurations and predicates to model.py and update enum values to be lowercase throughout
