### 2026-06-06 afae771
Update gazetteer YAML configurations to use lowercase types and update join conditions to use structured predicates
