+++ geoparser/gazetteer/configs/swissnames3d.yaml
@@ -7,38 +7,38 @@ sources:
     attributes:
       original:
         - name: UUID
-          type: TEXT
+          type: text
           index: true
         - name: OBJEKTART
-          type: TEXT
+          type: text
         - name: OBJEKTKLAS
-          type: TEXT
+          type: text
         - name: HOEHE
-          type: REAL
+          type: real
         - name: GEBAEUDENU
-          type: INTEGER
+          type: integer
         - name: NAME_UUID
-          type: TEXT
+          type: text
         - name: NAME
-          type: TEXT
+          type: text
         - name: STATUS
-          type: TEXT
+          type: text
         - name: SPRACHCODE
-          type: TEXT
+          type: text
         - name: NAMEN_TYP
-          type: TEXT
+          type: text
         - name: NAMENGRUPP
-          type: TEXT
+          type: text
         - name: geometry
-          type: GEOMETRY
+          type: geometry
           index: true
           srid: 2056
       derived:
         - name: NAME_NO_PARENS
-          type: TEXT
+          type: text
           expression: "CASE WHEN instr(NAME, '(') > 0 THEN trim(substr(NAME, 1, instr(NAME, '(') - 1)) ELSE NAME END"
         - name: NAME_NO_CODE
-          type: TEXT
+          type: text
           expression: "CASE WHEN length(NAME) > 3 AND substr(NAME, -3, 1) = ' ' AND substr(NAME, -2, 2) GLOB '[A-Z][A-Z]' THEN trim(substr(NAME, 1, length(NAME) - 3)) ELSE NAME END"
     view:
       select:
@@ -89,15 +89,27 @@ sources:
           column: KANTONSNUM
           alias: KANTONSNUM
       join:
-        - type: LEFT JOIN
+        - method: left join
           source: swissBOUNDARIES3D_HOHEITSGEBIET
-          condition: ST_Within(swissNAMES3D_PKT.geometry, swissBOUNDARIES3D_HOHEITSGEBIET.geometry)
-        - type: LEFT JOIN
+          condition:
+            type: spatial
+            predicate: within
+            left: swissNAMES3D_PKT.geometry
+            right: swissBOUNDARIES3D_HOHEITSGEBIET.geometry
+        - method: left join
           source: swissBOUNDARIES3D_BEZIRKSGEBIET
-          condition: swissBOUNDARIES3D_HOHEITSGEBIET.BEZIRKSNUM = swissBOUNDARIES3D_BEZIRKSGEBIET.BEZIRKSNUM
-        - type: LEFT JOIN
+          condition:
+            type: attribute
+            predicate: equals
+            left: swissBOUNDARIES3D_HOHEITSGEBIET.BEZIRKSNUM
+            right: swissBOUNDARIES3D_BEZIRKSGEBIET.BEZIRKSNUM
+        - method: left join
           source: swissBOUNDARIES3D_KANTONSGEBIET
-          condition: swissBOUNDARIES3D_BEZIRKSGEBIET.KANTONSNUM = swissBOUNDARIES3D_KANTONSGEBIET.KANTONSNUM
+          condition:
+            type: attribute
+            predicate: equals
+            left: swissBOUNDARIES3D_BEZIRKSGEBIET.KANTONSNUM
+            right: swissBOUNDARIES3D_KANTONSGEBIET.KANTONSNUM
     features:
       identifier:
         - column: UUID
@@ -113,36 +125,36 @@ sources:
     attributes:
       original:
         - name: UUID
-          type: TEXT
+          type: text
           index: true
         - name: OBJEKTART
-          type: TEXT
+          type: text
         - name: OBJEKTKLAS
-          type: TEXT
+          type: text
         - name: KUNSTBAUTE
-          type: TEXT
+          type: text
         - name: NAME_UUID
-          type: TEXT
+          type: text
[…]
