+++ geoparser/gazetteer/configs/geonames-cities.yaml
@@ -8,68 +8,68 @@ sources:
     attributes:
       original:
         - name: geonameid
-          type: INTEGER
+          type: integer
           index: true
         - name: name
-          type: TEXT
+          type: text
         - name: asciiname
-          type: TEXT
+          type: text
         - name: alternatenames
-          type: TEXT
+          type: text
         - name: latitude
-          type: REAL
+          type: real
         - name: longitude
-          type: REAL
+          type: real
         - name: feature_class
-          type: TEXT
+          type: text
         - name: feature_code
-          type: TEXT
+          type: text
         - name: country_code
-          type: TEXT
+          type: text
           index: true
         - name: cc2
-          type: TEXT
+          type: text
         - name: admin1_code
-          type: TEXT
+          type: text
         - name: admin2_code
-          type: TEXT
+          type: text
         - name: admin3_code
-          type: TEXT
+          type: text
         - name: admin4_code
-          type: TEXT
+          type: text
         - name: population
-          type: INTEGER
+          type: integer
         - name: elevation
-          type: INTEGER
+          type: integer
         - name: dem
-          type: INTEGER
+          type: integer
         - name: timezone
-          type: TEXT
+          type: text
         - name: modification_date
-          type: TEXT
+          type: text
       derived:
         - name: geometry
-          type: GEOMETRY
+          type: geometry
           expression: "'POINT(' || longitude || ' ' || latitude || ')'"
           index: true
           srid: 4326
         - name: feature_code_full
-          type: TEXT
+          type: text
           expression: "feature_class || '.' || feature_code"
           index: true
         - name: admin1_code_full
-          type: TEXT
+          type: text
           expression: "country_code || '.' || admin1_code"
           index: true
         - name: admin2_code_full
-          type: TEXT
+          type: text
           expression: "country_code || '.' || admin1_code || '.' || admin2_code"
           index: true
         - name: name_no_parens
-          type: TEXT
+          type: text
           expression: "CASE WHEN instr(name, '(') > 0 THEN trim(substr(name, 1, instr(name, '(') - 1)) ELSE name END"
         - name: asciiname_no_parens
-          type: TEXT
+          type: text
           expression: "CASE WHEN instr(asciiname, '(') > 0 THEN trim(substr(asciiname, 1, instr(asciiname, '(') - 1)) ELSE asciiname END"
     view:
       select:
@@ -130,18 +130,34 @@ sources:
         - source: cities500
           column: geometry
       join:
-        - type: LEFT JOIN
+        - method: left join
           source: countryInfo
-          condition: cities500.country_code = countryInfo.ISO
-        - type: LEFT JOIN
+          condition:
+            type: attribute
+            predicate: equals
+            left: cities500.country_code
+            right: countryInfo.ISO
+        - method: left join
           source: admin1CodesASCII
-          condition: cities500.admin1_code_full = admin1CodesASCII.code
-        - type: LEFT JOIN
+          condition:
+            type: attribute
+            predicate: equals
+            left: cities500.admin1_code_full
+            right: admin1CodesASCII.code
+        - method: left join
           source: admin2Codes
-          condition: cities500.admin2_code_full = admin2Codes.code
-        - type: LEFT JOIN
+          condition:
+            type: attribute
+            predicate: equals
+            left: cities500.admin2_code_full
+            right: admin2Codes.code
+        - method: left join
           source: featureCodes
-          condition: cities500.feature_code_full = featureCodes.code
+          condition:
+            type: attribute
+            predicate: equals
+            left: cities500.feature_code_full
+            right: featureCodes.code
     features:
       identifier:
[…]

[…]
