### 2026-06-06 645e603
Update processing stages and refactor pipeline execution
