+++ geoparser/gazetteer/installer/stages/spatial.py
@@ -1,0 +1,268 @@
+from typing import Any, Dict, Optional
+
+import geopandas as gpd
+import pandas as pd
+import sqlalchemy as sa
+
+from geoparser.db.db import get_connection
+from geoparser.gazetteer.installer.model import (
+    DataType,
+    GeometryTransform,
+    SourceConfig,
+    SpatialConditionConfig,
+)
+from geoparser.gazetteer.installer.queries.spatial import spatial_join_column_name
+from geoparser.gazetteer.installer.stages.base import Stage
+from geoparser.gazetteer.installer.utils.progress import create_progress_bar
+
+# Temporary table used to bulk-apply the precomputed spatial-join mapping.
+_MAPPING_TABLE = "_geoparser_spatial_join_mapping"
+
+
+class SpatialStage(Stage):
+    """
+    Precomputes spatial join conditions declared in a source's view.
+
+    Spatial relationships (e.g. which administrative polygon contains each
+    point) are computed in Python using GeoPandas' in-memory spatial index,
+    replacing the need for a spatial database extension. For each spatial
+    join, the matched right-table ``rowid`` is stored in a key column on the
+    left table so the view can join on plain equality.
+    """
+
+    def __init__(self, source_map: Dict[str, SourceConfig]):
+        """
+        Initialize the spatial stage.
+
+        Args:
+            source_map: Mapping of source name to its configuration, used to
+                resolve the spatial reference system of each geometry column
+        """
+        super().__init__(
+            name="Spatial",
+            description="Precompute spatial joins",
+        )
+        self.source_map = source_map
+
+    def execute(self, source: SourceConfig, context: Dict[str, Any]) -> None:
+        """
+        Precompute all spatial join conditions for a source's view.
+
+        Args:
+            source: Source configuration
+            context: Shared context (unused, kept for interface consistency)
+        """
+        if source.view is None or not source.view.join:
+            return
+
+        spatial_conditions = [
+            join_item.condition
+            for join_item in source.view.join
+            if join_item.condition.is_spatial
+        ]
+
+        for condition in spatial_conditions:
+            self._precompute_join(condition)
+
+    def _precompute_join(self, condition: SpatialConditionConfig) -> None:
+        """
+        Precompute a single spatial join and store the result column.
+
+        Args:
+            condition: Spatial join condition
+        """
+        left_table = condition.left_source
+        right_table = condition.right_source
+        column = spatial_join_column_name(right_table)
+
+        left_srid = self._get_srid(left_table, condition.left_column)
+        right_srid = self._get_srid(right_table, condition.right_column)
+
+        with create_progress_bar(
+            1,
+            f"Joining {left_table}.{condition.left_column}",
+            "column",
+        ) as pbar:
+            left_geometries = self._load_geometries(
+                left_table, condition.left_column, left_srid, condition.left_transform
+            )
+            right_geometries = self._load_geometries(
+                right_table,
+                condition.right_column,
+                right_srid,
+                condition.right_transform,
+            )
+
+            # Align coordinate reference systems before evaluating the predicate
+            if left_srid != right_srid:
+                right_geometries = right_geometries.to_crs(left_geometries.crs)
+
+            mapping = self._compute_mapping(
+                left_geometries, right_geometries, condition.predicate.value
+            )
+            self._store_mapping(left_table, column, mapping)
+
+            pbar.update(1)
+
+    def _get_srid(self, source_name: str, column_name: str) -> int:
+        """
+        Resolve the SRID of a geometry column from the source configuration.
+
+        Args:
[…]

[…]
