### 2026-06-06 3d095af
Remove SpatiaLite workflow and related binary verification steps from CI configuration
