+++ .github/workflows/ci.yml
@@ -31,12 +31,6 @@ jobs:
       run: |
         release=$(bash .github/stg_release.sh)
         poetry version $release
-    - name: Verify spatialite binaries are present
-      shell: bash
-      run: |
-        echo "Checking spatialite binaries..."
-        find geoparser/db/extensions/spatialite/binaries -name "mod_spatialite.*" -type f | sort
-        echo "All binaries will be included in the wheel"
     - name: Verify spellfix binaries are present
       shell: bash
       run: |
