### 2026-06-06 19692af
Update macOS compatibility guide and gazetteer documentation to remove SpatiaLite references and clarify geometry handling
