### 2026-06-06 8605a55
Refactor tests to remove SpatiaLite references and update geometry handling
