+++ tests/unit/test_gazetteer/test_installer/test_model.py
@@ -11,17 +11,23 @@ import pytest
 import yaml
 
 from geoparser.gazetteer.installer.model import (
+    AttributeConditionConfig,
+    AttributePredicate,
     AttributesConfig,
     ColumnConfig,
     DataType,
     DerivedAttributeConfig,
     FeatureConfig,
     GazetteerConfig,
+    GeometryTransform,
+    JoinConditionType,
     NameColumnConfig,
     OriginalAttributeConfig,
     SelectConfig,
     SourceConfig,
     SourceType,
+    SpatialConditionConfig,
+    SpatialPredicate,
     ViewConfig,
     ViewJoinConfig,
 )
@@ -33,23 +39,23 @@ class TestDataTypeEnum:
 
     def test_has_text_type(self):
         """Test that DATA_TYPE has TEXT."""
-        assert DataType.TEXT == "TEXT"
+        assert DataType.TEXT == "text"
 
     def test_has_integer_type(self):
         """Test that DataType has INTEGER."""
-        assert DataType.INTEGER == "INTEGER"
+        assert DataType.INTEGER == "integer"
 
     def test_has_real_type(self):
         """Test that DataType has REAL."""
-        assert DataType.REAL == "REAL"
+        assert DataType.REAL == "real"
 
     def test_has_blob_type(self):
         """Test that DataType has BLOB."""
-        assert DataType.BLOB == "BLOB"
+        assert DataType.BLOB == "blob"
 
     def test_has_geometry_type(self):
         """Test that DataType has GEOMETRY."""
-        assert DataType.GEOMETRY == "GEOMETRY"
+        assert DataType.GEOMETRY == "geometry"
 
 
 @pytest.mark.unit
@@ -239,19 +245,181 @@ class TestSelectConfig:
 class TestViewJoinConfig:
     """Test ViewJoinConfig model."""
 
-    def test_creates_with_valid_data(self):
-        """Test creating with valid join data."""
+    def test_creates_with_attribute_condition(self):
+        """Test creating with an attribute equality condition."""
         # Act
         join = ViewJoinConfig(
-            type="LEFT JOIN",
+            method="left join",
             source="table2",
-            condition="table1.id = table2.ref_id",
+            condition=AttributeConditionConfig(
+                predicate="equals",
+                left="table1.id",
+                right="table2.ref_id",
+            ),
         )
 
         # Assert
-        assert join.type == "LEFT JOIN"
+        assert join.method == "left join"
         assert join.source == "table2"
-        assert join.condition == "table1.id = table2.ref_id"
+        assert join.condition.type == JoinConditionType.ATTRIBUTE
+        assert join.condition.predicate == AttributePredicate.EQUALS
+        assert not join.condition.is_spatial
+
+    def test_creates_with_spatial_condition(self):
+        """Test creating a join whose condition is a spatial join."""
+        # Act
+        join = ViewJoinConfig(
+            method="left join",
+            source="regions",
+            condition=SpatialConditionConfig(
+                predicate="within",
+                left="places.geometry",
+                right="regions.geometry",
+            ),
+        )
+
+        # Assert
+        assert join.condition.type == JoinConditionType.SPATIAL
+        assert join.condition.predicate == SpatialPredicate.WITHIN
+        assert join.condition.is_spatial
+
+    def test_parses_attribute_condition_from_mapping(self):
+        """Test that a mapping with type 'attribute' parses to the right model."""
+        # Act
+        join = ViewJoinConfig(
+            method="left join",
+            source="regions",
+            condition={
+                "type": "attribute",
+                "predicate": "equals",
+                "left": "places.admin_code",
+                "right": "regions.code",
+            },
+        )
+
+        # Assert
[…]
