+++ tests/unit/test_gazetteer/test_installer/test_stages/test_indexing.py
@@ -159,6 +159,38 @@ class TestIndexingStageCollectIndexedColumns:
         assert len(result) == 1
         assert result[0] == ("id", DataType.INTEGER)
 
+    def test_excludes_geometry_columns(self):
+        """Test that geometry columns are not indexed in the database."""
+        # Arrange
+        source = SourceConfig(
+            name="test",
+            url="http://example.com/data.shp",
+            file="data.shp",
+            type=SourceType.SPATIAL,
+            attributes=AttributesConfig(
+                original=[
+                    OriginalAttributeConfig(
+                        name="id", type=DataType.INTEGER, index=True
+                    ),
+                    OriginalAttributeConfig(
+                        name="geometry",
+                        type=DataType.GEOMETRY,
+                        index=True,
+                        srid=4326,
+                    ),
+                ]
+            ),
+        )
+
+        stage = IndexingStage()
+
+        # Act
+        result = stage._collect_indexed_columns(source)
+
+        # Assert
+        assert len(result) == 1
+        assert result[0] == ("id", DataType.INTEGER)
+
     def test_collects_both_original_and_derived_indexed_columns(self):
         """Test that both original and derived indexed columns are collected."""
         # Arrange

[…]
