### 2026-06-06 4e14090
Refactor gazetteer config model to enhance column reference handling and update source type terminology
