+++ geoparser/gazetteer/installer/model.py
@@ -6,8 +6,8 @@ import yaml
 from pydantic import BaseModel, Field, field_validator, model_validator
 
 
-class SourceType(str, Enum):
-    """Type of source data."""
+class SourceKind(str, Enum):
+    """Kind of source data."""
 
     TABULAR = "tabular"
     SPATIAL = "spatial"
@@ -95,38 +95,70 @@ class AttributesConfig(BaseModel):
     derived: t.List[DerivedAttributeConfig] = []
 
 
+class ColumnConfig(BaseModel):
+    """
+    Reference to a column in a specific source.
+
+    Written in configuration as the string ``source.column`` and parsed into
+    its ``source`` and ``column`` parts. Used wherever a config needs to point
+    at a specific column in a specific source (e.g. ``select`` items and join
+    operands).
+    """
+
+    source: str
+    column: str
+
+    @model_validator(mode="before")
+    @classmethod
+    def parse_reference(cls, data: t.Any) -> t.Any:
+        """Allow a ``'source.column'`` string in place of an explicit mapping."""
+        if isinstance(data, str):
+            parts = data.split(".")
+            if len(parts) != 2 or not all(parts):
+                raise ValueError(
+                    f"Column reference '{data}' must be in the format "
+                    "'source.column'"
+                )
+            return {"source": parts[0], "column": parts[1]}
+        return data
+
+    @property
+    def sql(self) -> str:
+        """The column reference rendered as ``source.column``."""
+        return f"{self.source}.{self.column}"
+
+
 class SelectConfig(BaseModel):
     """Configuration for a SELECT clause element."""
 
-    source: str  # Source/view name (replaces 'table')
-    column: str
+    column: ColumnConfig
     alias: t.Optional[str] = None
+
+
+class JoinOperandConfig(BaseModel):
+    """
+    One side (left or right) of a join condition.
+
+    References a column and, for spatial joins, may apply a geometry transform
+    (e.g. ``centroid``) before the predicate is evaluated.
+    """
+
+    column: ColumnConfig
+    transform: t.Optional[GeometryTransform] = None
 
 
 class ConditionConfig(BaseModel):
     """
     Base configuration for a join condition.
 
-    A condition relates a ``left`` and ``right`` reference, each given as
-    ``source.column``. The concrete ``type`` (attribute or spatial) determines
-    how the condition is evaluated.
+    A condition relates a ``left`` and ``right`` operand, each referencing a
+    column. The concrete ``type`` (attribute or spatial) determines how the
+    condition is evaluated.
     """
 
     type: JoinConditionType
-    left: str  # Reference as "source.column"
-    right: str  # Reference as "source.column"
-
-    @field_validator("left", "right")
-    @classmethod
-    def validate_reference(cls, v: str) -> str:
-        """Validate that references use the 'source.column' format."""
-        parts = v.split(".")
-        if len(parts) != 2 or not all(parts):
-            raise ValueError(
-                f"Join condition reference '{v}' must be in the "
-                "format 'source.column'"
-            )
-        return v
+    left: JoinOperandConfig
+    right: JoinOperandConfig
 
     @property
     def is_spatial(self) -> bool:
@@ -135,23 +167,23 @@ class ConditionConfig(BaseModel):
 
     @property
     def left_source(self) -> str:
-        """Source/table name of the left reference."""
-        return self.left.split(".", 1)[0]
+        """Source/table name of the left operand."""
+        return self.left.column.source
 
     @property
     def left_column(self) -> str:
-        """Column name of the left reference."""
-        return self.left.split(".", 1)[1]
+        """Column name of the left operand."""
+        return self.left.column.column
 
     @property
     def right_source(self) -> str:
[…]
