### 2026-06-06 dd6b3eb
Update gazetteer configurations for updated config model
