+++ geoparser/gazetteer/configs/geonames-cities.yaml
@@ -3,7 +3,7 @@ sources:
   - name: cities500
     url: https://download.geonames.org/export/dump/cities500.zip
     file: cities500.txt
-    type: tabular
+    kind: tabular
     separator: "\t"
     attributes:
       original:
@@ -73,106 +73,84 @@ sources:
           expression: "CASE WHEN instr(asciiname, '(') > 0 THEN trim(substr(asciiname, 1, instr(asciiname, '(') - 1)) ELSE asciiname END"
     view:
       select:
-        - source: cities500
-          column: geonameid
-        - source: cities500
-          column: name
-        - source: cities500
-          column: name_no_parens
-        - source: cities500
-          column: asciiname
-        - source: cities500
-          column: asciiname_no_parens
-        - source: cities500
-          column: alternatenames
-        - source: cities500
-          column: latitude
-        - source: cities500
-          column: longitude
-        - source: cities500
-          column: feature_class
-        - source: cities500
-          column: feature_code
-        - source: featureCodes
-          column: name
+        - column: cities500.geonameid
+        - column: cities500.name
+        - column: cities500.name_no_parens
+        - column: cities500.asciiname
+        - column: cities500.asciiname_no_parens
+        - column: cities500.alternatenames
+        - column: cities500.latitude
+        - column: cities500.longitude
+        - column: cities500.feature_class
+        - column: cities500.feature_code
+        - column: featureCodes.name
           alias: feature_name
-        - source: cities500
-          column: country_code
-        - source: countryInfo
-          column: Country
+        - column: cities500.country_code
+        - column: countryInfo.Country
           alias: country_name
-        - source: cities500
-          column: cc2
-        - source: cities500
-          column: admin1_code
-        - source: admin1CodesASCII
-          column: name
+        - column: cities500.cc2
+        - column: cities500.admin1_code
+        - column: admin1CodesASCII.name
           alias: admin1_name
-        - source: cities500
-          column: admin2_code
-        - source: admin2Codes
-          column: name
+        - column: cities500.admin2_code
+        - column: admin2Codes.name
           alias: admin2_name
-        - source: cities500
-          column: admin3_code
-        - source: cities500
-          column: admin4_code
-        - source: cities500
-          column: population
-        - source: cities500
-          column: elevation
-        - source: cities500
-          column: dem
-        - source: cities500
-          column: timezone
-        - source: cities500
-          column: modification_date
-        - source: cities500
-          column: geometry
+        - column: cities500.admin3_code
+        - column: cities500.admin4_code
+        - column: cities500.population
+        - column: cities500.elevation
+        - column: cities500.dem
+        - column: cities500.timezone
+        - column: cities500.modification_date
+        - column: cities500.geometry
       join:
         - method: left join
-          source: countryInfo
           condition:
             type: attribute
             predicate: equals
-            left: cities500.country_code
-            right: countryInfo.ISO
+            left:
+              column: cities500.country_code
+            right:
+              column: countryInfo.ISO
         - method: left join
-          source: admin1CodesASCII
           condition:
             type: attribute
             predicate: equals
-            left: cities500.admin1_code_full
-            right: admin1CodesASCII.code
+            left:
+              column: cities500.admin1_code_full
+            right:
+              column: admin1CodesASCII.code
         - method: left join
-          source: admin2Codes
           condition:
             type: attribute
             predicate: equals
-            left: cities500.admin2_code_full
-            right: admin2Codes.code
+            left:
[…]
