### 2026-06-06 38debf7
Update gazetteer documentation for new gazetteer config model
