+++ docs/guides/gazetteers.rst
@@ -218,7 +218,7 @@ Sources can be either tabular (CSV, TSV) or spatial (shapefiles, GeoPackage). Fo
      - name: places
        url: https://example.com/data.zip  # Downloaded and extracted
        file: places.csv                   # File within the ZIP
-       type: tabular
+       kind: tabular
        separator: ","
 
 For local files, provide both the ``path`` (directory containing the file) and ``file`` (filename):
@@ -229,7 +229,7 @@ For local files, provide both the ``path`` (directory containing the file) and `
      - name: local_data
        path: /path/to/data/directory
        file: data.csv
-       type: tabular
+       kind: tabular
        separator: "\t"
 
 Defining Attributes
@@ -237,7 +237,7 @@ Defining Attributes
 
 Each source must declare its attributes in two categories: original attributes that exist in the source file, and derived attributes computed from SQL expressions.
 
-Original attributes match columns in the source file. Specify their data types (TEXT, INTEGER, REAL, GEOMETRY) and optionally mark them for indexing:
+Original attributes match columns in the source file. Specify their data types (text, integer, real, geometry) and optionally mark them for indexing:
 
 .. code-block:: yaml
 
@@ -283,29 +283,24 @@ If you have multiple sources, you need to define a view that specifies which col
 
    view:
      select:
-       - source: places        # Source table name
-         column: geonameid     # Column to include
-       - source: places
-         column: name
-       - source: places
-         column: latitude
-       - source: places
-         column: longitude
-       - source: admin_names   # From a different source
-         column: name
-         alias: admin_name     # Rename column in view
-       - source: places
-         column: geometry
+       - column: places.geonameid    # source.column to include
+       - column: places.name
+       - column: places.latitude
+       - column: places.longitude
+       - column: admin_names.name     # From a different source
+         alias: admin_name            # Rename column in view
+       - column: places.geometry
      join:
-       - method: left join     # Join method
-         source: admin_names   # Source to join
+       - method: left join            # Join method
          condition:
-           type: attribute                      # attribute | spatial
-           predicate: equals                    # equals
-           left: places.admin_code              # reference as source.column
-           right: admin_names.code
+           type: attribute            # attribute | spatial
+           predicate: equals          # equals
+           left:
+             column: places.admin_code    # source.column reference
+           right:
+             column: admin_names.code
 
-A join ``condition`` always has the same shape: a ``type`` (``attribute`` or ``spatial``), a ``predicate`` relating a ``left`` and ``right`` reference (each written as ``source.column``). An ``attribute`` condition produces a plain SQL join (currently the ``equals`` predicate), enriching your main features table with data from auxiliary tables.
+Both a ``select`` item and a join operand reference a column with a single ``source.column`` string. A join ``condition`` always has the same shape: a ``type`` (``attribute`` or ``spatial``), a ``predicate``, and ``left``/``right`` operands. The joined source is inferred from the right operand, so it does not need to be repeated. An ``attribute`` condition produces a plain SQL join (currently the ``equals`` predicate), enriching your main features table with data from auxiliary tables.
 
 Spatial Joins
 ~~~~~~~~~~~~~
@@ -316,46 +311,48 @@ For determining spatial relationships (e.g., which administrative region contain
 
    join:
      - method: left join
-       source: municipalities
        condition:
          type: spatial                # attribute | spatial
          predicate: within            # within | intersects | contains
-         left: places.geometry        # geometry reference as source.column
-         right: municipalities.geometry
+         left:
+           column: places.geometry    # geometry source.column reference
+         right:
+           column: municipalities.geometry
 
[…]
