### 2026-06-06 1cc2009
Update tests for new gazetteer config model
