+++ tests/unit/test_gazetteer/test_installer/test_model.py
@@ -20,12 +20,14 @@ from geoparser.gazetteer.installer.model import (
     FeatureConfig,
     GazetteerConfig,
     GeometryTransform,
+    IdentifierColumnConfig,
     JoinConditionType,
+    JoinOperandConfig,
     NameColumnConfig,
     OriginalAttributeConfig,
     SelectConfig,
     SourceConfig,
-    SourceType,
+    SourceKind,
     SpatialConditionConfig,
     SpatialPredicate,
     ViewConfig,
@@ -59,16 +61,16 @@ class TestDataTypeEnum:
 
 
 @pytest.mark.unit
-class TestSourceTypeEnum:
-    """Test SourceType enum."""
+class TestSourceKindEnum:
+    """Test SourceKind enum."""
 
-    def test_has_tabular_type(self):
-        """Test that SourceType has TABULAR."""
-        assert SourceType.TABULAR == "tabular"
+    def test_has_tabular_kind(self):
+        """Test that SourceKind has TABULAR."""
+        assert SourceKind.TABULAR == "tabular"
 
-    def test_has_spatial_type(self):
-        """Test that SourceType has SPATIAL."""
-        assert SourceType.SPATIAL == "spatial"
+    def test_has_spatial_kind(self):
+        """Test that SourceKind has SPATIAL."""
+        assert SourceKind.SPATIAL == "spatial"
 
 
 @pytest.mark.unit
@@ -222,20 +224,20 @@ class TestAttributesConfig:
 class TestSelectConfig:
     """Test SelectConfig model."""
 
-    def test_creates_with_source_and_column(self):
-        """Test creating with source and column."""
+    def test_creates_with_column_reference(self):
+        """Test creating with a 'source.column' reference."""
         # Act
-        select = SelectConfig(source="table1", column="col1")
+        select = SelectConfig(column="table1.col1")
 
         # Assert
-        assert select.source == "table1"
-        assert select.column == "col1"
+        assert select.column.source == "table1"
+        assert select.column.column == "col1"
         assert select.alias is None
 
     def test_creates_with_alias(self):
         """Test creating with alias."""
         # Act
-        select = SelectConfig(source="table1", column="col1", alias="renamed_col")
+        select = SelectConfig(column="table1.col1", alias="renamed_col")
 
         # Assert
         assert select.alias == "renamed_col"
@@ -250,11 +252,10 @@ class TestViewJoinConfig:
         # Act
         join = ViewJoinConfig(
             method="left join",
-            source="table2",
             condition=AttributeConditionConfig(
                 predicate="equals",
-                left="table1.id",
-                right="table2.ref_id",
+                left=JoinOperandConfig(column="table1.id"),
+                right=JoinOperandConfig(column="table2.ref_id"),
             ),
         )
 
@@ -270,11 +271,10 @@ class TestViewJoinConfig:
         # Act
         join = ViewJoinConfig(
             method="left join",
-            source="regions",
             condition=SpatialConditionConfig(
                 predicate="within",
-                left="places.geometry",
-                right="regions.geometry",
+                left=JoinOperandConfig(column="places.geometry"),
+                right=JoinOperandConfig(column="regions.geometry"),
             ),
         )
 
@@ -288,12 +288,11 @@ class TestViewJoinConfig:
         # Act
         join = ViewJoinConfig(
             method="left join",
-            source="regions",
             condition={
                 "type": "attribute",
                 "predicate": "equals",
-                "left": "places.admin_code",
-                "right": "regions.code",
+                "left": {"column": "places.admin_code"},
+                "right": {"column": "regions.code"},
             },
         )
 
@@ -306,12 +305,11 @@ class TestViewJoinConfig:
         # Act
         join = ViewJoinConfig(
[…]
