### 2026-06-13 be26b87
Add rapidfuzz for fuzzy string matching and replace Spellfix extension with custom soundex and levenshtein functions
