### 2026-06-14 3972c6b
Update tests for replacement of Spellfix with Soundex and Levenshtein
