### 2026-06-14 fe06477
Remove macOS compatibility guide and update installation documentation to reflect changes in Python requirements
