### 2026-06-20 87a1a24
Add optimized writes context manager for improved database performance during gazetteer installation
