### 2026-06-26 8e7e956
Add unit tests for optimized writes context manager in database operations
