### 2026-06-26 601c312
Implement chunked processing for database operations in gazetteer installer
