### 2026-06-26 2f2f2e6
Update and extend tests for chunked processing in gazetteer installer
