+++ tests/unit/test_gazetteer/test_installer/test_queries/test_dml.py
@@ -33,11 +33,14 @@ class TestTransformationBuilderBuildDerivationUpdate:
 
         # Act
         sql = builder.build_derivation_update(
-            "test_table", "full_name", "first_name || ' ' || last_name"
+            "test_table", "full_name", "first_name || ' ' || last_name", 1, 20000
         )
 
         # Assert
-        assert sql == "UPDATE test_table SET full_name = first_name || ' ' || last_name"
+        assert sql == (
+            "UPDATE test_table SET full_name = first_name || ' ' || last_name "
+            "WHERE rowid BETWEEN 1 AND 20000"
+        )
 
     def test_sanitizes_table_name(self):
         """Test that table name is sanitized."""
@@ -46,7 +49,7 @@ class TestTransformationBuilderBuildDerivationUpdate:
 
         # Act & Assert
         with pytest.raises(ValueError, match="Invalid identifier"):
-            builder.build_derivation_update("table-name", "col", "expression")
+            builder.build_derivation_update("table-name", "col", "expression", 1, 100)
 
     def test_sanitizes_column_name(self):
         """Test that column name is sanitized."""
@@ -55,7 +58,7 @@ class TestTransformationBuilderBuildDerivationUpdate:
 
         # Act & Assert
         with pytest.raises(ValueError, match="Invalid identifier"):
-            builder.build_derivation_update("table", "col-name", "expression")
+            builder.build_derivation_update("table", "col-name", "expression", 1, 100)
 
 
 @pytest.mark.unit
@@ -83,7 +86,9 @@ class TestFeatureRegistrationBuilderBuildFeatureInsert:
         builder = FeatureRegistrationBuilder()
 
         # Act
-        sql = builder.build_feature_insert(source, source_id=123)
+        sql = builder.build_feature_insert(
+            source, source_id=123, rowid_start=1, rowid_end=20000
+        )
 
         # Assert
         assert "INSERT OR IGNORE INTO feature" in sql
@@ -92,6 +97,7 @@ class TestFeatureRegistrationBuilderBuildFeatureInsert:
         assert "CAST(id AS TEXT) as location_id_value" in sql
         assert "FROM test_source" in sql
         assert "WHERE id IS NOT NULL" in sql
+        assert "test_source.rowid BETWEEN 1 AND 20000" in sql
 
     def test_raises_error_when_source_has_no_features(self):
         """Test that error is raised when source has no feature configuration."""
@@ -111,7 +117,9 @@ class TestFeatureRegistrationBuilderBuildFeatureInsert:
 
         # Act & Assert
         with pytest.raises(ValueError, match="has no feature configuration"):
-            builder.build_feature_insert(source, source_id=123)
+            builder.build_feature_insert(
+                source, source_id=123, rowid_start=1, rowid_end=20000
+            )
 
 
 @pytest.mark.unit
@@ -142,7 +150,9 @@ class TestFeatureRegistrationBuilderBuildNameInsert:
         builder = FeatureRegistrationBuilder()
 
         # Act
-        sql = builder.build_name_insert(source, source_id=123, name_column="name")
+        sql = builder.build_name_insert(
+            source, source_id=123, name_column="name", rowid_start=1, rowid_end=20000
+        )
 
         # Assert
         assert "INSERT OR IGNORE INTO name" in sql
@@ -152,6 +162,7 @@ class TestFeatureRegistrationBuilderBuildNameInsert:
         assert "FROM test_source s" in sql
         assert "JOIN feature f ON f.source_id = 123" in sql
         assert "WHERE s.name IS NOT NULL AND s.name != ''" in sql
+        assert "s.rowid BETWEEN 1 AND 20000" in sql
 
     def test_raises_error_when_source_has_no_features(self):
         """Test that error is raised when source has no feature configuration."""
[…]
