### 2026-06-26 d60e06b
Enhanced progress reporting in the installer using rich to provide more compact and more modern feedback on processing stages
