+++ tests/unit/test_gazetteer/test_installer/test_utils/test_progress.py
@@ -1,13 +1,36 @@
 """
 Unit tests for geoparser/gazetteer/installer/utils/progress.py
 
-Tests the create_progress_bar utility function.
+Tests the create_progress_bar utility function and the source_progress group.
 """
 
-import pytest
-from tqdm.auto import tqdm
+from unittest.mock import patch
 
-from geoparser.gazetteer.installer.utils.progress import create_progress_bar
+import pytest
+from rich.console import Console
+from rich.progress import (
+    DownloadColumn,
+    MofNCompleteColumn,
+    SpinnerColumn,
+    TimeElapsedColumn,
+    TransferSpeedColumn,
+)
+
+from geoparser.gazetteer.installer.utils.progress import (
+    _CHILD_CONNECTOR,
+    _SOURCE_PROCESSED_LABEL,
+    _SOURCE_PROCESSING_LABEL,
+    _description_width,
+    _DescriptionColumn,
+    _layout_width,
+    _RichProgressBar,
+    _StepAwareColumn,
+    _StyledBarColumn,
+    create_progress_bar,
+    print_gazetteer_header,
+    print_gazetteer_summary,
+    source_progress,
+)
 
 
 @pytest.mark.unit
@@ -20,13 +43,10 @@ class TestCreateProgressBar:
         pbar = create_progress_bar(total=100, description="Processing")
 
         # Assert
-        assert isinstance(pbar, tqdm)
+        assert isinstance(pbar, _RichProgressBar)
         assert pbar.total == 100
-        assert pbar.desc == "Processing"
+        assert pbar.description == "Processing"
         assert pbar.unit == "items"
-
-        # Clean up
-        pbar.close()
 
     def test_creates_progress_bar_with_custom_unit(self):
         """Test creating progress bar with custom unit."""
@@ -37,32 +57,340 @@ class TestCreateProgressBar:
 
         # Assert
         assert pbar.total == 1000
-        assert pbar.desc == "Downloading"
+        assert pbar.description == "Downloading"
         assert pbar.unit == "records"
 
-        # Clean up
-        pbar.close()
-
-    def test_creates_progress_bar_with_bytes_unit(self):
-        """Test creating progress bar with bytes unit enables scaling."""
+    def test_supports_context_manager_and_update(self):
+        """Test the bar works as a context manager and advances on update."""
         # Act
-        pbar = create_progress_bar(total=1024, description="Transferring", unit="B")
+        with create_progress_bar(total=10, description="Processing") as pbar:
+            pbar.update(3)
+            pbar.update(2)
+            task = pbar._progress.tasks[0]
+
+            # Assert
+            assert task.completed == 5
+            assert task.total == 10
+
+    @pytest.mark.parametrize("unit", ["B", "bytes", "items", "rows", "view"])
+    def test_layout_is_uniform_across_units(self, unit):
+        """Test that the column layout is identical regardless of unit."""
+        # Act
+        pbar = create_progress_bar(total=100, description="Processing", unit=unit)
+        column_types = [type(column) for column in pbar._progress.columns]
 
         # Assert
-        assert pbar.total == 1024
-        assert pbar.unit == "B"
-        assert pbar.unit_scale is True  # Should enable scaling for bytes
+        assert column_types == [
+            _DescriptionColumn,
+            _StyledBarColumn,
+            _StepAwareColumn,
+            TimeElapsedColumn,
+        ]
 
-        # Clean up
-        pbar.close()
-
-    def test_creates_progress_bar_without_scaling_for_non_bytes(self):
-        """Test that non-byte units don't enable scaling."""
+    @pytest.mark.parametrize("unit", ["B", "items"])
+    def test_no_spinner_column(self, unit):
[…]
