+++ geoparser/gazetteer/installer/utils/progress.py
@@ -1,16 +1,292 @@
-from tqdm.auto import tqdm
+from contextlib import contextmanager
+from contextvars import ContextVar
+from typing import Iterator, List, Optional
+
+from rich.console import Console
+from rich.progress import (
+    BarColumn,
+    Progress,
+    ProgressColumn,
+    Task,
+    TimeElapsedColumn,
+)
+from rich.rule import Rule
+from rich.spinner import Spinner
+from rich.table import Column, Table
+from rich.text import Text
+
+# Connector that visually nests an action (child) bar under its source.
+_CHILD_CONNECTOR = "  └─ "
+_CHILD_SPINNER_NAME = "dots"
+
+# Labels for the source (parent) line. Past tense once all steps complete so
+# each state reads naturally without symbols.
+_SOURCE_PROCESSING_LABEL = "Processing "
+_SOURCE_PROCESSED_LABEL = "Processed "
+# Status colors. A source is "in progress" (cyan) until all of its steps
+# complete, then "done" (green); action bars use the in-progress color.
+_RUNNING_STYLE = "cyan"
+_DONE_STYLE = "green"
+
+# Holds the progress group of the source currently being installed. Progress
+# bars created deep inside the pipeline stages read this to attach themselves
+# to the shared, per-source display instead of creating their own.
+_active_group: ContextVar[Optional["_SourceProgress"]] = ContextVar(
+    "active_progress_group", default=None
+)
+
+
+class _DescriptionColumn(ProgressColumn):
+    """
+    Render the description so sources and actions are easy to tell apart.
+
+    A source (parent) task is shown as "Processing <name>" while its pipeline
+    runs, then "Processed <name>" once all steps complete. An action (child)
+    task is dimmed and nested under its source with a tree connector and an
+    animated spinner while the step is running.
+    """
+
+    def __init__(self, description_width: int) -> None:
+        super().__init__(
+            table_column=Column(
+                width=description_width, no_wrap=True, overflow="ellipsis"
+            )
+        )
+        self._child_spinner = Spinner(_CHILD_SPINNER_NAME, style="dim")
+
+    def render(self, task: Task) -> Text:
+        description = str(task.description)
+        if task.fields.get("is_step_task"):
+            if task.finished:
+                return Text.assemble(
+                    (_SOURCE_PROCESSED_LABEL, _DONE_STYLE),
+                    (description, "bold"),
+                )
+            return Text.assemble(
+                (_SOURCE_PROCESSING_LABEL, _RUNNING_STYLE),
+                (description, "bold"),
+            )
+        if task.fields.get("is_child"):
+            line = Text(_CHILD_CONNECTOR, style="dim")
+            if not task.finished:
+                line.append_text(self._child_spinner.render(task.get_time()))
+                line.append(" ")
+            line.append(description, style="dim")
+            return line
+        return Text(description)
+
+
+class _StepAwareColumn(ProgressColumn):
+    """
+    Render the absolute-progress readout for a task.
+
+    The per-source step (parent) task is shown as a "completed/total" step
+    count; ordinary task bars are shown as a percentage.
+    """
+
+    def render(self, task: Task) -> Text:
+        if task.fields.get("is_step_task"):
+            total = int(task.total) if task.total else 0
+            return Text(f"{int(task.completed)}/{total}", style="progress.percentage")
+        return Text(f"{int(task.percentage)}%", style="progress.percentage")
+
+
+class _StyledBarColumn(BarColumn):
+    """
+    Bar column that colors bars by status so the hierarchy and state are
+    visible at a glance: a source bar turns from in-progress to done when all
+    of its steps complete, while action bars use the in-progress color.
+    """
+
+    def render(self, task: Task):
[…]
