### 2026-06-26 d1ba3ec
Update integration tests to replace occurrences of "Encamp" with "Canillo" for more confident prediction threshold clearing
