### 2026-06-26 0f5df50
Add database compatibility check to prevent legacy schema usage
