### 2026-06-27 6b8c40c
Remove SQLite-specific check from database compatibility function
