### 2026-06-27 ed2589e
Add installed_at field to Gazetteer model and update installation logic to raise an error if an uninstalled gazetteer is queried
