### 2026-06-27 4ef78bf
Add fixture for installed geonames and update tests to ensure gazetteer installation is validated before use
