### 2026-06-27 687a527
Fix integration test for SentenceTransformerResolver to verify context extraction and truncation behavior in a more targeted manner
