### 2026-06-27 f5b5660
Rename the gazetteer install CLI command from `download` to `install`
