### 2026-06-28 1c01280
Remove obsolete macOS installation note from README
