### 2026-07-09 8937eb1
First draft of gazetteer installer pipeline redesign
