+++ geoparser/gazetteer/artifact.py
@@ -1,0 +1,413 @@
+"""
+Gazetteer artifact: the canonical, self-contained SQLite file per gazetteer.
+
+An artifact is produced by the build pipeline and is immutable afterwards. It
+has a fixed schema shared by every gazetteer:
+
+- ``metadata``: key/value pairs (schema version, gazetteer name, CRS, ...)
+- ``feature``: one row per feature (identifier, type, JSON attributes, WKB geometry)
+- ``name``: searchable names, one row per (feature, name)
+- ``name_fts``: FTS5 index over names (external content on ``name``)
+- ``name_soundex``: phonetic codes for fuzzy candidate retrieval
+
+This module owns the artifact schema (used by the build pipeline) and the
+read-only runtime access layer (used by the Gazetteer class).
+"""
+
+from __future__ import annotations
+
+import os
+import sqlite3
+import threading
+import typing as t
+from pathlib import Path
+
+from appdirs import user_data_dir
+
+from geoparser.db.functions import levenshtein, soundex
+from geoparser.gazetteer.feature import Feature
+
+# Bump when the artifact schema changes; artifacts with a different version
+# must be rebuilt.
+SCHEMA_VERSION = "1"
+
+ARTIFACT_SUFFIX = ".db"
+
+FTS_TOKENIZER = "unicode61 remove_diacritics 2 tokenchars '.'"
+
+# Base tables created before data is emitted. Search structures (FTS, soundex,
+# indexes) are created afterwards, once all rows are in place.
+BASE_SCHEMA = [
+    """
+    CREATE TABLE metadata (
+        key TEXT PRIMARY KEY,
+        value TEXT
+    )
+    """,
+    """
+    CREATE TABLE feature (
+        id INTEGER PRIMARY KEY,
+        identifier TEXT NOT NULL,
+        type TEXT NOT NULL,
+        attributes TEXT NOT NULL,
+        geometry BLOB
+    )
+    """,
+    """
+    CREATE TABLE name (
+        id INTEGER PRIMARY KEY,
+        feature_id INTEGER NOT NULL REFERENCES feature(id),
+        text TEXT NOT NULL
+    )
+    """,
+]
+
+SEARCH_SCHEMA = [
+    "CREATE UNIQUE INDEX idx_feature_identifier ON feature(identifier)",
+    "CREATE INDEX idx_name_feature_id ON name(feature_id)",
+    f"""
+    CREATE VIRTUAL TABLE name_fts USING fts5(
+        text,
+        content='name',
+        content_rowid='id',
+        tokenize="{FTS_TOKENIZER}"
+    )
+    """,
+    "INSERT INTO name_fts(rowid, text) SELECT id, text FROM name",
+    """
+    CREATE TABLE name_soundex (
+        id INTEGER PRIMARY KEY,
+        code TEXT
+    )
+    """,
+    "INSERT INTO name_soundex(id, code) SELECT id, soundex(text) FROM name",
+    "CREATE INDEX idx_name_soundex_code ON name_soundex(code)",
+]
+
+
+def gazetteers_dir() -> Path:
+    """
+    Return the directory holding installed gazetteer artifacts.
+
+    Defaults to the user data directory; can be overridden with the
+    ``GEOPARSER_GAZETTEERS_DIR`` environment variable.
+
+    Returns:
+        Path to the gazetteers directory
+    """
+    override = os.getenv("GEOPARSER_GAZETTEERS_DIR")
+    if override:
+        return Path(override)
+    return Path(user_data_dir("geoparser", "")) / "gazetteers"
+
+
+def artifact_path(gazetteer_name: str) -> Path:
+    """
+    Return the artifact path for a gazetteer name.
+
+    Args:
[…]
