+++ tests/unit/test_gazetteer/test_installer/test_queries/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for geoparser/gazetteer/installer/queries/"""
