+++ geoparser/gazetteer/installer/__init__.py
@@ -1,1 +1,0 @@
-from geoparser.gazetteer.installer.installer import GazetteerInstaller

[…]
