### 2026-07-09 cd1c18b
Add gazetteer build pipeline components
