+++ geoparser/gazetteer/build/builder.py
@@ -1,0 +1,318 @@
+"""
+Gazetteer builder: compiles a declarative config into a SQLite artifact.
+
+The build pipeline is:
+
+1. Acquire: download/extract the input files (cached across builds)
+2. Stage: load each input into a transient DuckDB table
+3. Project: compile and run one projection per feature block, producing
+   canonical (identifier, type, attributes, geometry) and (identifier, name)
+   rows with explicit merge policies for duplicate identifiers
+4. Emit: copy the projected rows into a temporary SQLite artifact
+5. Finalize: build FTS/soundex/indexes and metadata, vacuum, and atomically
+   move the artifact into place
+
+The staging database and all intermediate tables are discarded after the
+build; the artifact is the only output.
+"""
+
+import contextlib
+import os
+import shutil
+import typing as t
+from pathlib import Path
+
+import duckdb
+
+from geoparser.gazetteer import artifact
+from geoparser.gazetteer.build.acquire import Acquirer
+from geoparser.gazetteer.build.compile import ProjectionCompiler
+from geoparser.gazetteer.build.emit import create_artifact_db, emit, finalize
+from geoparser.gazetteer.build.progress import (
+    print_build_header,
+    print_build_summary,
+    step,
+)
+from geoparser.gazetteer.build.stage import Stager, quote_literal
+from geoparser.gazetteer.config import GazetteerConfig
+
+
+@contextlib.contextmanager
+def _sqlite_tmpdir(directory: Path) -> t.Iterator[None]:
+    """
+    Temporarily point SQLite's temporary storage at ``directory``.
+
+    SQLite has no per-connection temp-directory setting, so its location is
+    controlled through the ``SQLITE_TMPDIR`` environment variable, which is
+    read when temporary files are created.
+
+    Args:
+        directory: Directory SQLite should use for temporary files
+    """
+    previous = os.environ.get("SQLITE_TMPDIR")
+    os.environ["SQLITE_TMPDIR"] = str(directory)
+    try:
+        yield
+    finally:
+        if previous is None:
+            os.environ.pop("SQLITE_TMPDIR", None)
+        else:
+            os.environ["SQLITE_TMPDIR"] = previous
+
+
+class GazetteerBuilder:
+    """
+    Builds gazetteer artifacts from declarative configuration files.
+    """
+
+    # Fraction of physical memory the staging engine may use before it starts
+    # spilling intermediate data to disk.
+    _MEMORY_FRACTION = 0.6
+    # Headroom always left for the OS, the Python process and the SQLite
+    # writer, so the whole build never approaches the machine's memory cap.
+    _MEMORY_HEADROOM_MB = 1536
+    # Never bound the staging engine below this, or builds become impractically
+    # slow.
+    _MIN_MEMORY_MB = 512
+
+    def build(
+        self, config_path: t.Union[str, Path], keep_downloads: bool = False
+    ) -> Path:
+        """
+        Build and install a gazetteer from a configuration file.
+
+        Args:
+            config_path: Path to the YAML configuration file
+            keep_downloads: Whether to keep downloaded files after the build
+
+        Returns:
+            Path of the installed artifact
+        """
+        config = GazetteerConfig.from_yaml(config_path)
+        print_build_header(config.name)
+
+        target_path = artifact.artifact_path(config.name)
+        target_path.parent.mkdir(parents=True, exist_ok=True)
+
+        downloads_dir = target_path.parent / ".downloads" / config.name
+        acquirer = Acquirer(downloads_dir)
+        build_dir = target_path.parent / f".build-{config.name}"
+        if build_dir.exists():
+            shutil.rmtree(build_dir)
+        build_dir.mkdir(parents=True)
+
+        try:
+            file_paths = {
[…]
