+++ geoparser/gazetteer/build/emit.py
@@ -1,0 +1,168 @@
+"""
+Emission of the projected features into the SQLite artifact.
+
+Rows are copied from the DuckDB build tables into a temporary SQLite file in
+batches; search structures (FTS5, soundex, indexes) and metadata are built
+afterwards, then the file is vacuumed and atomically moved into place.
+"""
+
+import sqlite3
+import typing as t
+from datetime import datetime, timezone
+from pathlib import Path
+
+import duckdb
+
+from geoparser.gazetteer import artifact
+from geoparser.gazetteer.config import GazetteerConfig
+
+# Number of rows copied per batch from DuckDB to SQLite
+BATCH_SIZE = 50000
+
+
+def create_artifact_db(path: Path) -> sqlite3.Connection:
+    """
+    Create a fresh artifact database with the base schema.
+
+    The connection is tuned for bulk writes with a bounded memory footprint:
+    the artifact is a throwaway temporary file that is only moved into place
+    after a fully successful build, so durability is irrelevant, but memory
+    use must stay bounded even for very large gazetteers (tens of millions of
+    rows). All temporary data spills to disk rather than RAM.
+
+    Args:
+        path: Path of the artifact file to create
+
+    Returns:
+        SQLite connection to the new database
+    """
+    if path.exists():
+        path.unlink()
+    connection = sqlite3.connect(path, isolation_level=None)
+    connection.execute("PRAGMA synchronous=OFF")
+    # No rollback journal: avoids buffering the whole bulk-load transaction in
+    # memory, which would otherwise grow without bound on large builds.
+    connection.execute("PRAGMA journal_mode=OFF")
+    # Spill temporary data (index sorts, FTS rebuild, VACUUM) to disk, not RAM.
+    connection.execute("PRAGMA temp_store=FILE")
+    # Cap the page cache so memory stays bounded regardless of data size
+    # (negative value = cache size in KiB, here 64 MiB).
+    connection.execute("PRAGMA cache_size=-65536")
+    for statement in artifact.BASE_SCHEMA:
+        connection.execute(statement)
+    return connection
+
+
+def copy_rows(
+    duckdb_connection: duckdb.DuckDBPyConnection,
+    sqlite_connection: sqlite3.Connection,
+    select_sql: str,
+    insert_sql: str,
+) -> int:
+    """
+    Copy rows from a DuckDB query into a SQLite table in batches.
+
+    Args:
+        duckdb_connection: DuckDB connection to read from
+        sqlite_connection: SQLite connection to write to
+        select_sql: DuckDB query producing the rows
+        insert_sql: SQLite INSERT statement with positional placeholders
+
+    Returns:
+        Number of copied rows
+    """
+    cursor = duckdb_connection.execute(select_sql)
+    copied = 0
+    sqlite_connection.execute("BEGIN")
+    while True:
+        batch = cursor.fetchmany(BATCH_SIZE)
+        if not batch:
+            break
+        sqlite_connection.executemany(insert_sql, batch)
+        copied += len(batch)
+    sqlite_connection.execute("COMMIT")
+    return copied
+
+
+def emit(
+    duckdb_connection: duckdb.DuckDBPyConnection,
+    sqlite_connection: sqlite3.Connection,
+) -> t.Tuple[int, int]:
+    """
+    Copy the projected features and names into the artifact.
+
+    Expects the DuckDB build tables ``_features_final`` (id, identifier, type,
+    attributes, geometry) and ``_names_final`` (feature_id, text) to exist.
+
+    Args:
+        duckdb_connection: DuckDB connection holding the build tables
+        sqlite_connection: SQLite connection of the artifact being built
+
+    Returns:
+        Tuple of (feature count, name count)
[…]
