+++ geoparser/gazetteer/build/__init__.py
@@ -1,0 +1,3 @@
+from geoparser.gazetteer.build.builder import GazetteerBuilder
+
+__all__ = ["GazetteerBuilder"]

[…]
