### 2026-07-10 62b2406
Refactor progress tracking in gazetteer build pipeline
