+++ geoparser/gazetteer/build/progress.py
@@ -1,22 +1,135 @@
 """
 Console output helpers for the gazetteer build pipeline.
+
+Build activity is grouped by pipeline stage (acquire, stage, project, emit,
+finalize), mirroring how the old installer grouped progress by source: each
+stage owns a persistent bar that reads "<verb>ing ..." while it runs and
+"<verb>ed ..." once every item in it has completed, turning from cyan to
+green. Individual items within a stage (a download, a staged input, a
+projected feature type, ...) render as dimmed, indented bars nested under
+their stage and disappear as soon as they finish, so only the stage bars and
+the currently active item remain on screen. There are no checkmarks and no
+byte counts for downloads; progress within an item is shown as a percentage
+(or, for indeterminate work, an animated bar with no percentage).
 """
 
 import typing as t
 from contextlib import contextmanager
+from contextvars import ContextVar
 
 from rich.console import Console
-from rich.progress import (
-    BarColumn,
-    DownloadColumn,
-    Progress,
-    TextColumn,
-    TimeElapsedColumn,
-)
+from rich.progress import BarColumn, Progress, ProgressColumn, Task, TimeElapsedColumn
 from rich.rule import Rule
-from rich.table import Table
+from rich.spinner import Spinner
+from rich.table import Column, Table
+from rich.text import Text
+
+# Connector that visually nests an item (child) bar under its stage.
+_CHILD_CONNECTOR = "  └─ "
+_CHILD_SPINNER_NAME = "dots"
+
+# Status colors. A stage is "in progress" (cyan) until all of its items
+# complete, then "done" (green); item bars always use the in-progress color
+# since they disappear on completion.
+_RUNNING_STYLE = "cyan"
+_DONE_STYLE = "green"
 
 _console = Console(stderr=True)
+
+# The Progress shared by every stage and item of the current build.
+_active_progress: ContextVar[t.Optional[Progress]] = ContextVar(
+    "active_build_progress", default=None
+)
+# The stage currently accepting items, so item() can attach without callers
+# threading a stage object through unrelated modules (e.g. acquire.py).
+_active_stage: ContextVar[t.Optional["Stage"]] = ContextVar(
+    "active_build_stage", default=None
+)
+
+
+class _DescriptionColumn(ProgressColumn):
+    """
+    Render stage and item descriptions so the two are easy to tell apart.
+
+    A stage task is shown as its running label in bold cyan while any of its
+    items are outstanding, then its done label in bold green once complete.
+    An item task is dimmed and nested under its stage with a tree connector
+    and an animated spinner while it runs.
+    """
+
+    def __init__(self, width: int) -> None:
+        super().__init__(
+            table_column=Column(width=width, no_wrap=True, overflow="ellipsis")
+        )
+        self._child_spinner = Spinner(_CHILD_SPINNER_NAME, style="dim")
+
+    def render(self, task: Task) -> Text:
+        if task.fields.get("is_stage"):
+            if task.finished:
+                return Text(str(task.fields["done_label"]), style=f"bold {_DONE_STYLE}")
+            return Text(str(task.description), style=f"bold {_RUNNING_STYLE}")
+        line = Text(_CHILD_CONNECTOR, style="dim")
+        line.append_text(self._child_spinner.render(task.get_time()))
+        line.append(" ")
+        line.append(str(task.description), style="dim")
+        return line
+
+
+class _ProgressReadoutColumn(ProgressColumn):
+    """
+    Render the progress readout for a task.
+
+    A stage task is shown as a "completed/total" item count; an item task is
+    shown as a percentage when it tracks a known quantity, or left blank for
+    indeterminate work (its bar animates instead).
+    """
+
+    def render(self, task: Task) -> Text:
[…]
