### 2026-07-10 f0ecb17
Simplify gazetteer configuration schema by moving to more SQL-driven definitions
