+++ geoparser/gazetteer/build/stage.py
@@ -1,7 +1,7 @@
 """
-Staging of gazetteer input files into transient DuckDB tables.
+Staging of gazetteer source files into transient DuckDB tables.
 
-Tabular inputs (those with a ``delimiter``) are loaded with DuckDB's CSV
+Tabular sources (those with a ``delimiter``) are loaded with DuckDB's CSV
 reader; everything else is loaded with the spatial extension's ``ST_Read``
 (shapefiles, GeoPackage, GeoJSON, ...). Staged spatial tables always expose
 their geometry column under the name ``geometry``.
@@ -12,11 +12,12 @@ from pathlib import Path
 
 import duckdb
 
-from geoparser.gazetteer.config import DataType, InputConfig
+from geoparser.gazetteer.config import DataType, SourceConfig
+from geoparser.gazetteer.config.schema import GEOMETRY_ATTRIBUTE
 
-GEOMETRY_COLUMN = "geometry"
+GEOMETRY_COLUMN = GEOMETRY_ATTRIBUTE
 
-# Mapping from config column types to DuckDB types
+# Mapping from config attribute types to DuckDB types
 _DUCKDB_TYPES = {
     DataType.TEXT: "VARCHAR",
     DataType.INTEGER: "BIGINT",
@@ -64,22 +65,22 @@ class Stager:
         """
         self.connection = connection
 
-    def stage(self, input_config: InputConfig, file_path: Path) -> int:
+    def stage(self, source_config: SourceConfig, file_path: Path) -> int:
         """
-        Load an input file into a staging table named after the input.
+        Load a source file into a staging table named after the source.
 
         Args:
-            input_config: Input configuration
-            file_path: Path to the resolved input file
+            source_config: Source configuration
+            file_path: Path to the resolved source file
 
         Returns:
             Number of staged rows
         """
-        if input_config.is_tabular:
-            self._stage_tabular(input_config, file_path)
+        if source_config.is_tabular:
+            self._stage_tabular(source_config, file_path)
         else:
-            self._stage_spatial(input_config, file_path)
-        table = quote_identifier(input_config.name)
+            self._stage_spatial(source_config, file_path)
+        table = quote_identifier(source_config.name)
         return self.connection.execute(f"SELECT count(*) FROM {table}").fetchone()[0]
 
     def columns(self, table_name: str) -> t.List[str]:
@@ -99,41 +100,44 @@ class Stager:
         ).fetchall()
         return [row[0] for row in rows]
 
-    def _stage_tabular(self, input_config: InputConfig, file_path: Path) -> None:
+    def _stage_tabular(self, source_config: SourceConfig, file_path: Path) -> None:
         """Load a delimited text file via DuckDB's CSV reader."""
         options = [
-            f"delim={quote_literal(input_config.delimiter)}",
-            f"skip={input_config.skip_rows}",
+            f"delim={quote_literal(source_config.delimiter)}",
+            f"skip={source_config.skip_rows}",
+            "header=false",
             "null_padding=true",
         ]
-        quote = input_config.quote if input_config.quote is not None else '"'
+        quote = source_config.quote if source_config.quote is not None else '"'
         options.append(f"quote={quote_literal(quote)}")
         if not quote:
             options.append("escape=''")
 
-        if input_config.columns is not None:
-            column_spec = ", ".join(
-                f"{quote_literal(column.name)}: "
-                f"{quote_literal(_DUCKDB_TYPES[column.type])}"
-                for column in input_config.columns
-            )
-            options.append("header=false")
-            options.append(f"columns={{{column_spec}}}")
-        else:
-            options.append("header=true")
+        column_spec = ", ".join(
+            f"{quote_literal(attribute.name)}: "
[…]
