+++ geoparser/gazetteer/config/__init__.py
@@ -1,37 +1,17 @@
 from geoparser.gazetteer.config.schema import (
-    AttributeConfig,
-    AttributeMerge,
-    ColumnDef,
+    AttributeDef,
+    DataConfig,
     DataType,
     FeatureConfig,
     GazetteerConfig,
-    GeometryConfig,
-    GeometryMerge,
-    GeometryTransform,
-    InputConfig,
-    LookupConfig,
-    MatchConfig,
-    MergeConfig,
-    NameConfig,
-    PointConfig,
-    SpatialPredicate,
+    SourceConfig,
 )
 
 __all__ = [
-    "AttributeConfig",
-    "AttributeMerge",
-    "ColumnDef",
+    "AttributeDef",
+    "DataConfig",
     "DataType",
     "FeatureConfig",
     "GazetteerConfig",
-    "GeometryConfig",
-    "GeometryMerge",
-    "GeometryTransform",
-    "InputConfig",
-    "LookupConfig",
-    "MatchConfig",
-    "MergeConfig",
-    "NameConfig",
-    "PointConfig",
-    "SpatialPredicate",
+    "SourceConfig",
 ]
