### 2026-07-10 2657760
Update swissnames3d.yaml to change data types for GEBAEUDENU and EINWOHNERK from integer to text
