### 2026-07-10 95f430d
Refactor gazetteer configuration to use SQL-style data definitions
