+++ geoparser/gazetteer/config/schema.py
@@ -12,10 +12,13 @@ feature store. It has two top-level concepts:
   ``join`` other sources to enrich its rows. A feature's ``source`` in the
   artifact is the name of the source it is built from.
 
-Values (``identifier``, ``names``, ``geometry`` and each ``data`` attribute)
-are column references or scalar SQL expressions evaluated over the block's
-source, aliased ``src``; a bare column name refers to that source, and columns
-of joined sources are referenced by qualification (``<source>.<column>``).
+Values (``identifier``, ``names``, ``geometry`` and each ``data`` entry) are
+column references or scalar SQL expressions evaluated over the block's source,
+aliased ``src``; a bare column name refers to that source, and columns of
+joined sources are referenced by qualification (``<source>.<column>``). A
+``data`` entry names the key it is stored under with a trailing ``AS <alias>``
+(as in a SQL ``SELECT``); a bare or qualified column reference may omit it, in
+which case the column's own name is the key.
 
 Joins are written as raw SQL join clauses (e.g.
 ``"LEFT JOIN countryInfo ON src.country_code = countryInfo.ISO"``); the whole
@@ -164,40 +167,49 @@ class SourceConfig(BaseModel):
         return self
 
 
-class DataConfig(BaseModel):
+# Matches a bare or dot-qualified identifier (``NAME``, ``g.NAME``), each
+# component optionally double-quoted (``"My Column".other``).
+_QUALIFIED_IDENTIFIER = re.compile(
+    r'^(?:"[^"]+"|[A-Za-z_][A-Za-z0-9_]*)'
+    r'(?:\.(?:"[^"]+"|[A-Za-z_][A-Za-z0-9_]*))*$'
+)
+
+# Matches a trailing SQL alias, as in ``SELECT <expr> AS <alias>``.
+_ALIAS_SUFFIX = re.compile(
+    r'\s+AS\s+("[^"]+"|[A-Za-z_][A-Za-z0-9_]*)\s*$', re.IGNORECASE
+)
+
+
+def split_data_value(value: str) -> t.Tuple[str, t.Optional[str]]:
     """
-    A value stored in the feature's JSON data object.
+    Split a ``data`` entry into its SQL expression and output key.
 
-    ``attribute`` is a column of the block's source (bare), a column of a joined
-    source (qualified as ``<source>.<column>``) or a scalar SQL expression;
-    ``alias`` is the key it is stored under. When ``alias`` is omitted,
-    ``attribute`` must be a bare column of the block's source and is used as the
-    key. The string shorthand ``"name"`` is equivalent to ``{attribute: name}``.
+    A trailing ``AS <alias>`` (as in a SQL ``SELECT``) names the key
+    explicitly. Otherwise the entry must be a bare or dot-qualified column
+    reference (``NAME``, ``g.NAME``), whose last component names the key; a
+    plain expression without ``AS`` has no derivable key, so its key is
+    ``None``.
+
+    Args:
+        value: Raw ``data`` entry, e.g. ``"NAME"`` or ``"g.NAME AS GEMEINDE_NAME"``
+
+    Returns:
+        Tuple of (SQL expression, output key or None)
     """
+    match = _ALIAS_SUFFIX.search(value)
+    if match:
+        alias = match.group(1)
+        if alias.startswith('"'):
+            alias = alias[1:-1]
+        return value[: match.start()].strip(), alias
 
-    attribute: str
-    alias: t.Optional[str] = None
-
-    @property
-    def output_name(self) -> str:
-        """The key the value is stored under."""
-        return self.alias or self.attribute
-
-    @model_validator(mode="before")
-    @classmethod
-    def coerce_shorthand(cls, value: t.Any) -> t.Any:
-        if isinstance(value, str):
[…]
