+++ tests/integration/test_gazetteer/test_build_integration.py
@@ -132,8 +132,8 @@ class TestDuplicateIdentifierMerge:
                     names:
                       - "name"
                     data:
-                      - attribute: "name"
-                      - attribute: "height"
+                      - "name"
+                      - "height"
                 """
             )
         )
