+++ geoparser/gazetteer/configs/geonames.yaml
@@ -152,28 +152,24 @@ features:
       - "CASE WHEN instr(asciiname, '(') > 0 THEN trim(substr(asciiname, 1, instr(asciiname, '(') - 1)) ELSE asciiname END"
       - "unnest(string_split(alternatenames, ','))"
     data:
-      - attribute: "name"
-      - attribute: "asciiname"
-      - attribute: "latitude"
-      - attribute: "longitude"
-      - attribute: "feature_class"
-      - attribute: "feature_code"
-      - attribute: "featureCodes_en.name"
-        alias: "feature_name"
-      - attribute: "country_code"
-      - attribute: "countryInfo.Country"
-        alias: "country_name"
-      - attribute: "cc2"
-      - attribute: "admin1_code"
-      - attribute: "admin1CodesASCII.name"
-        alias: "admin1_name"
-      - attribute: "admin2_code"
-      - attribute: "admin2Codes.name"
-        alias: "admin2_name"
-      - attribute: "admin3_code"
-      - attribute: "admin4_code"
-      - attribute: "population"
-      - attribute: "elevation"
-      - attribute: "dem"
-      - attribute: "timezone"
-      - attribute: "modification_date"
+      - "name"
+      - "asciiname"
+      - "latitude"
+      - "longitude"
+      - "feature_class"
+      - "feature_code"
+      - "featureCodes_en.name AS feature_name"
+      - "country_code"
+      - "countryInfo.Country AS country_name"
+      - "cc2"
+      - "admin1_code"
+      - "admin1CodesASCII.name AS admin1_name"
+      - "admin2_code"
+      - "admin2Codes.name AS admin2_name"
+      - "admin3_code"
+      - "admin4_code"
+      - "population"
+      - "elevation"
+      - "dem"
+      - "timezone"
+      - "modification_date"

[…]
