### 2026-07-11 138c2d3
Enhance progress tracking in gazetteer build pipeline with detailed item bars and improved query progress reporting
