+++ geoparser/gazetteer/build/builder.py
@@ -29,11 +29,14 @@ from geoparser.gazetteer.build.acquire import Acquirer
 from geoparser.gazetteer.build.compile import ProjectionCompiler
 from geoparser.gazetteer.build.emit import create_artifact_db, emit, finalize
 from geoparser.gazetteer.build.progress import (
+    allocate,
     build_display,
     item,
     print_build_header,
     print_build_summary,
+    scale_poll,
     stage,
+    track,
 )
 from geoparser.gazetteer.build.stage import Stager, quote_literal
 from geoparser.gazetteer.config import GazetteerConfig
@@ -77,6 +80,15 @@ class GazetteerBuilder:
     # slow.
     _MIN_MEMORY_MB = 512
 
+    # Relative weight given to each part of a feature's projection when
+    # splitting one item bar's range between them (see progress.allocate):
+    # the feature query does the heavy join/aggregation, name queries are
+    # cheap selects over the same staged source, and merging duplicate
+    # geometries touches only the few rows that actually duplicate.
+    _FEATURE_QUERY_WEIGHT = 6
+    _NAME_QUERY_WEIGHT = 2
+    _MERGE_WEIGHT = 1
+
     def build(
         self, config_path: t.Union[str, Path], keep_downloads: bool = False
     ) -> Path:
@@ -111,7 +123,9 @@ class GazetteerBuilder:
                 try:
                     self._configure_staging(connection, build_dir)
                     if self._needs_spatial(config):
-                        self._load_spatial_extension(connection)
+                        with item("Loading spatial extension", total=100) as bar:
+                            self._load_spatial_extension(connection)
+                            bar.set_progress(100)
                     self._stage_inputs(connection, config, file_paths)
                     self._project(connection, config)
                     feature_count, name_count = self._emit_artifact(
@@ -145,12 +159,19 @@ class GazetteerBuilder:
             connection: DuckDB connection used for staging and projection
             build_dir: Directory holding the transient build files
         """
-        # DuckDB prints its own progress bar straight to the terminal for
-        # long-running queries by default. It writes outside of Rich's
-        # control, which desyncs the live build display (stray bars flashing
-        # in, stage rows appearing to repeat); our own progress reporting
-        # already covers the whole build, so disable DuckDB's.
-        connection.execute("SET enable_progress_bar = false")
+        # DuckDB tracks query progress internally only while its progress bar
+        # feature is enabled. By default it also prints that progress straight
+        # to the terminal, outside Rich's control, which would desync the live
+        # build display (stray bars flashing in, stage rows appearing to
+        # repeat); ``enable_progress_bar_print`` keeps the tracking without
+        # the printing, so `connection.query_progress()` (used to drive our
+        # own item bars, see ``_project`` and ``_stage_inputs``) works while
+        # the terminal stays under our control. ``progress_bar_time = 0``
+        # makes tracking start immediately rather than after DuckDB's default
+        # delay for what it guesses will be a short query.
+        connection.execute("SET enable_progress_bar = true")
+        connection.execute("SET enable_progress_bar_print = false")
+        connection.execute("SET progress_bar_time = 0")
 
         temp_dir = build_dir / "duckdb-temp"
         temp_dir.mkdir(parents=True, exist_ok=True)
@@ -226,8 +247,13 @@ class GazetteerBuilder:
         stager = Stager(connection)
         with stage("Staging sources", "Staged sources", len(config.sources)) as group:
             for source_config in config.sources:
-                with item(f"Staging source '{source_config.name}'"):
-                    stager.stage(source_config, file_paths[source_config.name])
+                with item(f"Staging source '{source_config.name}'", total=100) as bar:
+                    # Stager reports its own progress: a spatial source runs
[…]

[…]
