### 2026-07-11 8ce5e48
Improve progress tracking and reporting for gazetteer installation pipeline
