+++ geoparser/gazetteer/build/progress.py
@@ -1,16 +1,32 @@
 """
 Console output helpers for the gazetteer build pipeline.
 
-Build activity is grouped by pipeline stage (acquire, stage, project, emit,
-finalize), mirroring how the old installer grouped progress by source: each
-stage owns a persistent bar that reads "<verb>ing ..." while it runs and
-"<verb>ed ..." once every item in it has completed, turning from cyan to
-green. Individual items within a stage (a download, a staged input, a
-projected feature type, ...) render as dimmed, indented bars nested under
-their stage and disappear as soon as they finish, so only the stage bars and
-the currently active item remain on screen. There are no checkmarks and no
-byte counts for downloads; progress within an item is shown as a percentage
-(or, for indeterminate work, an animated bar with no percentage).
+Build activity is grouped by pipeline stage (preparing sources, deriving
+features, building the artifact; see :mod:`builder`), mirroring how the old
+installer grouped progress by source: each stage owns a persistent bar that
+reads "<verb>ing ..." while it runs and "<verb>ed ..." once every item in it
+has completed, turning from cyan to green. Individual items within a stage (a
+download, a staged source, a query run against a feature block, ...) render
+as dimmed, indented bars nested under their stage and disappear as soon as
+they finish, so only the stage bars and the currently active item remain on
+screen. There are no checkmarks and no byte counts for downloads; progress
+within an item is shown as a percentage (or, for indeterminate work, an
+animated bar with no percentage). A determinate item that finishes
+successfully is snapped to 100% and held there briefly (see
+:data:`_COMPLETION_PAUSE_SECONDS`) before it disappears, so it visibly
+completes rather than seeming to vanish mid-step, whether it was polling
+normally or sitting at some percentage right up to the end. A failed item
+disappears immediately instead, since there is no completed state to show.
+
+Every item that appears leaves a mark on its stage: the stage advances by one
+the moment its item disappears, so the stage's own percentage always
+reflects real, finished work rather than jumping only once several items
+have quietly come and gone. A stage's ``total_items`` is a best
+estimate (the exact number of items some units of work will show, e.g. an
+acquisition's download-then-extract, is only known once it runs); the count
+grows on the fly if advances outrun it, so the displayed total never falls
+behind reality, and is snapped to match whatever actually completed once the
+stage exits.
 
 Database operations (a single DuckDB query, or a batch of SQLite statements)
 report a percentage too, using :func:`track`, which runs the operation on a
@@ -21,13 +37,14 @@ track progress for every query shape, and statement-count progress does not
 account for statements taking unequal time, but they give a useful sense of
 motion for otherwise silent, long-running steps. Their items are always
 created determinate (``total=100``) so they read "0%" immediately rather than
-ever showing an animated, indeterminate bar, even briefly. When an item
-covers several queries in sequence, :func:`scale_poll` and :func:`allocate`
-give each query its own slice of the item's 0-100 range, so the bar only
-ever climbs, never resetting to a lower number when the next query starts.
+ever showing an animated, indeterminate bar, even briefly. Each individual
+query gets its own item, shown and disposed of in turn, rather than several
+queries sharing one bar (which would either not reflect their real, separate
+progress or require guessing at how to divide the bar between them).
 """
 
 import threading
+import time
 import typing as t
 from contextlib import contextmanager
 from contextvars import ContextVar
@@ -42,6 +59,12 @@ from rich.text import Text
 # Connector that visually nests an item (child) bar under its stage.
 _CHILD_CONNECTOR = "  └─ "
 _CHILD_SPINNER_NAME = "dots"
+
[…]
