### 2026-07-11 9e6826b
Add label suffixing to progress bars with identical descriptions
