### 2026-07-11 42adb32
Rename gazetteer build pipeline stages and reorganize related files
