+++ geoparser/gazetteer/config/__init__.py
@@ -1,17 +1,0 @@
-from geoparser.gazetteer.config.schema import (
-    AttributeDef,
-    DataType,
-    FeatureConfig,
-    GazetteerConfig,
-    SourceConfig,
-    split_data_value,
-)
-
-__all__ = [
-    "AttributeDef",
-    "DataType",
-    "FeatureConfig",
-    "GazetteerConfig",
-    "SourceConfig",
-    "split_data_value",
-]

[…]
