### 2026-07-11 f8eedff
Add unit tests for gazetteer build stages and new gazetteer modules
