+++ tests/unit/test_gazetteer/test_build/test_builder.py
@@ -1,0 +1,174 @@
+"""
+Unit tests for geoparser/gazetteer/build/builder.py
+
+Exercises GazetteerBuilder's internal helpers directly (memory sizing,
+temp-directory handling, spatial extension loading) and runs small,
+hand-crafted builds to trigger error paths (no features produced, a stale
+leftover build directory) that are impractical to reach through the
+public build() entry point alone.
+"""
+
+import os
+import textwrap
+from unittest.mock import MagicMock
+
+import duckdb
+import pytest
+
+from geoparser.gazetteer import artifact
+from geoparser.gazetteer.build.builder import GazetteerBuilder, _sqlite_tmpdir
+from geoparser.gazetteer.build.schema import GazetteerConfig
+
+
+@pytest.mark.unit
+class TestSqliteTmpdir:
+    """Test the _sqlite_tmpdir() context manager."""
+
+    def test_sets_and_restores_previous_value(self, monkeypatch, tmp_path):
+        """A pre-existing SQLITE_TMPDIR is restored after the context exits."""
+        monkeypatch.setenv("SQLITE_TMPDIR", "/original/tmpdir")
+
+        with _sqlite_tmpdir(tmp_path):
+            assert os.environ["SQLITE_TMPDIR"] == str(tmp_path)
+
+        assert os.environ["SQLITE_TMPDIR"] == "/original/tmpdir"
+
+    def test_removes_var_when_none_was_set_before(self, monkeypatch, tmp_path):
+        """SQLITE_TMPDIR is removed again if it wasn't set beforehand."""
+        monkeypatch.delenv("SQLITE_TMPDIR", raising=False)
+
+        with _sqlite_tmpdir(tmp_path):
+            assert os.environ["SQLITE_TMPDIR"] == str(tmp_path)
+
+        assert "SQLITE_TMPDIR" not in os.environ
+
+
+@pytest.mark.unit
+class TestMemoryLimit:
+    """Test GazetteerBuilder._memory_limit_mb() and _physical_memory_bytes()."""
+
+    def test_returns_none_when_physical_memory_is_unknown(self, monkeypatch):
+        """No memory limit is set when the machine's RAM can't be determined."""
+        builder = GazetteerBuilder()
+        monkeypatch.setattr(builder, "_physical_memory_bytes", lambda: None)
+
+        assert builder._memory_limit_mb() is None
+
+    def test_returns_a_bounded_value_when_memory_is_known(self, monkeypatch):
+        """A known RAM size yields a positive, bounded MiB limit."""
+        builder = GazetteerBuilder()
+        monkeypatch.setattr(
+            builder, "_physical_memory_bytes", lambda: 8 * 1024 * 1024 * 1024
+        )
+
+        assert builder._memory_limit_mb() > 0
+
+    def test_physical_memory_bytes_returns_none_on_error(self, monkeypatch):
+        """A sysconf() failure is treated as 'unknown', not a crash."""
+
+        def _raise(_name):
+            raise OSError("not supported")
+
+        monkeypatch.setattr(os, "sysconf", _raise)
+
+        assert GazetteerBuilder._physical_memory_bytes() is None
+
+
+@pytest.mark.unit
+class TestLoadSpatialExtension:
+    """Test GazetteerBuilder._load_spatial_extension()."""
+
+    def test_wraps_duckdb_errors_in_a_clear_runtime_error(self):
+        """A failure to install/load the extension raises a clear RuntimeError."""
+        connection = MagicMock()
+        connection.install_extension.side_effect = duckdb.Error("network unreachable")
+        builder = GazetteerBuilder()
+
+        with pytest.raises(
+            RuntimeError, match="Failed to load the DuckDB spatial extension"
+        ):
[…]
