+++ tests/unit/test_gazetteer/test_build/test_stages/test_load.py
@@ -1,0 +1,244 @@
+"""
+Unit tests for geoparser/gazetteer/build/stages/load.py
+
+Tests the Loader against small tabular and spatial fixture files, and the
+error paths around a spatial source's declared geometry column.
+"""
+
+import json
+from pathlib import Path
+
+import duckdb
+import pytest
+
+from geoparser.gazetteer.build.schema import SourceConfig
+from geoparser.gazetteer.build.stages.load import (
+    Loader,
+    quote_identifier,
+    quote_literal,
+)
+
+
+def make_tabular_source(**overrides) -> SourceConfig:
+    data = {
+        "name": "places",
+        "path": "/tmp/unused.csv",
+        "file": "places.csv",
+        "delimiter": ",",
+        "attributes": [
+            {"name": "id", "type": "integer"},
+            {"name": "name", "type": "text"},
+        ],
+    }
+    data.update(overrides)
+    return SourceConfig.model_validate(data)
+
+
+def make_spatial_source(**overrides) -> SourceConfig:
+    data = {
+        "name": "shape",
+        "path": "/tmp/unused.geojson",
+        "file": "shape.geojson",
+        "attributes": [
+            {"name": "id", "type": "integer"},
+            {"name": "geometry", "type": "geometry"},
+        ],
+    }
+    data.update(overrides)
+    return SourceConfig.model_validate(data)
+
+
+@pytest.fixture
+def connection():
+    con = duckdb.connect()
+    con.load_extension("spatial")
+    yield con
+    con.close()
+
+
+@pytest.fixture
+def loader(connection) -> Loader:
+    return Loader(connection)
+
+
+class _FakeGeometryColumnsResult:
+    """A minimal stand-in for a DuckDB result set."""
+
+    def __init__(self, rows):
+        self._rows = rows
+
+    def fetchall(self):
+        return self._rows
+
+
+class _FakeMultiGeometryConnection:
+    """
+    Wraps a real DuckDB connection but reports two geometry columns.
+
+    Used to exercise Loader's "multiple geometry columns" error path, which
+    no supported file format naturally produces (every GDAL driver exposes
+    exactly one geometry field per layer).
+    """
+
+    def __init__(self, real: duckdb.DuckDBPyConnection):
+        self._real = real
+
+    def execute(self, sql, *args, **kwargs):
+        if "data_type LIKE 'GEOMETRY%'" in sql:
+            return _FakeGeometryColumnsResult([("geometry",), ("geometry2",)])
+        return self._real.execute(sql, *args, **kwargs)
+
+    def __getattr__(self, name):
+        return getattr(self._real, name)
+
+
+@pytest.mark.unit
+class TestQuoting:
+    """Test quote_identifier() and quote_literal()."""
+
+    def test_quote_identifier_escapes_embedded_quotes(self):
+        assert quote_identifier('my "column"') == '"my ""column"""'
+
+    def test_quote_literal_escapes_embedded_quotes(self):
+        assert quote_literal("it's") == "'it''s'"
+
+
+@pytest.mark.unit
+class TestLoadTabular:
+    """Test Loader.load() for tabular (delimited) sources."""
+
+    def test_loads_rows_with_declared_types(self, loader, tmp_path):
+        """A tabular file is loaded into a table matching its declared schema."""
+        data_file = tmp_path / "places.csv"
+        data_file.write_text("1,Paris\n2,Berlin\n")
+        source = make_tabular_source()
+
+        row_count = loader.load(source, data_file)
+
+        assert row_count == 2
+        assert loader.columns("places") == ["id", "name"]
+
+    def test_respects_skip_rows(self, loader, tmp_path):
[…]
