+++ tests/unit/test_gazetteer/test_build/test_stages/test_compile.py
@@ -23,8 +23,7 @@ def build_compiler(config_data: dict) -> t.Tuple[GazetteerConfig, ProjectionComp
     """Validate a config and build a compiler with a catalog from its sources."""
     config = GazetteerConfig.model_validate(config_data)
     catalog = {
-        source.name: [a.name for a in source.attributes]
-        for source in config.sources
+        source.name: [a.name for a in source.attributes] for source in config.sources
     }
     return config, ProjectionCompiler(config, catalog)
 
@@ -67,9 +66,7 @@ def run_features(
         merged_geometry = {}
         duplicate_query = compiler.duplicate_geometry_query(feature)
         if duplicate_query is not None:
-            for identifier, geometry in connection.execute(
-                duplicate_query
-            ).fetchall():
+            for identifier, geometry in connection.execute(duplicate_query).fetchall():
                 merged_geometry[identifier] = geometry
         for identifier, source, data, geometry in connection.execute(
             compiler.feature_query(feature)
@@ -127,8 +124,12 @@ class TestBasicProjection:
         return {
             "name": "testgaz",
             "sources": [
-                tabular("places", ("id", "integer"), ("name", "text"),
-                        ("population", "integer"))
+                tabular(
+                    "places",
+                    ("id", "integer"),
+                    ("name", "text"),
+                    ("population", "integer"),
+                )
             ],
             "features": [feature],
         }
@@ -156,9 +157,7 @@ class TestBasicProjection:
     def test_name_expression(self, connection):
         """Names can be scalar SQL expressions."""
         self.make_places(connection)
-        config, compiler = build_compiler(
-            self.config_data(names=["upper(name)"])
-        )
+        config, compiler = build_compiler(self.config_data(names=["upper(name)"]))
 
         names = run_names(connection, compiler, config)
 
@@ -225,8 +224,9 @@ class TestSplitNames:
         config, compiler = build_compiler(
             {
                 "name": "testgaz",
-                "sources": [tabular("places", ("id", "integer"),
-                                    ("alternates", "text"))],
+                "sources": [
+                    tabular("places", ("id", "integer"), ("alternates", "text"))
+                ],
                 "features": [
                     {
                         "source": "places",
@@ -264,8 +264,11 @@ class TestDuplicateMerge:
         }
         return {
             "name": "testgaz",
-            "sources": [tabular("places", ("id", "integer"), ("name", "text"),
-                                ("height", "integer"))],
+            "sources": [
+                tabular(
+                    "places", ("id", "integer"), ("name", "text"), ("height", "integer")
+                )
+            ],
             "features": [feature],
         }
 
@@ -313,8 +316,15 @@ class TestGeometry:
         config, compiler = build_compiler(
             {
                 "name": "testgaz",
-                "sources": [tabular("places", ("id", "integer"), ("name", "text"),
-                                    ("lon", "real"), ("lat", "real"))],
+                "sources": [
+                    tabular(
+                        "places",
+                        ("id", "integer"),
+                        ("name", "text"),
+                        ("lon", "real"),
+                        ("lat", "real"),
+                    )
+                ],
                 "features": [
                     {
                         "source": "places",
@@ -347,8 +357,14 @@ class TestGeometry:
         config, compiler = build_compiler(
             {
                 "name": "testgaz",
-                "sources": [spatial("places", ("id", "integer"), ("name", "text"),
-                                    ("geometry", "geometry"))],
+                "sources": [
+                    spatial(
+                        "places",
+                        ("id", "integer"),
+                        ("name", "text"),
+                        ("geometry", "geometry"),
+                    )
+                ],
                 "features": [
                     {
                         "source": "places",
[…]
