### 2026-07-11 447ff4c
Fix handling of sysconf for Windows in memory limit tests and update absolute path handling in YAML configuration tests
