### 2026-07-18 56994f7
Add SpanEncoderResolver module for improved mention-aware toponym disambiguation
