+++ tests/unit/test_modules/test_resolvers/test_spanencoder.py
@@ -1,0 +1,153 @@
+"""
+Unit tests for geoparser/modules/resolvers/spanencoder.py
+
+Tests span/token utilities and serialization with mocked heavy dependencies.
+"""
+
+from unittest.mock import Mock, patch
+
+import pytest
+import torch
+
+
+def _make_resolver(**kwargs):
+    from geoparser.modules.resolvers.spanencoder import SpanEncoderResolver
+
+    with (
+        patch("geoparser.modules.resolvers.spanencoder.AutoConfig"),
+        patch("geoparser.modules.resolvers.spanencoder.AutoModel") as mock_model,
+        patch("geoparser.modules.resolvers.spanencoder.AutoTokenizer"),
+        patch("geoparser.modules.resolvers.spanencoder.Gazetteer"),
+    ):
+        mock_model.from_pretrained.return_value.config.hidden_size = 768
+        mock_model.from_pretrained.return_value.to = Mock()
+        return SpanEncoderResolver(**kwargs)
+
+
+@pytest.mark.unit
+class TestSpanEncoderResolverInitialization:
+    def test_creates_with_default_parameters(self):
+        resolver = _make_resolver()
+
+        assert resolver.name == "SpanEncoderResolver"
+        assert resolver.model_name == "answerdotai/ModernBERT-base"
+        assert resolver.gazetteer_name == "geonames"
+        assert resolver.min_similarity == 0.6
+        assert resolver.max_tiers == 3
+        assert resolver.embedding_dim == 256
+
+    def test_unknown_gazetteer_without_map_raises(self):
+        with pytest.raises(ValueError, match="not configured"):
+            _make_resolver(gazetteer_name="unknown")
+
+    def test_custom_attribute_map(self):
+        custom_map = {"name": "n", "type": "t", "level1": "l1"}
+        resolver = _make_resolver(
+            gazetteer_name="custom", attribute_map=custom_map
+        )
+
+        assert resolver.attribute_map == custom_map
+
+
+@pytest.mark.unit
+class TestCharSpanToTokenSpan:
+    def test_exact_alignment(self):
+        from geoparser.modules.resolvers.spanencoder import SpanEncoderResolver
+
+        # Tokens: [CLS] "Alexandria" "is" [SEP] with offsets
+        offsets = [(0, 0), (0, 10), (11, 13), (0, 0)]
+
+        assert SpanEncoderResolver._char_span_to_token_span(offsets, 0, 10) == (1, 1)
+
+    def test_multi_token_span(self):
+        from geoparser.modules.resolvers.spanencoder import SpanEncoderResolver
+
+        offsets = [(0, 0), (0, 7), (8, 14), (15, 20), (0, 0)]
+
+        assert SpanEncoderResolver._char_span_to_token_span(offsets, 0, 14) == (1, 2)
+
+    def test_partial_overlap_included(self):
+        from geoparser.modules.resolvers.spanencoder import SpanEncoderResolver
+
+        # A token partially overlapping the char span is included
+        offsets = [(0, 0), (0, 5), (5, 12), (0, 0)]
+
+        assert SpanEncoderResolver._char_span_to_token_span(offsets, 3, 8) == (1, 2)
+
+    def test_no_overlap_returns_none(self):
+        from geoparser.modules.resolvers.spanencoder import SpanEncoderResolver
+
+        offsets = [(0, 0), (0, 5), (0, 0)]
+
+        assert SpanEncoderResolver._char_span_to_token_span(offsets, 10, 15) is None
+
+
+@pytest.mark.unit
+class TestSerializeCandidate:
+    def test_full_attributes(self):
+        resolver = _make_resolver()
+        candidate = Mock()
+        candidate.data = {
+            "name": "Alexandria",
+            "feature_name": "populated place",
+            "country_name": "United States",
[…]
