+++ geoparser/modules/resolvers/spanencoder.py
@@ -1,0 +1,979 @@
+import json
+import random
+import typing as t
+from pathlib import Path
+from typing import Dict, List, Optional, Tuple, Union
+
+import torch
+import torch.nn.functional as F
+from torch import nn
+from transformers import AutoConfig, AutoModel, AutoTokenizer, logging
+
+from geoparser.gazetteer.gazetteer import Gazetteer
+from geoparser.modules.resolvers import Resolver
+
+if t.TYPE_CHECKING:
+    from geoparser.db.models.feature import Feature
+
+logging.set_verbosity_error()
+
+HEAD_FILE = "span_projection_head.pt"
+CONFIG_FILE = "span_encoder_config.json"
+
+
+class SpanEncoderResolver(Resolver):
+    """
+    A resolver that disambiguates the exact place mention using span pooling.
+
+    Both sides of the comparison are encoded by a single shared transformer:
+
+    - Mentions are embedded by encoding their document context and pooling the
+      hidden states of the mention's own tokens (start state, end state, and
+      span mean), so each mention in a document gets its own embedding.
+    - Gazetteer candidates are embedded by serializing their textual attributes
+      into a structured string (``name | type | admin3 > admin2 > admin1``) and
+      pooling the name's tokens the same way, with the remaining attributes
+      acting as context.
+
+    Resolution scores candidates by cosine similarity between the two
+    embeddings, mirroring the tiered candidate search of the original
+    SentenceTransformerResolver, and always assigns the best-scoring candidate
+    encountered if no candidate passes the similarity threshold.
+    """
+
+    NAME = "SpanEncoderResolver"
+
+    # Gazetteer-specific attribute mappings for entity serialization
+    GAZETTEER_ATTRIBUTE_MAP = {
+        "geonames": {
+            "name": "name",
+            "type": "feature_name",
+            "level1": "country_name",
+            "level2": "admin1_name",
+            "level3": "admin2_name",
+        },
+        "geonames-cities": {
+            "name": "name",
+            "type": "feature_name",
+            "level1": "country_name",
+            "level2": "admin1_name",
+            "level3": "admin2_name",
+        },
+        "swissnames3d": {
+            "name": "NAME",
+            "type": "OBJEKTART",
+            "level1": "KANTON_NAME",
+            "level2": "BEZIRK_NAME",
+            "level3": "GEMEINDE_NAME",
+        },
+    }
+
+    def __init__(
+        self,
+        model_name: str = "answerdotai/ModernBERT-base",
+        gazetteer_name: str = "geonames",
+        min_similarity: float = 0.6,
+        max_tiers: int = 3,
+        attribute_map: dict = None,
+        embedding_dim: int = 256,
+        max_context_tokens: int = 512,
+        max_document_tokens: int = 8192,
+    ):
+        """
+        Initialize the SpanEncoderResolver.
+
+        Args:
+            model_name: HuggingFace model name or path to a fine-tuned
+                SpanEncoderResolver directory
+            gazetteer_name: Name of the gazetteer to search
+            min_similarity: Similarity threshold used to stop candidate search
+                expansion (the best candidate overall is assigned regardless)
+            max_tiers: Maximum number of tiers to expand through search methods
+            attribute_map: Optional custom attribute mapping for the gazetteer
+            embedding_dim: Dimensionality of the shared embedding space
+            max_context_tokens: Token window around a mention used during
+                training and for documents longer than max_document_tokens
+            max_document_tokens: Maximum document length for single-pass encoding
+        """
+        super().__init__(
+            model_name=model_name,
+            gazetteer_name=gazetteer_name,
+            min_similarity=min_similarity,
+            max_tiers=max_tiers,
+            attribute_map=attribute_map,
+            embedding_dim=embedding_dim,
+            max_context_tokens=max_context_tokens,
[…]
