### 2026-07-19 55141db
Enhance memory management in GazetteerBuilder by adding fallback memory limits and thread count adjustments
