### 2026-07-19 5617d95
Enhance the _sqlite_tmpdir context manager to handle temporary directory settings for both Unix and Windows environments
