### 2026-07-19 735f0c1
Add disk space validation to GazetteerBuilder and update configurations with measured disk size requirements
