### 2026-07-19 7b5215e
Refactor memory management in GazetteerBuilder to improve CPU and memory usage detection
