### 2026-07-19 0253999
Update installation documentation to include typical install times for built-in gazetteers
