### 2026-07-19 b1dd6f6
Add installation tip for memory requirement when building gazetteers
