### 2026-07-19 802f244
Update unit tests for GazetteerBuilder by refining memory limit checks and adding preference for cgroup limits over sysconf values
