### 2026-07-19 c013494
Add unit tests for GazetteerBuilder memory and CPU detection
