### 2026-07-19 6492bea
Update progress messages in gazetteer installer changing "Reading" to "Loading" and "Extracting" to "Unpacking"
