### 2026-07-19 54d93d7
Update SQL join references in gazetteer configurations to use aliases for consistency
