+++ docs/guides/gazetteers.rst
@@ -330,37 +330,37 @@ Column references throughout a block (in ``identifier``, ``geometry``, ``names``
    # or
    geometry: "ST_Point(longitude, latitude)"  # Built from coordinate columns
 
-**Data** lists exactly what goes into the feature's ``data`` dictionary. Each entry is a column or scalar expression, written exactly as it would appear in a SQL ``SELECT``, with an optional trailing ``AS <alias>`` naming the key it is stored under. A bare or qualified column reference (``name``, ``countryInfo.Country``) may omit the alias, in which case its own column name (the last component, for a qualified reference) is the key; a plain expression has no name of its own, so it always needs one:
+**Data** lists exactly what goes into the feature's ``data`` dictionary. Each entry is a column or scalar expression, written exactly as it would appear in a SQL ``SELECT``, with an optional trailing ``AS <alias>`` naming the key it is stored under. A bare or qualified column reference (``name``, ``c.Country``) may omit the alias, in which case its own column name (the last component, for a qualified reference) is the key; a plain expression has no name of its own, so it always needs one:
 
 .. code-block:: yaml
 
    data:
-     - "population"                          # Bare column, stored as "population"
-     - "countryInfo.Country AS country_name" # Column of a joined source, renamed
-     - "upper(name) AS name_upper"           # Expression (alias required)
+     - "population"                 # Bare column, stored as "population"
+     - "c.Country AS country_name"  # Column of a joined source, renamed
+     - "upper(name) AS name_upper"  # Expression (alias required)
 
 Joins
 ~~~~~
 
-A feature block can ``join`` other sources to enrich its rows. Each join is a **raw SQL join clause** appended to the block's source: the whole joined table becomes available (there is no separate value selection—pick what you need in ``data`` using qualified references). Within a join's ``ON`` condition, bare names are the block's source columns and joined columns are qualified (``<source>.<column>``). Joins are applied in order, so a later join can reference a table joined earlier.
+A feature block can ``join`` other sources to enrich its rows. Each join is a **raw SQL join clause** appended to the block's source: the whole joined table becomes available (there is no separate value selection—pick what you need in ``data`` using qualified references). Within a join's ``ON`` condition, bare names are the block's source columns and joined columns are qualified (``<alias>.<column>``). Give each joined table a short alias to reference it conveniently. Joins are applied in order, so a later join can reference a table joined earlier.
 
 An **attribute** join equates columns (or expressions) of the two sides:
 
 .. code-block:: yaml
 
    joins:
-     - "LEFT JOIN countryInfo ON country_code = countryInfo.ISO"
-     - "LEFT JOIN admin1CodesASCII ON country_code || '.' || admin1_code = admin1CodesASCII.code"
+     - "LEFT JOIN countryInfo c ON country_code = c.ISO"
+     - "LEFT JOIN admin1CodesASCII a1 ON country_code || '.' || admin1_code = a1.code"
 
 The joined columns are then read in ``data`` by qualification:
 
 .. code-block:: yaml
 
    data:
-     - "countryInfo.Country AS country_name"
-     - "admin1CodesASCII.name AS admin1_name"
+     - "c.Country AS country_name"
+     - "a1.name AS admin1_name"
 
[…]
