### 2026-07-25 6d1441e
Add Pleiades gazetteer configuration and initial draft of step-by-step instructions for building a custom gazetteer
