+++ docs/index.rst
@@ -61,6 +61,7 @@ The Irchel Geoparser is released under the `MIT License <https://github.com/dguz
    guides/modules
    guides/training
    guides/gazetteers
+   guides/custom-gazetteers
 
 .. toctree::
    :maxdepth: 1

[…]
