### 2026-07-26 1667c80
Integrate custom gazetteer documentation into gazetteer guide, remove Pleiades from built-in gazetteers and add automatic geometry transformation
