### 2026-07-26 d30eb4e
Add unit tests for bare references and enhance CSV loading error handling
