### 2026-07-26 7a334df
Update gazetteer documentation to clarify attribute mapping and improve descriptions
