+++ docs/guides/gazetteers.rst
@@ -514,7 +514,7 @@ A gazetteer is only as good as its names, and so far each place has exactly one.
    433032    Pompei     la            Pompei         Pompei
    433032    Colonia …  la            Colonia …      Colonia …
 
-To reach them, the feature block *joins* that file. A join is written as a raw SQL join clause, appended to the block's source; the whole joined table becomes available, and you pick what you need from it afterwards. Declare ``names`` as a second source (again with all nineteen of its columns — abbreviated here), then join it:
+To reach them, the feature block *joins* that file. A join is written as a raw SQL join clause, appended to the block's source; the whole joined table becomes available, and you pick what you need from it afterwards. Declare ``names`` as a second source — all nineteen columns, in file order, exactly as for ``places`` — and then join it:
 
 .. code-block:: yaml
 
@@ -532,9 +532,24 @@ To reach them, the feature block *joins* that file. A join is written as a raw S
        attributes:
          - name: "created"
            type: text
-         # ... remaining columns in file order ...
+         - name: "description"
+           type: text
+         - name: "details"
+           type: text
+         - name: "provenance"
+           type: text
+         - name: "title"
+           type: text
+         - name: "uri"
+           type: text
+         - name: "id"
+           type: text
          - name: "place_id"
            type: integer
+         - name: "name_type"
+           type: text
+         - name: "language_tag"
+           type: text
          - name: "attested_form"
            type: text
          - name: "romanized_form_1"
@@ -543,7 +558,16 @@ To reach them, the feature block *joins* that file. A join is written as a raw S
            type: text
          - name: "romanized_form_3"
            type: text
-         # ...
+         - name: "association_certainty"
+           type: text
+         - name: "transcription_accuracy"
+           type: text
+         - name: "transcription_completeness"
+           type: text
+         - name: "year_after_which"
+           type: integer
+         - name: "year_before_which"
+           type: integer
 
    features:
      - source: places
@@ -561,6 +585,8 @@ To reach them, the feature block *joins* that file. A join is written as a raw S
          - "title"
          # ... as before ...
 
+Note that ``id`` is ``text`` in this file and ``integer`` in ``places.csv``. Each source is declared on its own terms: the type describes the column in *that* file, and here it holds a slug rather than a number. What has to match is the pair of columns the join compares, ``places.id`` and ``names.place_id``, both integers.
+
 Two conventions make join clauses short. Give every joined table a **short alias** (``n`` here) and refer to its columns through it (``n.attested_form``). Columns of the block's *own* source are written **bare** (``id``, ``title``), everywhere in the block including inside the join condition — you never write a prefix for them. Use ``LEFT JOIN`` rather than ``JOIN`` unless you deliberately want to drop places that have no match: an inner join here would silently discard the 15,301 places that have no row in ``names.csv``.
 
 The name count rises from 42,242 to 77,923, and Pompeii now carries its Latin, Greek, and Italian names:
@@ -577,7 +603,7 @@ The name count rises from 42,242 to 77,923, and Pompeii now carries its Latin, G
 Step 6: Clean Up the Names
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Names are rarely usable exactly as stored, because source data mixes names with editorial notation. A quick look through Pleiades titles shows three patterns:
+Names sometimes need work before they are usable, because source data mixes names with editorial notation. A quick look through Pleiades titles shows three patterns:
 
 .. code-block:: text
 
[…]
