### 2026-07-26 981fd2e
Use a dedicated function for delimited text, to ensure consistent newline handling across platforms
