### 2026-07-26 bff47be
Switch order of GeoNames and GeoNames Cities in gazetteer and installation docs
