### 2026-07-26 51028de
Update dependencies and Python version constraints in pyproject.toml and adjust GitHub Actions workflow to include Python 3.14
