+++ pyproject.toml
@@ -14,10 +14,10 @@ include = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.10,<3.14"
+python = ">=3.10,<3.15"
 accelerate = "^1.11.0"
 appdirs = "^1.4.4"
-datasets = "^4.3.0"
+datasets = "^5.0.0"
 duckdb = "^1.5.4"
 fastapi = "^0.121.0"
 markupsafe = "^3.0.3"
@@ -27,15 +27,15 @@ python-multipart = "^0.0.20"
 pyyaml = "^6.0.3"
 rapidfuzz = "^3.14.5"
 requests = "^2.32.5"
-rich = "^13.9.4"
+rich = "^15.0.0"
 sentence-transformers = "^5.1.2"
 shapely = "^2.1.2"
-spacy = "^3.8.7"
+spacy = "^3.8.7,!=3.8.14" # 3.8.14 has cp314 wheels on GitHub but not on PyPI (explosion/spaCy#13949)
 sqlalchemy = "^2.0.44"
-sqlmodel = "^0.0.27"
+sqlmodel = "^0.0.39"
 torch = "^2.9.0"
-transformers = "^4.57.1"
-typer = "^0.20.0"
+transformers = "^5.0.0"
+typer = "^0.27.0"
 uvicorn = "^0.38.0"
 werkzeug = "^3.1.3"
 
