### 2026-07-26 96f8726
Rewrite CONTRIBUTING.md with improved guidance on local developmen and contribution processes
