+++ README.md
@@ -50,7 +50,7 @@ Full documentation is available at **[docs.geoparser.app](https://docs.geoparser
 
 ## Contributing
 
-Contributions are welcome! Please feel free to submit a Pull Request. For major changes, please open an issue first to discuss what you would like to change.
+Questions, bug reports, and ideas are always welcome via [issues](https://github.com/dguzh/geoparser/issues). Pull requests are appreciated too — see [CONTRIBUTING.md](CONTRIBUTING.md) for local setup and development guidelines.
 
 ## Acknowledgments
 
