### 2026-07-26 83ce9ca
Add download statistics badge to README
