### 2026-07-26 d3bf69a
Update copyright years in LICENSE and documentation
