### 2026-07-26 7112ed1
Update version to 0.5.0 in pyproject.toml and documentation configuration
