+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "geoparser"
-version = "0.4.0"
+version = "0.5.0"
 description = "A customizable geoparsing library for unstructured text"
 authors = ["Diego Gomes <diego.gomes@uzh.ch>"]
 readme = "README.md"
