### 2026-08-07 83c9d45
Add document retrieval by IDs and update Geoparser initialization
