+++ tests/unit/test_project/test_project.py
@@ -5,7 +5,7 @@ Tests the Project class with mocked dependencies.
 """
 
 from unittest.mock import ANY, Mock, patch
-from uuid import UUID
+from uuid import UUID, uuid4
 
 import pytest
 
@@ -100,6 +100,144 @@ class TestProjectCreateDocuments:
         assert "Doc 1" in call_args_list
         assert "Doc 2" in call_args_list
         assert "Doc 3" in call_args_list
+
+    @patch("geoparser.project.project.ProjectRepository")
+    @patch("geoparser.project.project.DocumentRepository")
+    def test_returns_document_ids_in_input_order(
+        self, mock_doc_repo, mock_project_repo
+    ):
+        """Test that create_documents returns the new IDs in the order of the texts."""
+        # Arrange
+
+        mock_existing_project = Mock()
+        mock_existing_project.id = UUID("12345678-1234-5678-1234-567812345678")
+        mock_project_repo.get_by_name.return_value = mock_existing_project
+
+        created_ids = [uuid4(), uuid4(), uuid4()]
+        mock_doc_repo.create.side_effect = [Mock(id=id) for id in created_ids]
+
+        project = Project("TestProject")
+
+        # Act
+        document_ids = project.create_documents(["Doc 1", "Doc 2", "Doc 3"])
+
+        # Assert
+        assert document_ids == created_ids
+
+
+@pytest.mark.unit
+class TestProjectGetDocumentsByIds:
+    """Test Project get_documents method when specific IDs are requested."""
+
+    @patch("geoparser.project.project.Context")
+    @patch("geoparser.project.project.ProjectRepository")
+    @patch("geoparser.project.project.DocumentRepository")
+    def test_returns_documents_in_requested_order(
+        self, mock_doc_repo, mock_project_repo, mock_context
+    ):
+        """Test that get_documents returns documents in the order the IDs were given."""
+        # Arrange
+
+        mock_existing_project = Mock()
+        mock_existing_project.id = UUID("12345678-1234-5678-1234-567812345678")
+        mock_project_repo.get_by_name.return_value = mock_existing_project
+
+        first_id, second_id = uuid4(), uuid4()
+        mock_doc1 = Mock(id=first_id, references=[])
+        mock_doc2 = Mock(id=second_id, references=[])
+
+        # The repository makes no promises about ordering
+        mock_doc_repo.get_by_ids.return_value = [mock_doc2, mock_doc1]
+
+        mock_context_instance = Mock()
+        mock_context_instance.get_recognizer_context.return_value = None
+        mock_context_instance.get_resolver_context.return_value = None
+        mock_context.return_value = mock_context_instance
+
+        project = Project("TestProject")
+
+        # Act
+        documents = project.get_documents(ids=[first_id, second_id])
+
+        # Assert
+        assert documents == [mock_doc1, mock_doc2]
+        mock_doc_repo.get_by_ids.assert_called_once_with(
+            ANY, project.id, [first_id, second_id]
+        )
+
+    @patch("geoparser.project.project.Context")
+    @patch("geoparser.project.project.ProjectRepository")
+    @patch("geoparser.project.project.DocumentRepository")
+    def test_accepts_ids_as_strings(
+        self, mock_doc_repo, mock_project_repo, mock_context
+    ):
+        """Test that get_documents accepts IDs given as strings."""
+        # Arrange
+
+        mock_existing_project = Mock()
+        mock_existing_project.id = UUID("12345678-1234-5678-1234-567812345678")
+        mock_project_repo.get_by_name.return_value = mock_existing_project
+
+        document_id = uuid4()
+        mock_doc = Mock(id=document_id, references=[])
+        mock_doc_repo.get_by_ids.return_value = [mock_doc]
+
[…]
