### 2026-08-07 2eb0cae
Refactor Geoparser initialization to require explicit recognizer and resolver parameters and removing deprecated parameters and associated warnings
