+++ geoparser/geoparser/geoparser.py
@@ -1,14 +1,10 @@
 import uuid
-import warnings
 from typing import List, Optional, Union
 
 from geoparser.db.models import Document
 from geoparser.modules.recognizers import Recognizer
 from geoparser.modules.resolvers import Resolver
 from geoparser.project import Project
-
-# Sentinel value to distinguish "not provided" from "explicitly None"
-_UNSET = object()
 
 
 class Geoparser:
@@ -17,127 +13,35 @@ class Geoparser:
 
     Provides a simple parse method for processing texts with configured recognizer and resolver.
     The Geoparser creates a new project for each parse operation, making it stateless by default.
+
+    The recognizer and resolver have to be provided explicitly, so that it is
+    always clear which modules a pipeline is built from::
+
+        from geoparser import Geoparser
+        from geoparser.modules import SentenceTransformerResolver, SpacyRecognizer
+
+        geoparser = Geoparser(
+            recognizer=SpacyRecognizer(),
+            resolver=SentenceTransformerResolver(gazetteer_name="geonames"),
+        )
     """
 
     def __init__(
         self,
-        recognizer: Optional[Recognizer] = _UNSET,
-        resolver: Optional[Resolver] = _UNSET,
-        spacy_model: Optional[str] = None,
-        transformer_model: Optional[str] = None,
+        recognizer: Optional[Recognizer],
+        resolver: Optional[Resolver],
     ):
         """
         Initialize a Geoparser instance.
 
         Args:
-            recognizer: The recognizer module to use for identifying references.
-                       If not provided, a default SpacyRecognizer will be created.
-                       Can be explicitly set to None to skip recognition step.
-            resolver: The resolver module to use for resolving references to referents.
-                     If not provided, a default SentenceTransformerResolver will be created.
-                     Can be explicitly set to None to skip resolution step.
-            spacy_model: (Deprecated) Name of spaCy model to use.
-                        Use SpacyRecognizer(model_name='...') instead.
-            transformer_model: (Deprecated) Name of transformer model to use.
-                              Use SentenceTransformerResolver(model_name='...') instead.
+            recognizer: The recognizer module to use for identifying references,
+                        or None to skip the recognition step.
+            resolver: The resolver module to use for resolving references to referents,
+                      or None to skip the resolution step.
         """
-        # Handle legacy parameters with deprecation warning
-        self._warn_deprecated_parameters(spacy_model, transformer_model)
-
-        # Set up recognizer
-        if recognizer is _UNSET:
-            # No recognizer provided, create default
-            from geoparser.modules import SpacyRecognizer
-
-            if spacy_model is not None:
-                self.recognizer = SpacyRecognizer(model_name=spacy_model)
-            else:
-                self.recognizer = SpacyRecognizer()
-        else:
-            # Recognizer explicitly provided (could be None to skip)
-            self.recognizer = recognizer
-
-        # Set up resolver
-        if resolver is _UNSET:
-            # No resolver provided, create default
-            from geoparser.modules import SentenceTransformerResolver
-
-            if transformer_model is not None:
-                self.resolver = SentenceTransformerResolver(
-                    model_name=transformer_model
-                )
-            else:
-                self.resolver = SentenceTransformerResolver()
-        else:
-            # Resolver explicitly provided (could be None to skip)
-            self.resolver = resolver
-
-    @staticmethod
-    def _warn_deprecated_parameters(
-        spacy_model: Optional[str], transformer_model: Optional[str]
-    ) -> None:
-        """
-        Show deprecation warning for legacy parameters.
-
-        Args:
-            spacy_model: Legacy spacy_model parameter value
[…]
