### 2026-08-07 fd82574
Enhance error messaging for missing gazetteers and optimize initialization order in SentenceTransformerResolver
