### 2026-08-07 5fadab5
Update documentation to reflect new Geoparser initialization with explicit recognizer and resolver and new id based document retrieval from projects
