### 2026-08-07 35bb030
Refactor Geoparser methods to ensure consistent handling of single and multiple document inputs, updating documentation and tests accordingly
