### 2026-08-09 a7ba29e
Restructure CI around pull requests and tag-driven releases
Tests and lint ran twice for every commit, once for the push event and
once for the pull_request event on the same SHA, and on every push to
every branch. They now run on pull requests into main plus the merge
result on the trunk, which keeps the README badges and the shared
Actions cache alive while cutting the bulk of the redundancy.

Releases move from the staging branch to tags, so stg_release.sh and its
scraping of `pip index versions` are gone. Release candidates publish to
real PyPI as PEP 440 pre-releases, which verifies depen
[…]
