+++ .github/workflows/release.yml
@@ -29,7 +29,7 @@ jobs:
     outputs:
       prerelease: ${{ steps.version.outputs.prerelease }}
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v7
         with:
           # Needed to prove the tagged commit is reachable from main.
           fetch-depth: 0
