### 2026-08-09 ac72d0f
Derive the docs version from pyproject and tighten CONTRIBUTING docs/conf.py hardcoded release = "0.5.0", so the rendered version would have drifted at every release. It now reads pyproject.toml, which works on Read the Docs where poetry install --no-root means geoparser itself is never importable. The release section of CONTRIBUTING had grown explanatory: it justified design decisions and explained what git tags and GitHub Releases are before getting to anything project-specific. Rewritten as instructions, with the per-tag behaviour as a table and the failure modes as a troubleshooting list.
