+++ CONTRIBUTING.md
@@ -98,57 +98,53 @@ A few practical tips that make reviews easier:
 - Add or update tests when behavior changes
 - Update docs when user-facing behavior changes
 
-CI runs on pull requests only, never on branch pushes, so a commit is tested once rather than twice. The full matrix is three operating systems across Python 3.10–3.14. Pushing again to an open pull request cancels the previous run.
+CI runs on pull requests into `main` and on `main` itself, never on feature-branch pushes. The matrix is three operating systems across Python 3.10–3.14. Pushing again to an open pull request cancels the previous run.
 
-If you add a dependency, commit the updated `poetry.lock` alongside `pyproject.toml`, and prefer permissively licensed packages — geoparser is MIT-licensed and its dependencies should not impose stricter terms on users.
+If you add a dependency, commit the updated `poetry.lock` alongside `pyproject.toml`. Prefer permissively licensed packages; geoparser is MIT-licensed.
 
 ## Releasing
 
-Releases are driven entirely by tags, and the tag name is the version — there is no release branch. Tags carry no `v` prefix.
+For maintainers. Releases are driven by tags: the tag name is the version, tags carry no `v` prefix, and there is no release branch.
 
-**Step 1 — set the version.** The version lives in `pyproject.toml`, so bumping it is an ordinary code change. Set the *final* version even if you plan to cut release candidates first.
-
-**Step 2 — tag the merged commit.** A tag is a separate ref, not a branch push, so the branch ruleset does not apply and this works even though `main` rejects direct pushes:
+Bump the version with `poetry version <version>` in the last pull request of the cycle, setting the final version even when release candidates come first. Once it is merged, tag from `main`:
 
 ```bash
 git switch main && git pull
-git tag 0.6.0rc1 && git push origin 0.6.0rc1
+VERSION=$(poetry version --short)
+git tag "${VERSION}rc1" && git push origin "${VERSION}rc1"
 ```
 
-**Step 3 — verify the published artifact.** PEP 440 hides pre-releases from resolvers, so `--pre` is required to see it at all and ordinary users cannot install it by accident:
+Check the candidate in a clean environment. `--pre` is required, since pre-releases are hidden from resolvers:
 
 ```bash
-python -m venv /tmp/rc && /tmp/rc/bin/pip install --pre geoparser==0.6.0rc1
-/tmp/rc/bin/python -c "import geoparser; print(geoparser.__version__)"
+python -m venv /tmp/rc
+/tmp/rc/bin/pip install --pre "geoparser==${VERSION}rc1"
+/tmp/rc/bin/python -c "from importlib.metadata import version; print(version('geoparser'))"
 ```
 
-**Step 4 — cut the real release.** No second version bump is needed; a release-candidate tag stamps its own version onto the build, while a plain tag uses the version already in `pyproject.toml`:
+Then tag the release itself. No second version bump is needed:
 
 ```bash
-git tag 0.6.0 && git push origin 0.6.0
+git tag "$VERSION" && git push origin "$VERSION"
 ```
 
-If the candidate needs fixes, merge them through a pull request as usual and tag `0.6.0rc2`.
+If the candidate needs fixes, merge them through a pull request and tag `${VERSION}rc2`.
 
-### What the tag decides
+### What each tag produces
 
-A tag containing `rc` publishes to PyPI as a pre-release, which `pip install geoparser` will not pick up, and leaves nothing on the Releases page. A plain `MAJOR.MINOR.PATCH` tag publishes a full release, signs the artifacts with Sigstore, and creates the GitHub Release.
+| Tag | PyPI | GitHub Release | Read the Docs |
+| --- | --- | --- | --- |
+| `1.4.0rc1` | pre-release, needs `--pre` | none | inactive until activated |
[…]
