+++ .github/workflows/coverage-preview.yml
@@ -25,7 +25,7 @@ jobs:
         with:
           python-version: '3.11'
       - run: pip install smokeshow
-      - uses: dawidd6/action-download-artifact@v6
+      - uses: dawidd6/action-download-artifact@v21
         with:
           workflow: test.yml
           commit: ${{ github.event.workflow_run.head_sha }}
