### 2026-08-15 9386964
Initial draft of extensive documentation overhaul
