+++ docs/guides/results.rst
@@ -1,0 +1,299 @@
+.. _results:
+
+Working with Results
+====================
+
+Parsing text is rarely the last step. This guide covers what comes out of a pipeline, how to get at it, and how to turn it into the formats you will actually analyse in — tables, spreadsheets, and spatial files.
+
+The Shape of the Output
+-----------------------
+
+Whatever you parse, you get **documents**. Each document holds the **toponyms** found in it, and each toponym may hold the **location** it resolved to.
+
+.. code-block:: python
+
+   from geoparser import Geoparser
+   from geoparser.modules import SentenceTransformerResolver, SpacyRecognizer
+
+   geoparser = Geoparser(
+       recognizer=SpacyRecognizer(),
+       resolver=SentenceTransformerResolver(gazetteer_name="geonames"),
+   )
+
+   document = geoparser.parse(
+       "The conference was held in Zurich, with satellite events in Geneva and Basel."
+   )
+
+   for toponym in document.toponyms:
+       print(toponym.text, toponym.start, toponym.end, toponym.location)
+
+Three levels, each answering a different question:
+
+.. list-table::
+   :header-rows: 1
+   :widths: 22 78
+
+   * - Object
+     - What you get from it
+   * - ``Document``
+     - ``text``, ``id``, and ``toponyms``
+   * - ``Reference`` (a toponym)
+     - ``text`` as written, ``start`` and ``end`` character offsets, and ``location``
+   * - ``Feature`` (a location)
+     - ``identifier``, ``data`` attributes, ``geometry``, ``crs``, ``names``, ``source``
+
+The offsets are the useful part people overlook: they let you go back to the text. ``document.text[toponym.start:toponym.end]`` is the toponym, and widening the slice gives you the sentence it sat in — which is what you want for keyword-in-context tables or for checking a suspicious result by eye.
+
+Always Check for ``None``
+-------------------------
+
+``toponym.location`` is ``None`` when the name was recognized but not resolved. This is normal and expected: the place may be absent from the gazetteer, or no candidate may have been convincing enough. The resolver prefers admitting uncertainty to guessing.
+
+.. code-block:: python
+
+   for toponym in document.toponyms:
+       if toponym.location is None:
+           print(f"{toponym.text}: unresolved")
+           continue
+       data = toponym.location.data
+       print(f"{toponym.text}: {data.get('name')}, {data.get('country_name')}")
+
+Read attributes with ``.get()`` rather than ``data["name"]``. Which keys exist depends on the gazetteer, and even on the source within one gazetteer, so subscripting works until it suddenly does not. Geometry needs the same care — a place known by name but never located has none:
+
+.. code-block:: python
+
+   geometry = toponym.location.geometry
+   if geometry is not None:
+       print(geometry.x, geometry.y)
+
+Flattening to Rows
+------------------
+
+Almost every downstream use starts by turning the nested structure into one row per toponym. This is the single most useful piece of code to have on hand:
+
+.. code-block:: python
+
+   def to_rows(documents):
+       """Flatten parsed documents into one dictionary per toponym."""
+       rows = []
+       for document in documents:
+           for toponym in document.toponyms:
+               location = toponym.location
+               data = location.data if location else {}
+               rows.append(
+                   {
+                       "document_id": str(document.id),
+                       "toponym": toponym.text,
+                       "start": toponym.start,
+                       "end": toponym.end,
+                       "resolved": location is not None,
+                       "gazetteer_id": location.identifier if location else None,
+                       "name": data.get("name"),
+                       "country": data.get("country_name"),
+                       "feature_type": data.get("feature_name"),
+                       "latitude": data.get("latitude"),
+                       "longitude": data.get("longitude"),
+                   }
[…]
