+++ docs/guides/custom-gazetteers.rst
@@ -1,0 +1,872 @@
+.. _custom-gazetteers:
+
+Building Custom Gazetteers
+==========================
+
+The pre-configured gazetteers cover the modern world and one country in detail, but no fixed set of gazetteers can cover every research question. If you work on a region, a period, or a domain that is not represented — a national placename register, an excavation catalogue, a historical map index, your own field data — you can turn that data into a gazetteer of your own, and the library then treats it exactly like a pre-configured one.
+
+You do this by writing a YAML configuration file that describes your source files and how their rows map onto places. There is no plugin to write and no code to run: you declare where the files are, what columns they have, and which of those columns are the identifier, the names, the geometry, and the attributes of a place. The build pipeline does the rest.
+
+The rest of this guide walks through that process end to end on a real dataset, then documents every configuration key in full.
+
+What You Are Building
+---------------------
+
+Before writing any configuration, it helps to know exactly what the build produces, because that is what your configuration has to describe.
+
+The Canonical Feature Model
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Every gazetteer, however heterogeneous its sources, is projected into a single model. A gazetteer is a set of **features**, and each feature has exactly five things:
+
+.. list-table::
+   :header-rows: 1
+   :widths: 15 85
+
+   * - Field
+     - Meaning
+   * - ``identifier``
+     - A string that identifies the place within this gazetteer and never changes. It is what gets stored in annotations, so it must be stable across rebuilds and unique across the whole gazetteer.
+   * - ``names``
+     - Every string the place should be findable by: its main name, historical spellings, transliterations, names in other languages, abbreviations. Names are not ranked or labelled — they are a set of search keys.
+   * - ``geometry``
+     - One geometry (point, line, polygon, or a multi-part combination), in the gazetteer's coordinate reference system. It may be absent: a place that is known by name but not located is still a perfectly valid feature.
+   * - ``data``
+     - A free-form dictionary of attributes: type, hierarchy, population, dates, links, descriptions — whatever your source offers and your work needs. There is no fixed schema, and different sources within one gazetteer may store entirely different keys.
+   * - ``source``
+     - Which of the configuration's sources the feature came from. Set automatically; useful for telling apart features of different kinds in one gazetteer.
+
+The finished gazetteer is a single self-contained SQLite file (an *artifact*) holding those features plus the full-text and phonetic indexes used for searching. Nothing else is installed, and the artifact is never modified after the build.
+
+Writing a configuration is therefore an exercise in answering five questions about your data: what is one place, what identifies it, what is it called, where is it, and what else do I want to know about it.
+
+How the Build Runs
+~~~~~~~~~~~~~~~~~~
+
+Understanding the three stages the build reports makes its error messages much easier to place:
+
+1. **Preparing sources.** Each source file is downloaded or located on disk, extracted if it is a ZIP archive, and loaded into a table in a temporary analytical database (DuckDB). Errors here are about files and columns: a missing file, a column count that does not match, a value that will not convert to its declared type.
+2. **Compiling features.** For each ``features`` block, the declared identifier, names, geometry, and data are compiled into SQL over the block's source and its joins, and run. Errors here are about your expressions: an unknown column name, an invalid join clause, an identifier that collides with another block's.
[…]

[…]
