### 2026-08-16 432fd88
Compress working with results guide, update awkward phrases and remove next steps sections
