+++ docs/guides/results.rst
@@ -3,7 +3,7 @@
 Working with Results
 ====================
 
-Parsing text is rarely the last step. This guide covers what comes out of a pipeline, how to get at it, and how to turn it into the formats you will actually analyse in — tables, spreadsheets, and spatial files.
+This guide describes what a pipeline gives you back and how to read it: the three kinds of object a parse produces, the attributes each one carries, and the two cases — an unresolved name and a missing attribute — that any code reading results has to allow for.
 
 The Shape of the Output
 -----------------------
@@ -27,6 +27,12 @@ Whatever you parse, you get **documents**. Each document holds the **toponyms** 
    for toponym in document.toponyms:
        print(toponym.text, toponym.start, toponym.end, toponym.location)
 
+.. code-block:: text
+
+   Zurich 27 33 Feature(geonames:2657896)
+   Geneva 60 66 Feature(geonames:2660646)
+   Basel 71 76 Feature(geonames:2661604)
+
 Three levels, each answering a different question:
 
 .. list-table::
@@ -42,12 +48,39 @@ Three levels, each answering a different question:
    * - ``Feature`` (a location)
      - ``identifier``, ``data`` attributes, ``geometry``, ``crs``, ``names``, ``source``
 
-The offsets are the useful part people overlook: they let you go back to the text. ``document.text[toponym.start:toponym.end]`` is the toponym, and widening the slice gives you the sentence it sat in — which is what you want for keyword-in-context tables or for checking a suspicious result by eye.
+Full signatures for all three are in the :doc:`../api/models` reference.
 
-Always Check for ``None``
--------------------------
+Toponyms
+--------
 
-``toponym.location`` is ``None`` when the name was recognized but not resolved. This is normal and expected: the place may be absent from the gazetteer, or no candidate may have been convincing enough. The resolver prefers admitting uncertainty to guessing.
+A toponym is a place name as it was found in the text. ``text`` is the name as written, which need not be the gazetteer's name for the place: the toponym above reads ``Zurich``, while GeoNames names that city ``Zürich``.
+
+``start`` and ``end`` are character offsets into ``document.text``, with ``end`` exclusive, so ``document.text[toponym.start:toponym.end]`` gives the toponym back and a wider slice gives the text around it.
+
+Locations
+---------
+
+A location is a ``Feature``: one entry in the gazetteer the resolver was using.
+
+``data`` holds that gazetteer's attributes for the place, as a dictionary. Which keys exist depends on the gazetteer, and can differ between sources within one gazetteer, so read them with ``.get()``:
+
+.. code-block:: python
+
+   data = toponym.location.data
+   print(data.get("name"), data.get("country_name"), data.get("feature_name"))
+
+``identifier`` is the gazetteer's stable id for the place. It is what to store when you need a reference that survives — in annotations, in exported data, in anything you come back to — and what to group or count by. Names are not unique: GeoNames has 122 places called Paris, so counting by ``data["name"]`` silently merges places that happen to share a name, while counting by ``identifier`` does not.
+
+``geometry`` is a Shapely object — usually a point, but lines, polygons, and multi-part geometries occur — expressed in the coordinate reference system named by ``crs``, which is ``EPSG:4326`` for both pre-configured gazetteers. Because it is a Shapely object, it can be measured, transformed, or handed to any library that reads Shapely geometries.
+
+``names`` lists every string the place is searchable by, and ``source`` names the gazetteer source the feature came from, which is how features of different kinds within one gazetteer can be told apart.
+
+What May Be Missing
+-------------------
+
+Two things are absent often enough that reading results means allowing for them.
+
+**A toponym may have no location.** ``toponym.location`` is ``None`` when a name was recognized but not resolved, either because the place is not in the gazetteer or because no candidate passed the resolver's similarity threshold.
 
 .. code-block:: python
 
[…]
