### 2026-08-16 9bdb29d
Update documentation with new Sphinx theme shibuya and include new logo and related assets
