### 2026-08-16 62b1f27
Adjust logo size and update badge links alignment
