### 2026-08-16 f6737c9
Fix demo documentation to reflect the total of 74 populated places mentioned in demo text
