### 2026-08-16 ee7cdb2
Bump version to 0.6.0 in pyproject.toml
