### 2026-08-22 ba1dfaf
Cite two recent studies that used the Irchel Geoparser.
