+++ paper/paper.bib
@@ -162,6 +162,28 @@
   doi = {10.1007/s42001-025-00448-x}
 }
 
+@article{hanny2026,
+  author = {Hanny, David and {Ghosh Dastidar}, Kanishka and Wieland, Marc and Granitzer, Michael and Resch, Bernd},
+  title = {{Towards multimodal geospatial reasoning: a foundation model approach for disaster detection from social media, news, and weather data}},
+  journal = {Natural Hazards},
+  volume = {122},
+  number = {11},
+  pages = {459},
+  year = {2026},
+  doi = {10.1007/s11069-026-08191-w}
+}
+
+@article{jalilian2026,
+  author = {Jalilian, Ehsaneddin and Hanny, David and Resch, Bernd},
+  title = {{Beyond unimodal metrics: a multimodal evaluation framework for geo-referenced content on social media}},
+  journal = {GeoInformatica},
+  volume = {30},
+  number = {2},
+  pages = {25},
+  year = {2026},
+  doi = {10.1007/s10707-026-00584-0}
+}
+
 @article{anderson2026,
   author = {Anderson, Cory J. and Giordano, Alberto},
   title = {{Natural Language Processing and the Spatial Humanities: Comparing Geoparsing of Holocaust Testimonies With GPT-4.1, Deep Learning, and Rules-Based NLP}},
