+++ paper/paper.md
@@ -50,7 +50,7 @@ The gazetteer layer itself is also customisable. While a GeoNames-based configur
 
 # Research Impact Statement
 
-The Irchel Geoparser has been used in both research and teaching contexts. For example, it supported a study on geographic bias in global climate disaster reporting, in which thousands of news articles were analysed and mapped to reveal patterns in the geographic focus of media attention [@kong2025]. @hanny2025 also used the Irchel Geoparser to analyse the locations of disasters described in social media content before carrying out a spatial-temporal analysis of emotions. @anderson2026 integrated the Irchel Geoparser in a forthcoming publication to explore the locations described in holocaust testimonies. It was also used in a number of Master's theses, for instance to investigate how newspaper discourses reflected the return of the wolf to Switzerland in the last two decades [@besse2025] and to explore global discourses of disaster resilience [@haus2025].
+The Irchel Geoparser has been used in both research and teaching contexts. For example, it supported a study on geographic bias in global climate disaster reporting, in which thousands of news articles were analysed and mapped to reveal patterns in the geographic focus of media attention [@kong2025]. @hanny2025 also used the Irchel Geoparser to analyse the locations of disasters described in social media content before carrying out a spatial-temporal analysis of emotions. It was later used to extract place mentions from Bluesky posts in a multimodal disaster-detection study [@hanny2026] and in an evaluation of multimodal topic models of disaster-related social media [@jalilian2026]. @anderson2026 integrated the Irchel Geoparser in a forthcoming publication to explore the locations described in holocaust testimonies. It was also used in a number of Master's theses, for instance to investigate how newspaper discourses reflected the return of the wolf to Switzerland in the last two decades [@besse2025] and to explore global discourses of disaster resilience [@haus2025].
 
 Furthermore, the straightforward installation, comprehensive documentation, and a complete demo available as a Docker image [@gomes2025demo] also make the Irchel Geoparser an accessible tool for teaching and for illustrating key concepts in geographic information retrieval and we have used it in multiple workshops and courses.
 
