### 2026-08-22 55c4e76
Update the JOSS paper manuscript
