+++ paper/paper.md
@@ -25,32 +25,32 @@ bibliography: paper.bib
 
 # Summary
 
-The Irchel Geoparser is an open-source Python library for recognising and resolving place names (toponyms) in unstructured text. Based on a modular architecture, it provides an end-to-end geoparsing pipeline in which the components responsible for toponym recognition, toponym resolution, and the underlying gazetteer data can be freely adapted or exchanged. The library includes robust, configurable built-in modules and a GeoNames-based gazetteer configuration, offering strong performance for English text out of the box. Users can also choose to fine-tune the built-in modules or integrate their own custom recognisers and resolvers. Similarly, it is possible to include domain-specific gazetteers, allowing geoparsing workflows to be tailored to diverse research and practical contexts.
+The Irchel Geoparser is an open-source Python library for recognising and resolving place names (toponyms) in unstructured text. Based on a modular architecture, it provides an end-to-end geoparsing pipeline in which the components responsible for toponym recognition, toponym resolution, and the underlying gazetteer data can be freely adapted or exchanged. The library includes robust, configurable built-in modules and a GeoNames-based gazetteer configuration that together provide a ready-to-use pipeline for English text. Users assemble this pipeline from the modules they choose. They can also fine-tune the built-in modules or integrate their own custom recognisers and resolvers. Similarly, it is possible to include domain-specific gazetteers, allowing geoparsing workflows to be tailored to diverse research and practical contexts.
 
 # Statement of Need
 
 Geoparsing is an essential step in linking text to geographical locations [@purves2018], enabling diverse applications including mapping news coverage, exploring historical documents, or analysing discourse across regions [@teitler2008; @grover2010; @chesnokova2019]. Typically, geoparsing is a two-stage process involving firstly toponym recognition and subsequently toponym resolution. Toponyms, or place names, are first identified in text, a special case of the more general problem of named entity recognition. Toponym recognition must differentiate between instances of words being used in other contexts - for example to describe generic objects (as for instance in the case of bath (a place to wash oneself and an English city)) or to name people and firms (for example (George) Washington or Zurich (Insurance)). Toponym resolution resolves candidate place names to a unique identifier, dealing with the problem of multiple instances of the same place name (for example London, UK vs. London, Ontario, Canada).
 
-The Irchel Geoparser was designed to provide a modular, out of the box functional Python library which could easily be incorporated into text processing chains using default settings. As such, it is well suited to teaching applications introducing for example graduate students to the geoparsing process. At the same time, because of its modular design, the entire processing chain is configurable to particular use cases - for example by swapping gazetteers to use those more appropriate to historical documents or retraining language models underlying toponym resolution modules on a specific corpus. Thus, the Irchel Geoparser is also well suited for use in research. Finally, we also aimed to provide a well documented Python library with an open licence allowing not only experimentation but also reproducibility of results.
+The Irchel Geoparser was designed to provide a modular Python library which could easily be incorporated into text processing chains by assembling a complete pipeline from its built-in modules. As such, it is well suited to teaching applications introducing for example graduate students to the geoparsing process. At the same time, because of its modular design, the entire processing chain is configurable to particular use cases - for example by swapping gazetteers to use those more appropriate to historical documents or retraining language models underlying toponym resolution modules on a specific corpus. Thus, the Irchel Geoparser is also well suited for use in research. Finally, we also aimed to provide a well documented Python library with an open licence allowing not only experimentation but also reproducibility of results.
 
 
 # State of the Field
 
[…]
