### 2026-08-31 #10365 @jGaboardi
@SamHames @cforgaci – These timelines work for me.

As a note, I will be out for parental leave from ≈September 18, so I may be slow to respond after that for a bit.

cc @dguzh @samhforbes
