+++ .github/workflows/coverage-preview.yml
@@ -21,7 +21,7 @@ jobs:
       # Listing the triggering run and downloading its artifacts.
       actions: read
     steps:
-      - uses: actions/setup-python@v5
+      - uses: actions/setup-python@v7
         with:
           python-version: '3.11'
       - run: pip install smokeshow
