+++ .github/workflows/release.yml
@@ -89,7 +89,7 @@ jobs:
       id-token: write
     steps:
       - name: Download the dists
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v8
         with:
           name: python-package-distributions
           path: dist/
@@ -110,7 +110,7 @@ jobs:
       id-token: write
     steps:
       - name: Download the dists
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v8
         with:
           name: python-package-distributions
           path: dist/
