### 2026-09-07 55c5c95
Add CITATION.cff
