+++ CITATION.cff
@@ -1,0 +1,13 @@
+cff-version: 1.2.0
+message: If you use this software, please cite it as below.
+title: 'The Irchel Geoparser: A Modular Python Library for Toponym Recognition and Resolution'
+type: software
+authors:
+- given-names: Diego
+  family-names: Gomes
+  orcid: https://orcid.org/0009-0003-8449-2603
+- given-names: Ross S.
+  family-names: Purves
+  orcid: https://orcid.org/0000-0002-9878-9243
+repository-code: https://github.com/dguzh/geoparser
+license: MIT
