### 2026-09-10 edfba6b
build: migrate from Poetry to uv
Translate pyproject.toml to PEP 621 with dependency groups, swap the build
backend to hatchling, and lock with uv. Dependency constraints are carried
over verbatim from the Poetry carets, so the resolved set is unchanged.

Update the CI, Read the Docs and release workflows to uv, and fix
docs/conf.py, which read the version out of [tool.poetry].

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
