+++ scripts/crap.py
@@ -1,0 +1,192 @@
+"""
+CRAP score gate.
+
+CRAP (Change Risk Anti-Patterns) combines how branchy a function is with how
+well it is tested::
+
+    CRAP = complexity^2 * (1 - coverage)^3 + complexity
+
+A fully covered function scores its own cyclomatic complexity, so the gate is
+simultaneously a complexity ceiling for tested code and a much harsher one for
+untested code. Run it after pytest has written a coverage data file.
+
+    uv run python scripts/crap.py --max-crap 30
+
+Exits non-zero when any measured function scores above the threshold.
+"""
+
+from __future__ import annotations
+
+import argparse
+import sys
+from dataclasses import dataclass
+from pathlib import Path
+
+from coverage import Coverage
+from coverage.exceptions import NoSource
+from radon.complexity import cc_visit
+
+# Mirrors [tool.coverage.run] omit: the annotator is a server-rendered UI that
+# the test suite does not measure, so it has no coverage to compute CRAP from.
+DEFAULT_OMIT = ("geoparser/annotator",)
+
+
+@dataclass(frozen=True)
+class Score:
+    """A single function's complexity, coverage and resulting CRAP score."""
+
+    path: str
+    name: str
+    lineno: int
+    complexity: int
+    coverage: float
+
+    @property
+    def crap(self) -> float:
+        """The CRAP score for this function."""
+        uncovered = 1.0 - self.coverage
+        return self.complexity**2 * uncovered**3 + self.complexity
+
+    def __str__(self) -> str:
+        return (
+            f"{self.crap:7.2f}  cc={self.complexity:<3} "
+            f"cov={self.coverage * 100:5.1f}%  {self.path}:{self.lineno} {self.name}"
+        )
+
+
+def _analyse(coverage: Coverage, path: Path) -> tuple[set[int], set[int]] | None:
+    """
+    Return the (statements, missing) line numbers for a file.
+
+    Args:
+        coverage: A loaded coverage session
+        path: The file to analyse
+
+    Returns:
+        A tuple of statement and missing line numbers, or None when the file
+        has no coverage data at all.
+    """
+    try:
+        _, statements, _, missing, _ = coverage.analysis2(str(path))
+    except NoSource:
+        return None
+    return set(statements), set(missing)
+
+
+def score_file(coverage: Coverage, path: Path, root: Path) -> list[Score]:
+    """
+    Score every function and method in one source file.
+
+    Args:
+        coverage: A loaded coverage session
+        path: The file to score
+        root: Repository root, used to render relative paths
+
+    Returns:
+        One Score per function, ordered as they appear in the file.
+    """
+    analysed = _analyse(coverage, path)
+    if analysed is None:
+        return []
+    statements, missing = analysed
+
+    relative = path.relative_to(root).as_posix()
+    scores = []
+    for block in cc_visit(path.read_text(encoding="utf-8")):
+        # cc_visit yields classes as well as functions; a class's own score is
+        # the sum of its methods, which would double-count them.
+        if block.letter == "C":
+            continue
+        span = range(block.lineno, block.endline + 1)
+        owned = statements.intersection(span)
+        # A function with no measurable statements (an overload stub, say)
+        # cannot be under-tested, so it counts as fully covered.
+        covered = 1.0
+        if owned:
+            covered = 1.0 - len(owned & missing) / len(owned)
+        scores.append(
+            Score(
+                path=relative,
+                name=block.fullname,
+                lineno=block.lineno,
+                complexity=block.complexity,
+                coverage=covered,
+            )
+        )
+    return scores
+
+
[…]
