+++ CONTRIBUTING.md
@@ -4,41 +4,45 @@ Thanks for contributing to Irchel Geoparser. For product usage, see the [documen
 
 ## Setup
 
-This project uses [Poetry](https://python-poetry.org/docs/) for dependency management. Install Poetry, then from the repository root:
+This project uses [uv](https://docs.astral.sh/uv/) for dependency management. Install uv, then from the repository root:
 
 ```bash
-poetry install
+uv sync
 ```
 
-That creates a virtual environment and installs runtime and development dependencies (including spaCy models used in tests).
+That creates `.venv/`, installs runtime and development dependencies (including the spaCy models used in tests) at the versions pinned in `uv.lock`, and installs geoparser itself in editable mode. uv downloads a suitable interpreter automatically, so no separate Python install is needed.
 
-Run tools through Poetry:
+Run tools through uv:
 
 ```bash
-poetry run <command>
+uv run <command>
 ```
 
 To activate the environment in your current shell instead:
 
 ```bash
-eval $(poetry env activate)
+source .venv/bin/activate
 ```
 
 Supported Python versions are `>=3.10,<3.15`. Keep changes compatible across that range.
 
 ## Code style
 
-Formatting and import hygiene are checked in CI on every pull request; the job fails if the code is not clean. Before opening a PR, run:
+[Ruff](https://docs.astral.sh/ruff/) handles formatting, import sorting, unused-code removal and linting in one tool; it replaces the former black + isort + autoflake trio. Its version is pinned in `pyproject.toml` and locked in `uv.lock`, so CI and your machine format identically.
 
 ```bash
-poetry run black .
-poetry run isort .
-poetry run autoflake --remove-all-unused-imports --in-place --recursive --exclude=__init__.py geoparser tests
+uv run ruff check --fix .
+uv run ruff format .
 ```
 
-- **black** formats the code (`required-version` is pinned in `pyproject.toml`)
-- **isort** sorts imports (`profile = "black"`)
-- **autoflake** removes unused imports (`__init__.py` is excluded)
+CI runs the checking form of both and fails if anything is unformatted or unclean:
+
+```bash
+uv run ruff check .
+uv run ruff format --check .
+```
+
+The enabled rule sets are declared under `[tool.ruff.lint]` in `pyproject.toml`. Prefer fixing a finding over silencing it; a `# noqa` needs a specific code and a reason.
 
 ## Tests
 
@@ -48,30 +52,64 @@ Tests live under `tests/` and are organized as:
 - `tests/integration/` — exercises real components together (models, DB, gazetteers)
 - `tests/e2e/` — full pipeline tests
 
-Markers `unit`, `integration`, and `e2e` are defined in `tests/pytest.ini`.
+Markers `unit`, `integration`, and `e2e` are declared under `[tool.pytest.ini_options]` in `pyproject.toml`, which is the single source of pytest configuration.
 
 Run the full suite:
 
 ```bash
-poetry run pytest
+uv run pytest
 ```
 
-Coverage is collected for `geoparser` (HTML report in `htmlcov/`; open `htmlcov/index.html`). `geoparser/annotator/` is omitted from coverage. Pull requests expect near-complete coverage of the measured package, so new functionality should ship with tests.
+Coverage is collected for `geoparser` (HTML report in `htmlcov/`; open `htmlcov/index.html`). `geoparser/annotator/` is omitted from coverage. CI enforces a hard floor on the combined coverage of the whole matrix:
+
+```bash
+uv run pytest --cov-fail-under=COVERAGE_GATE
+```
 
 Useful subsets:
 
 ```bash
-poetry run pytest tests/unit
-poetry run pytest tests/integration/test_geoparser_integration.py
+uv run pytest tests/unit
+uv run pytest tests/integration/test_geoparser_integration.py
 ```
+
+## Quality checks
+
[…]
