+++ .readthedocs.yaml
@@ -12,12 +12,14 @@ build:
     python: "3.12"
   jobs:
     post_create_environment:
-      # Install Poetry via pip
-      - pip install poetry
+      # uv installs into the environment Read the Docs already created, so the
+      # docs build uses exactly the versions pinned in uv.lock.
+      - pip install uv
     post_install:
-      # Install your package dependencies using Poetry
-      # Manually set VIRTUAL_ENV to work with Poetry in RTD
-      - VIRTUAL_ENV=$READTHEDOCS_VIRTUALENV_PATH poetry install --no-root
+      # Runtime dependencies are needed because autodoc imports geoparser from the
+      # source tree; the project package itself is not installed, matching the
+      # previous `poetry install --no-root`.
+      - UV_PROJECT_ENVIRONMENT=$READTHEDOCS_VIRTUALENV_PATH uv sync --locked --no-install-project --no-default-groups --group docs
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:

[…]
