### 2026-09-10 5614712
style: replace black/isort/autoflake with ruff
Ruff now owns formatting, import sorting, unused-code removal and linting.
Beyond the mechanical rewrites (PEP 585/604 annotations, f-string
conversions), this fixes findings the previous trio could not see:

- 72 unused locals, including two dead recognizer_id extractions in
  ReferenceRepository
- two caching tests that captured call counts and never asserted on them;
  they now fail if the cache stops working
- two blind 'pytest.raises(Exception)' narrowed to IntegrityError and
  ValidationError
- shared mutable argument defaults across the ann
[…]
