### 2026-09-10 41cd83b
fix: correct types found by ty in the core packages
ty surfaced two latent crashes, both now covered by tests:

- SentenceTransformerResolver._extract_context called sentences.index(None)
  when a reference span fell outside every sentence, failing with
  'None is not in list'; it now reports the offset it could not place.
- The same method assumed get_max_seq_length() returns an int.

Also: list() around Sequence returns declared as list, a shared
_load_context_record helper replacing four unchecked Optional accesses in
Context, corrected results annotations in the resolver (predict seeds the
[…]
