### 2026-09-10 fd8094e
fix: correct types found by ty in the gazetteer build pipeline
- Stage now exposes its Progress/TaskID through a _display accessor, so
  using a stage outside its 'with' block reports that instead of an
  AttributeError on None, and the 25 diagnostics in progress.py collapse
  to zero.
- A scalar_int() helper replaces four 'fetchone()[0]' calls that would
  have raised TypeError had the aggregate returned no row.
- track()'s run parameter is typed '() -> object'; the return value was
  always discarded and every caller passes an expression lambda.
- Optional config fields (source path, delimit
[…]
