### 2026-09-10 cb32306
fix: correct service, project and resolver type annotations
Recognizer and Resolver ids are the module's config hash, a str; the
service internals annotated them as uuid.UUID throughout. Trainability is
optional, so 'fit' is now looked up explicitly instead of via hasattr and
an untyped call. ManualResolver.__init__ and Project.create_referents
declared referents as tuples when both already accept None for an
ungeocoded reference, as their own predict signature and docstring say.

Every package except the annotator is now clean under ty.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
