### 2026-09-10 a2adf2b
fix: clear the remaining ty errors in the annotator
The annotator had no direct tests, and ty found real defects there:

- Nineteen routes typed their injected session/document as 'dict', hiding
  every attribute access on them.
- update_annotation, delete_annotation and annotate_many dereferenced an
  optional toponym lookup, so unmatched offsets produced a 500 rather than
  the ToponymNotFoundException the app already defines.
- DocumentRepository iterated recognizer output that may be None, and
  passed a possibly-absent upload filename straight to secure_filename.
- BaseRepository was not 
[…]
