+++ geoparser/annotator/db/crud/settings.py
@@ -12,21 +12,25 @@ from geoparser.annotator.db.models.settings import (
 from geoparser.annotator.exceptions import SessionSettingsNotFoundException
 
 
-class SessionSettingsRepository(BaseRepository):
+class SessionSettingsRepository(BaseRepository[AnnotatorSessionSettings]):
     model = AnnotatorSessionSettings
     exception_factory: t.Callable[[str, uuid.UUID], Exception] = lambda x, y: (
         SessionSettingsNotFoundException(f"{x} with ID {y} not found.")
     )
 
     @classmethod
-    def create(
+    # BaseRepository declares the widest input type (SQLModel); each repository
+    # deliberately accepts its own Create/Update model. Callers always go
+    # through the concrete repository, so the precise signature is worth more
+    # here than strict substitutability.
+    def create(  # ty: ignore[invalid-method-override]
         cls,
         db: DBSession,
         item: AnnotatorSessionSettingsCreate,
         exclude: list[str] | None = None,
         additional: dict[str, t.Any] | None = None,
     ) -> AnnotatorSessionSettings:
-        assert "session_id" in additional, (
+        assert additional and "session_id" in additional, (
             "settings cannot be created without link to session"
         )
         return super().create(db, item, exclude=exclude, additional=additional)
@@ -40,7 +44,7 @@ class SessionSettingsRepository(BaseRepository):
         return super().read_all(db, **filters)
 
     @classmethod
-    def update(
+    def update(  # ty: ignore[invalid-method-override]
         cls, db: DBSession, item: AnnotatorSessionSettingsUpdate
     ) -> AnnotatorSessionSettings:
         return super().update(db, item)

[…]
