### 2026-09-10 2b5afc8
refactor: split the three highest-complexity functions
Radon's only C-grade blocks, all now decomposed with behaviour unchanged:

- SourceConfig.validate_source (16) -> one method per concern: location,
  attribute names, and the tabular/spatial split.
- SentenceTransformerResolver.predict (16 -> 8) -> _all_resolved,
  _search_once and _search_tier; the search methods become a class
  constant now that two methods share them.
- _extract_context (13) -> _token_limit, _document_tokens, _sentences,
  _locate_sentence and _expand_window, each independently readable.

Worst CRAP score across the pa
[…]
