+++ .github/workflows/test.yml
@@ -27,10 +27,10 @@ env:
   # matrix. Single cells are not gated: platform-specific branches are
   # unreachable off their own operating system, so only the combination is a
   # fair measure. Raise this when coverage rises; never lower it to go green.
-  COVERAGE_THRESHOLD: "0"
+  COVERAGE_THRESHOLD: "99"
   # The worst tolerated CRAP score (complexity^2 * (1-coverage)^3 + complexity).
   # For fully covered code this is just a cyclomatic-complexity ceiling.
-  CRAP_THRESHOLD: "0"
+  CRAP_THRESHOLD: "10"
 
 jobs:
   pytest:

[…]
