### 2026-09-10 c1ba357
test: let mutation runs see the documented example config
mutmut runs the suite from a copied tree, and the schema tests validate
docs/examples/pleiades.yaml, so without also_copy every mutant failed for
want of a file and stats collection aborted.

Also caches the mutants/ directory in CI, so a pull request re-tests only
the mutants its changes affect rather than sweeping the whole package.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
