### 2026-09-10 aef5fae
ci: set the mutation baseline from a measured run
3764 mutants over the library, judged by the unit suite: 2353 killed, 956
survived, 287 with no covering test at all -- a 71.1% mutation score, at
8.56 mutations/second once the one-off test-to-code mapping is built.

The gap against ~100% line coverage is the finding, not a problem with the
run: a surviving mutant is a line the fast suite executes without checking.
Survivors cluster in the gazetteer build pipeline, which the integration
suite exercises for real but which mutation testing does not see.

MAX_SURVIVING_MUTANTS is a ratchet at the
[…]
