### 2026-09-10 9f9cb5b
perf: stop the CLI importing the world, and speed up the gates
python -m geoparser --help imported the annotator's FastAPI app, and with
it spaCy, thinc and torch, purely to register the 'annotator' command;
geoparser.gazetteer imported the whole build pipeline to re-export
GazetteerBuilder. Both are now resolved on use:

  python -m geoparser --help   8.6s -> 0.29s
  unit suite                   72.6s -> 28.7s
  full suite (gated)           ~75s, coverage overhead now under 1%

torch, spaCy, thinc, transformers and duckdb are no longer imported at all
for --help; a test pins that the package 
[…]
