### 2026-09-10 603cd66
perf: keep mutation testing off the merge path, and pin the CLI fast path
Mutation testing is now nightly and on demand rather than a pull request
gate. The merge path is ruff, ty, the test matrix, coverage and CRAP, which
finish in minutes; the mutants/ cache keeps warm runs incremental.

Judging mutants with the integration suite as well was measured and
rejected, with the numbers recorded in CONTRIBUTING: it removes every
'no tests' mutant and cuts survival from 29% to about 11%, but each mutant
it reaches rebuilds a real gazetteer -- roughly 23s apiece, some thirteen
hours for the package.
[…]
