### 2026-09-11 5efad2a
test: drive mutation survivors from 931 to 193
Scoping first: the gazetteer build pipeline is excluded from mutation. Its
mutants can only be judged by the integration suite, which rebuilds a real
gazetteer per mutant -- ~23s each, some thirteen hours for the package. That
exclusion also removed every segfault (164 -> 0), since those all came from
mutating native-extension code mutmut runs in-process, and took the sweep
from 9.7 to 34 mutants/second.

Then the survivors themselves, function by function. Most were real gaps the
tests now close:

- no test asserted that a clean database passes t
[…]
