### 2026-09-11 04f06af
test: drive mutation survivors from 193 to 146
More real gaps closed by tests rather than suppressed:

- the annotation import never had its data flow pinned, so swapping or
  dropping one of the three parallel lists went unnoticed
- run_recognizer, run_resolver and get_documents had no test for the tag they
  file results under, so mutating the "latest" default was invisible
- _best_referent had nothing asserting it picks the highest similarity, nor
  that a candidate exactly at the threshold is accepted
- _token_limit had no test that it reserves two tokens for [CLS] and [SEP]
- both service
[…]
