+++ tests/unit/test_services/test_prediction_recording.py
@@ -1,0 +1,158 @@
+"""
+Tests for how the services record what a pluggable module predicted.
+
+Recognizers and resolvers are supplied by the caller, so the services treat a
+short prediction list leniently -- the trailing items are simply left
+unprocessed -- and skip individual predictions that come back as None. Both
+behaviours are easy to break into either a crash or a silent early exit.
+"""
+
+from types import SimpleNamespace
+from unittest.mock import Mock, patch
+
+import pytest
+
+from geoparser.services.recognition import RecognitionService
+from geoparser.services.resolution import ResolutionService
+
+
+@pytest.mark.unit
+class TestRecordReferencePredictions:
+    """Writing a recognizer's spans."""
+
+    @staticmethod
+    def _record(documents, predictions):
+        """Run the recorder, returning the reference and recognition writes."""
+        service = RecognitionService(Mock())
+        references, recognitions = [], []
+        with (
+            patch.object(
+                service,
+                "_create_reference_record",
+                side_effect=lambda s, d, start, end, r: references.append(
+                    (d, start, end)
+                ),
+            ),
+            patch.object(
+                service,
+                "_create_recognition_record",
+                side_effect=lambda s, d, r: recognitions.append(d),
+            ),
+        ):
+            service._record_reference_predictions(Mock(), documents, predictions, "rec")
+        return references, recognitions
+
+    def test_tolerates_fewer_predictions_than_documents(self):
+        """
+        A recognizer that returns too few results is not an error.
+
+        The documents it did cover are recorded and the rest are left for a
+        later run, rather than the whole batch failing.
+        """
+        # Arrange
+        documents = [SimpleNamespace(id="d1"), SimpleNamespace(id="d2")]
+
+        # Act
+        references, recognitions = self._record(documents, [[(0, 5)]])
+
+        # Assert
+        assert references == [("d1", 0, 5)]
+        assert recognitions == ["d1"]
+
+    def test_skips_a_none_prediction_and_keeps_going(self):
+        """
+        None means "could not process this document", not "stop".
+
+        Breaking out here would silently drop every later document whenever
+        one came back unprocessed.
+        """
+        # Arrange
+        documents = [
+            SimpleNamespace(id="d1"),
+            SimpleNamespace(id="d2"),
+            SimpleNamespace(id="d3"),
+        ]
+
+        # Act
+        references, recognitions = self._record(documents, [[(0, 1)], None, [(2, 3)]])
+
+        # Assert
+        assert references == [("d1", 0, 1), ("d3", 2, 3)]
+        assert recognitions == ["d1", "d3"]
+
+    def test_marks_a_document_processed_only_once_per_recognizer(self):
+        """Each covered document gets exactly one recognition record."""
+        # Arrange
+        documents = [SimpleNamespace(id="d1")]
+
+        # Act
+        _, recognitions = self._record(documents, [[(0, 1), (2, 3)]])
+
+        # Assert
+        assert recognitions == ["d1"]
+
+
+@pytest.mark.unit
+class TestRecordReferentPredictions:
+    """Writing a resolver's referents."""
+
+    @staticmethod
+    def _record(references, predictions):
+        """Run the recorder, returning the referent and resolution writes."""
+        service = ResolutionService(Mock())
+        referents, resolutions = [], []
+        with (
+            patch.object(
+                service,
+                "_create_referent_record",
+                side_effect=lambda s, ref, g, i, r: referents.append((ref, g, i)),
+            ),
+            patch.object(
+                service,
+                "_create_resolution_record",
+                side_effect=lambda s, ref, r: resolutions.append(ref),
+            ),
+        ):
[…]
