### 2026-09-11 d5ff02a
test: drive mutation survivors from 146 to 120
Tests for the resolver's contrastive training pairs (which candidate is
labelled positive, and that each pair carries its own reference's context)
and for the reference-text span fallbacks, where a partial update has to keep
the end of the span it was not given.

The redundant 'and hasattr(document, "text")' guard in ReferenceRepository
is removed rather than tested: a Document row always has text, and the
foreign key means the lookup either finds one or the update fails, so the
branch was unreachable.

Co-Authored-By: Claude Opus 5 <noreply@anthr
[…]
