### 2026-09-11 a246b4b
test: drive mutation survivors from 120 to 91
Tests for Gazetteer.search's dispatch (which artifact query each method
reaches, and that the defaults are exact/one tier), the artifact's read-only
per-thread connection, ManualResolver's annotation lookup, the description
text a candidate is embedded as, and the two embedding caches.

Four pragmas, each with its reason: three zips whose two sides come from the
same call, so strict= cannot change anything, and the soundex first-letter
default whose equivalence was already demonstrated.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
