### 2026-09-11 2d993fd
feat: add GLiNER2 and Jina modules
