+++ geoparser/modules/resolvers/sentencetransformer.py
@@ -22,6 +22,11 @@ if t.TYPE_CHECKING:
 logging.set_verbosity_error()
 
 
+# Throughput and display only: neither changes the embeddings that come back.
+_ENCODE_BATCH_SIZE = 32  # pragma: no mutate
+_SHOW_ENCODE_PROGRESS = True  # pragma: no mutate
+
+
 class SentenceTransformerResolver(Resolver):
     """
     A resolver that uses SentenceTransformer to map reference contexts to gazetteer candidates.
@@ -73,6 +78,7 @@ class SentenceTransformerResolver(Resolver):
         min_similarity: float = 0.6,
         max_tiers: int = 3,
         attribute_map: dict | None = None,
+        **extra_config,
     ):
         """
         Initialize the SentenceTransformerResolver.
@@ -86,6 +92,9 @@ class SentenceTransformerResolver(Resolver):
                           If None, will look up gazetteer_name in GAZETTEER_ATTRIBUTE_MAP.
                           If provided, will be used directly.
                           Should have keys: "name", "type", "level1", "level2", "level3"
+            **extra_config: Additional configuration a subclass wants recorded
+                          in the module id, so two resolvers that differ only
+                          in a subclass parameter do not share one.
         """
         # Initialize parent with the parameters
         super().__init__(
@@ -94,6 +103,7 @@ class SentenceTransformerResolver(Resolver):
             min_similarity=min_similarity,
             max_tiers=max_tiers,
             attribute_map=attribute_map,
+            **extra_config,
         )
 
         # Store instance attributes directly from parameters
@@ -112,14 +122,8 @@ class SentenceTransformerResolver(Resolver):
         self.gazetteer = Gazetteer(gazetteer_name)
 
         # Initialize transformer and tokenizer
-        self.transformer = SentenceTransformer(model_name)
-        # Annotated explicitly: AutoTokenizer's return union includes backend
-        # types (and None) that do not carry .tokenize, which is all this class
-        # uses. Narrowing here types the four call sites correctly; from_pretrained
-        # does not actually return None for a resolvable model name.
-        self.tokenizer: PreTrainedTokenizerBase = AutoTokenizer.from_pretrained(  # ty: ignore[invalid-assignment]
-            model_name
-        )
+        self.transformer = self._load_transformer(model_name)
+        self.tokenizer = self._load_tokenizer(model_name)
 
         # Initialize spaCy model for sentence splitting
         self.nlp = self._load_spacy_model("xx_sent_ud_sm")
@@ -133,6 +137,35 @@ class SentenceTransformerResolver(Resolver):
         self.candidate_embeddings: dict[
             int, torch.Tensor
         ] = {}  # feature_id -> embedding
+
+    def _load_transformer(self, model_name: str, **kwargs) -> SentenceTransformer:
+        """
+        Load the embedding model.
+
+        A hook rather than a direct call so a subclass whose checkpoint ships
+        its own modelling code can pass the flags that need.
+
+        Args:
+            model_name: HuggingFace checkpoint to load
+            **kwargs: Extra arguments for the SentenceTransformer constructor
+
+        Returns:
+            The loaded model
+        """
+        return SentenceTransformer(model_name, **kwargs)
+
+    def _load_tokenizer(self, model_name: str, **kwargs) -> PreTrainedTokenizerBase:
+        """
+        Load the tokenizer used to size reference contexts.
+
+        Args:
+            model_name: HuggingFace checkpoint to load the tokenizer of
+            **kwargs: Extra arguments for ``AutoTokenizer.from_pretrained``
+
+        Returns:
+            The loaded tokenizer
+        """
[…]
