+++ geoparser/gazetteer/gazetteer.py
@@ -36,13 +36,14 @@ class Gazetteer:
         """
         path = artifact_path(gazetteer_name)
         if not path.exists():
-            # pragma: no mutate block - wording only; a test pins the type
+            # pragma: no mutate start - wording only; a test pins the type
             # and that the message names the missing gazetteer.
             raise ValueError(
                 f"Gazetteer '{gazetteer_name}' is not installed. Install it by running "
                 f"'python -m geoparser install {gazetteer_name}', or run "
                 "'python -m geoparser list' to see which gazetteers are installed."
             )
+            # pragma: no mutate end
         self.gazetteer_name = gazetteer_name
         self._artifact = GazetteerArtifact(path)
 

[…]
