### 2026-09-11 3ae4fe5
docs: add quality uplift implementation plan
