+++ docs/superpowers/plans/2026-09-11-quality-uplift-mkdocs-migration.md
@@ -1,0 +1,371 @@
+# Repository Quality Uplift and MkDocs Migration Implementation Plan
+
+> **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development (recommended) or superpowers:executing-plans to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
+
+**Goal:** Migrate the public documentation from Sphinx to strict MkDocs Material and make the requested deterministic quality gauntlet executable locally and blocking in CI without changing library behavior.
+
+**Architecture:** Keep `geoparser/` runtime behavior unchanged. Add small, side-effect-isolated quality tools under `scripts/`, test them through stable command contracts, use AST-only architecture checks for dependency cycles and layer violations, and keep generated coverage, mutation, build, and Docker artifacts outside the repository. Convert the existing public documentation into Markdown with a single MkDocs navigation and use `mkdocstrings` only for API surfaces that must follow Python signatures.
+
+**Tech Stack:** uv and `uv.lock`, Ruff, ty, pytest/pytest-cov, Hypothesis, pytest-bdd, radon/CRAP, mutmut, MkDocs Material, mkdocstrings, Python AST, Docker, GitHub Actions, GitHub Pages.
+
+---
+
+### Task 1: Establish the quality-contract tests
+
+**Files:**
+- Create: `tests/unit/test_quality/__init__.py`
+- Create: `tests/unit/test_quality/test_project_contract.py`
+- Modify: `pyproject.toml`
+
+- [ ] **Step 1: Write failing contract tests**
+
+Add tests that inspect `pyproject.toml` with `tomllib`. The first contract slice must require the `hypothesis` and `pytest-bdd` development dependencies and the `property`, `acceptance`, and `architecture` pytest markers. Later tasks extend this same contract test with the files and workflow assertions they introduce.
+
+- [ ] **Step 2: Run the focused tests and capture RED**
+
+Run:
+
+```bash
+uv run --no-sync pytest tests/unit/test_quality/test_project_contract.py -q
+```
+
+Expected result: failure because the new quality entrypoints, MkDocs configuration, Docker packaging, citation metadata, and dependencies do not yet exist.
+
+- [ ] **Step 3: Implement the minimum project contract**
+
+Add the dependency and marker declarations only. Keep dependency versions bounded and let `uv lock` update the lockfile.
+
+- [ ] **Step 4: Run the focused tests and verify GREEN**
+
+Run the same focused pytest command and require all contract assertions to pass.
+
+- [ ] **Step 5: Commit the contract foundation**
+
+```bash
+git add pyproject.toml uv.lock tests/unit/test_quality/test_project_contract.py tests/unit/test_quality/__init__.py
+git commit -m "test: define repository quality contracts"
+```
+
+### Task 2: Add the architecture and dependency-boundary checker
+
+**Files:**
+- Create: `scripts/check_architecture.py`
+- Create: `tests/unit/test_quality/test_architecture.py`
+- Modify: `pyproject.toml`
+
+- [ ] **Step 1: Write failing tests for the stable checker interface**
+
+Test `build_import_graph`, `find_cycles`, and `find_boundary_violations` against temporary packages. Cover an acyclic graph, a three-node cycle, a runtime forbidden import, and an import under `if TYPE_CHECKING:` that is allowed for annotations. Also test that the real `geoparser` package is checked from the repository root.
+
+- [ ] **Step 2: Run the focused tests and verify RED**
+
+```bash
+uv run --no-sync pytest tests/unit/test_quality/test_architecture.py -q
+```
+
[…]
