### 2026-09-11 a65f941
test: define repository quality contracts
