+++ pyproject.toml
@@ -71,6 +71,8 @@ lint = [
 test = [
     "pytest>=8.4.2,<10",
     "pytest-cov>=7.0.0,<8",
+    "hypothesis>=6.140.0,<7",
+    "pytest-bdd>=8.1.0,<9",
     "requests-mock>=1.12.1,<2",
     "mutmut>=3.7.0,<4",
     "en-core-web-sm",
@@ -129,6 +131,9 @@ markers = [
     "unit: Fast, isolated unit tests with mocked dependencies",
     "integration: Integration tests with real models and database",
     "e2e: End-to-end pipeline tests",
+    "property: Property-based tests over generated inputs",
+    "acceptance: Executable user-facing acceptance scenarios",
+    "architecture: Automated import and dependency-boundary checks",
 ]
 addopts = [
     "--strict-markers",
