### 2026-09-11 207f40e
feat: enforce import architecture boundaries
