+++ scripts/check_architecture.py
@@ -1,0 +1,212 @@
+"""Check import cycles and dependency boundaries in the package."""
+
+from __future__ import annotations
+
+import argparse
+import ast
+from collections.abc import Iterable, Mapping
+from pathlib import Path
+
+
+FORBIDDEN_IMPORTS: dict[str, set[str]] = {
+    "geoparser.db": {
+        "geoparser.context",
+        "geoparser.modules",
+        "geoparser.project",
+        "geoparser.services",
+    },
+    "geoparser.modules": {
+        "geoparser.context",
+        "geoparser.db",
+        "geoparser.project",
+        "geoparser.services",
+    },
+    "geoparser.context": {
+        "geoparser.modules",
+        "geoparser.project",
+        "geoparser.services",
+    },
+    "geoparser.services": {
+        "geoparser.context",
+        "geoparser.modules",
+        "geoparser.project",
+    },
+}
+
+
+def _module_name(path: Path, package_root: Path, package_name: str) -> str:
+    relative = path.relative_to(package_root).with_suffix("")
+    parts = relative.parts
+    if parts[-1] == "__init__":
+        parts = parts[:-1]
+    return ".".join((package_name, *parts))
+
+
+def _is_type_checking_test(test: ast.expr) -> bool:
+    """Return whether an ``if`` condition is a TYPE_CHECKING guard."""
+    if isinstance(test, ast.Name):
+        return test.id == "TYPE_CHECKING"
+    if isinstance(test, ast.Attribute):
+        return test.attr == "TYPE_CHECKING"
+    return False
+
+
+class _ImportVisitor(ast.NodeVisitor):
+    """Collect imports while excluding annotation-only imports."""
+
+    def __init__(self) -> None:
+        self.runtime: list[tuple[str | None, int, str | None, tuple[str, ...]]] = []
+
+    def visit_If(self, node: ast.If) -> None:
+        if _is_type_checking_test(node.test):
+            return
+        self.generic_visit(node)
+
+    def visit_Import(self, node: ast.Import) -> None:
+        for alias in node.names:
+            self.runtime.append((alias.name, node.lineno, None, ()))
+
+    def visit_ImportFrom(self, node: ast.ImportFrom) -> None:
+        self.runtime.append(
+            (
+                node.module,
+                node.lineno,
+                "." * node.level,
+                tuple(alias.name for alias in node.names),
+            )
+        )
+
+
+def _resolve_from_import(
+    current: str,
+    module: str | None,
+    dots: str | None,
+    names: tuple[str, ...],
+    packages: set[str],
+) -> Iterable[str]:
+    """Yield import targets that correspond to real internal modules."""
+    if dots == "":
+        base = module or ""
+    else:
+        level = len(dots)
+        current_parts = current.split(".")
+        current_is_package = current in packages
+        parent_parts = current_parts if current_is_package else current_parts[:-1]
+        base_parts = parent_parts[: len(parent_parts) - level + 1]
+        base = ".".join(base_parts)
+        if module:
+            base = f"{base}.{module}" if base else module
+
+    for name in names:
+        exact = f"{base}.{name}" if base else name
+        target = exact if exact in packages else base
+        if target in packages:
+            yield target
+
+
+def build_import_graph(package_root: Path, package_name: str) -> dict[str, set[str]]:
+    """Build a deterministic runtime import graph for one Python package."""
+    package_root = package_root.resolve()
+    files = sorted(package_root.rglob("*.py"))
+    modules = {
+        _module_name(path, package_root, package_name): path for path in files
+    }
+    graph = {module: set() for module in modules}
+    packages = set(modules)
+
+    for module, path in modules.items():
+        visitor = _ImportVisitor()
+        visitor.visit(ast.parse(path.read_text(encoding="utf-8"), filename=str(path)))
[…]
