### 2026-09-11 ffaba99
test: add property and acceptance coverage
