+++ tests/acceptance/__init__.py
@@ -1,0 +1,1 @@
+"""Executable user-facing acceptance scenarios."""
