+++ tests/property/__init__.py
@@ -1,0 +1,1 @@
+"""Property-based tests for pure matching helpers."""
