### 2026-09-11 8d07b9b
chore: enforce quality gates and migrate docs
