+++ scripts/quality_gauntlet.py
@@ -1,0 +1,269 @@
+"""Run the repository's deterministic quality stages in order."""
+
+from __future__ import annotations
+
+import argparse
+import os
+import subprocess
+import tempfile
+from collections.abc import Iterator
+from contextlib import contextmanager
+from dataclasses import dataclass
+from pathlib import Path
+
+Command = tuple[str, ...]
+
+
+@dataclass(frozen=True)
+class Stage:
+    """A named group of commands that share one working directory."""
+
+    name: str
+    commands: tuple[Command, ...]
+    cwd: Path
+
+
+def _uv(*arguments: str) -> Command:
+    return ("uv", "run", "--no-sync", *arguments)
+
+
+def build_stages(
+    root: Path,
+    artifact_dir: Path,
+    *,
+    skip_mutation: bool = False,
+    skip_docker: bool = False,
+    docker_tag: str = "geoparser:quality-check",
+) -> list[Stage]:
+    """Build the ordered quality stages for a repository checkout."""
+    root = root.resolve()
+    artifact_dir = artifact_dir.resolve()
+    coverage_report = artifact_dir / "coverage-html"
+    coverage_data = artifact_dir / ".coverage"
+
+    stages = [
+        Stage("baseline", (_uv("pytest", "--cov-fail-under=100"),), root),
+        Stage(
+            "ruff",
+            (
+                _uv("ruff", "check", "."),
+                _uv("ruff", "format", "--check", "."),
+            ),
+            root,
+        ),
+        Stage("ty", (_uv("ty", "check", "geoparser", "scripts", "tests"),), root),
+        Stage(
+            "dependencies",
+            (
+                ("uv", "lock", "--check"),
+                _uv(
+                    "deptry",
+                    "geoparser",
+                    "demo",
+                    "--per-rule-ignores",
+                    "DEP002=accelerate|python-multipart|peft|protobuf|sentencepiece,"
+                    "DEP004=plotly",
+                ),
+            ),
+            root,
+        ),
+        Stage(
+            "tests",
+            (
+                _uv(
+                    "pytest",
+                    "--cov-fail-under=100",
+                    f"--cov-report=html:{coverage_report}",
+                ),
+            ),
+            root,
+        ),
+        Stage(
+            "property",
+            (_uv("pytest", "tests/property", "-m", "property", "--no-cov", "-q"),),
+            root,
+        ),
+        Stage(
+            "acceptance",
+            (
+                _uv(
+                    "pytest",
+                    "tests/acceptance",
+                    "-m",
+                    "acceptance",
+                    "--no-cov",
+                    "-q",
+                ),
+            ),
+            root,
+        ),
+        Stage(
+            "architecture",
+            (_uv("python", "scripts/check_architecture.py", "--package", "geoparser"),),
+            root,
+        ),
+        Stage(
+            "crap",
+            (
+                _uv(
+                    "python",
+                    "scripts/crap.py",
+                    "--max-crap",
+                    "5.99",
+                    "--data-file",
+                    str(coverage_data),
+                ),
+            ),
+            root,
+        ),
+    ]
+
+    if not skip_mutation:
+        stages.append(
+            Stage(
+                "mutation",
+                (
+                    _uv("mutmut", "run"),
+                    _uv("mutmut", "export-cicd-stats"),
+                    _uv(
+                        "python",
+                        "scripts/mutation_gate.py",
+                        "--max-survivors",
+                        "0",
+                        "--stats",
+                        "mutants/mutmut-cicd-stats.json",
+                    ),
+                ),
+                root,
+            )
+        )
+
[…]
