+++ geoparser/db/__init__.py
@@ -1,0 +1,1 @@
+"""Database persistence primitives for projects and parsed documents."""

[…]
