### 2026-09-11 0a361a8
fix(ci): pin setup-uv to a resolvable release
