+++ .github/workflows/release.yml
@@ -44,7 +44,7 @@ jobs:
             exit 1
           fi
 
-      - uses: astral-sh/setup-uv@v10
+      - uses: astral-sh/setup-uv@v10.1.0
         with:
           enable-cache: true
 
