### 2026-09-11 c36411f
fix: make CPU affinity detection portable
