### 2026-09-11 cf49603
fix: make quality tests portable
