### 2026-09-12 adc0fc2
fix: make quality gates portable and deterministic
