+++ pyproject.toml
@@ -74,6 +74,7 @@ test = [
     "pytest-cov>=7.0.0,<8",
     "hypothesis>=6.140.0,<7",
     "pytest-bdd>=8.1.0,<9",
+    "tomli>=2.0.1,<3; python_version < \"3.11\"",
     "requests-mock>=1.12.1,<2",
     "mutmut>=3.7.0,<4",
     "en-core-web-sm",
