### 2026-09-12 cf5ff19
fix: support legacy quality tooling on Python 3.10
