### 2026-09-12 66e9cc9
fix: report actionable mutation failures
