### 2026-09-12 d7ca8d9
test: bound training artifact retention
