### 2026-09-12 4e8b336
fix: order training cleanup before temp teardown
