### 2026-09-12 2cb17d5
test: pin artifact read-only URI semantics
