### 2026-09-12 fffdab0
style: format artifact wiring test
