### 2026-09-12 46c51a5
refactor: extract context sizing into a pure module
Choosing how much text around a reference fits an encoder's token budget is
arithmetic over sentence costs, but it lived on SentenceTransformerResolver
and reached into the transformer, the tokenizer, the spaCy pipeline, the model
name and two caches. Exercising it meant constructing three mocked models, so
the twenty tests it had all went through that fixture.

Move the decision into geoparser.modules.resolvers.context, behind a Sentence
value object carrying the text, its span and what it costs to encode. The
module imports nothing from the
[…]
