+++ tests/property/test_context_invariants.py
@@ -1,0 +1,186 @@
+"""
+Property tests for the context window.
+
+The window is chosen by arithmetic over sentence costs, and the arithmetic has
+to hold for any document and any budget, not only the handful of shapes the
+unit tests spell out. These generate documents and budgets and assert the
+properties the resolver depends on: the reference is always in the context,
+the context is a contiguous run of whole sentences, it respects the budget
+whenever that is possible at all, and it is as large as the budget allows.
+"""
+
+import pytest
+from hypothesis import given, settings
+from hypothesis import strategies as st
+
+from geoparser.modules.resolvers.context import (
+    Sentence,
+    expand_window,
+    locate_sentence,
+    select_context,
+)
+
+PROPERTY_SETTINGS = settings(deadline=None, derandomize=True, database=None)
+
+COSTS = st.lists(st.integers(min_value=0, max_value=40), min_size=1, max_size=12)
+BUDGETS = st.integers(min_value=0, max_value=120)
+
+
+def _document(costs: list[int]) -> list[Sentence]:
+    """A document whose sentences carry the given token costs."""
+    sentences = []
+    offset = 0
+    for index, cost in enumerate(costs):
+        text = f"s{index}" * max(cost, 1)
+        sentences.append(
+            Sentence(text=text, start=offset, end=offset + len(text), cost=cost)
+        )
+        offset += len(text) + 1
+    return sentences
+
+
+@st.composite
+def documents_with_target(draw):
+    """A document, and the index of one of its sentences."""
+    costs = draw(COSTS)
+    sentences = _document(costs)
+    return sentences, draw(st.integers(min_value=0, max_value=len(sentences) - 1))
+
+
+@pytest.mark.property
+@PROPERTY_SETTINGS
+@given(documents_with_target(), BUDGETS)
+def test_the_window_always_contains_the_target_sentence(document, budget):
+    """
+    Whatever the budget, the reference's own sentence is in the context.
+
+    Losing it would hand the encoder text that does not contain the place
+    name it is being asked to disambiguate.
+    """
+    sentences, target = document
+
+    window = expand_window(sentences, target, budget)
+
+    assert sentences[target] in window
+
+
+@pytest.mark.property
+@PROPERTY_SETTINGS
+@given(documents_with_target(), BUDGETS)
+def test_the_window_is_a_contiguous_run_in_document_order(document, budget):
+    """The context is a slice of the document, not a selection from it."""
+    sentences, target = document
+
+    window = expand_window(sentences, target, budget)
+
+    first = sentences.index(window[0])
+    assert window == sentences[first : first + len(window)]
+
+
+@pytest.mark.property
+@PROPERTY_SETTINGS
+@given(documents_with_target(), BUDGETS)
+def test_the_window_fits_the_budget_unless_the_target_alone_cannot(document, budget):
+    """
+    The only context allowed to exceed the budget is a single oversized
+    sentence, which the encoder will truncate rather than lose entirely.
+    """
+    sentences, target = document
+
+    window = expand_window(sentences, target, budget)
+
+    total = sum(sentence.cost for sentence in window)
+    assert total <= budget or window == [sentences[target]]
+
+
+@pytest.mark.property
+@PROPERTY_SETTINGS
+@given(documents_with_target(), BUDGETS)
+def test_neither_neighbour_of_the_window_would_have_fitted(document, budget):
+    """
+    The window is maximal: it stopped because nothing more fitted.
+
+    Without this, a window could satisfy every other property by simply being
+    smaller than it needed to be, wasting the budget the encoder was given.
+    """
+    sentences, target = document
+
[…]
