### 2026-09-12 7d9f71c
test: fail the architecture check when a pure module gains a dependency
The value of the context module is that it depends on nothing: that is what
lets it be tested directly instead of through whatever heavy collaborators its
callers happen to construct. Nothing stopped the next change from importing a
convenient helper and quietly undoing it.

Add PURE_MODULES to the architecture checker and find_impure_modules to
enforce it. A listed module may import the standard library and nothing else;
anything from the package or from a third-party distribution is reported.
Relative imports count as in
[…]
