### 2026-09-12 256d8eb
fix: fail the mutation gate when a run checked nothing
mutmut writes a stats file and exits zero even when its own baseline test run
fails. Every mutant is then left "not checked" and the export reads
killed: 0, survived: 0 -- which a gate that looks only at the survivor count
reports as a clean sheet at a 100% score. A green build that verified nothing
is the most dangerous way for this check to be wrong.

This was hit twice while extracting the context module: once when an
architecture test that reads the real package source failed inside the
rewritten mutants/ copy, and once when a filtered
[…]
