+++ tests/unit/test_quality/test_mutation_gate.py
@@ -61,3 +61,76 @@ def test_mutation_gate_reports_unavailable_mutant_diagnostics(
         "mutmut results failed with exit code 127: mutmut missing"
         in capsys.readouterr().err
     )
+
+
+def _unaccounted_stats_path(tmp_path, **counts: int) -> str:
+    """A stats file whose totals need not add up, for the accounting checks."""
+    path = tmp_path / "mutmut-cicd-stats.json"
+    path.write_text(json.dumps(counts), encoding="utf-8")
+    return str(path)
+
+
+def test_mutation_gate_fails_when_the_run_checked_nothing(
+    tmp_path, capsys: Any
+) -> None:
+    """
+    A run that generated mutants but checked none of them is not a pass.
+
+    mutmut exits zero and writes a stats file even when its own baseline test
+    run failed, leaving every mutant "not checked". Reading only the survivor
+    count then reports a clean sheet for a gate that verified nothing at all,
+    which is the most dangerous way for this check to be wrong.
+    """
+    result = mutation_gate.main(
+        [
+            "--max-survivors",
+            "0",
+            "--stats",
+            _unaccounted_stats_path(tmp_path, killed=0, survived=0, total=2028),
+        ]
+    )
+
+    assert result == 1
+    assert "2028" in capsys.readouterr().err
+
+
+def test_mutation_gate_fails_when_some_mutants_were_never_checked(
+    tmp_path, capsys: Any
+) -> None:
+    """A partially completed run is reported rather than silently accepted."""
+    result = mutation_gate.main(
+        [
+            "--max-survivors",
+            "0",
+            "--stats",
+            _unaccounted_stats_path(
+                tmp_path, killed=900, survived=0, no_tests=100, total=2028
+            ),
+        ]
+    )
+
+    assert result == 1
+    assert "1028" in capsys.readouterr().err
+
+
+def test_mutation_gate_accepts_a_run_that_accounts_for_every_mutant(
+    tmp_path, capsys: Any
+) -> None:
+    """Every outcome category counts towards the accounting, not just kills."""
+    result = mutation_gate.main(
+        [
+            "--max-survivors",
+            "0",
+            "--stats",
+            _unaccounted_stats_path(
+                tmp_path,
+                killed=1812,
+                survived=0,
+                timeout=3,
+                no_tests=212,
+                total=2027,
+            ),
+        ]
+    )
+
+    assert result == 0
