### 2026-09-14 12734dc
feat: add pure evaluation metrics for pilot evidence
Scoring a pipeline against gold annotations is arithmetic over spans and
identifiers, so it lives on its own: stdlib only, no gazetteer, no models,
no imports from the rest of the package. That is what lets the property
tests generate annotations directly instead of standing up three
checkpoints to reach the same code.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
