### 2026-09-14 5302fbf
feat: run a bounded real-model Andorra pilot
Thirteen fixed sentences with hand-written gold spans, driven through the
real GLiNER2 recognizer and the Jina embedding/reranker resolver against
the 3,268-feature Andorra gazetteer fixture. The point is evidence that
the wiring works end to end, not a benchmark.

The run stays inside a laptop's budget: one recognition batch, GLiNER
released before Jina is constructed, a 128-token context bound for these
short documents, and float32 only for the CPU reranker.

Its output is not versioned. ADR 0003 keeps generated evidence out of
source control, and
[…]
