+++ scripts/pilot.py
@@ -1,0 +1,509 @@
+"""Run a real, reproducible Andorra geoparsing pilot and write its evidence."""
+
+from __future__ import annotations
+
+import argparse
+import gc
+import json
+import os
+import time
+import uuid
+from collections.abc import Mapping, Sequence
+from dataclasses import dataclass
+from pathlib import Path
+from typing import TYPE_CHECKING, Any
+
+if TYPE_CHECKING:
+    from geoparser.evaluation import Annotation
+
+Report = dict[str, Any]
+
+
+@dataclass(frozen=True, slots=True)
+class PilotSpan:
+    """A hand-written gold span for one pilot document."""
+
+    start: int
+    end: int
+    identifier: str
+
+
+@dataclass(frozen=True, slots=True)
+class PilotCase:
+    """One fixed pilot document and its gold Andorra annotations."""
+
+    case_id: str
+    text: str
+    gold: tuple[PilotSpan, ...]
+
+
+ANDORRA_ATTRIBUTE_MAP = {
+    "name": "name",
+    "type": "feature_name",
+    "level1": "country_name",
+    "level2": "admin1_name",
+    "level3": "admin2_name",
+}
+
+# The fixed pilot sentences are all far shorter than this. Bounding the
+# encoder window avoids paying for the checkpoint's 8K-token maximum on a
+# memory-constrained laptop while preserving every token in this fixture.
+MAX_SEQUENCE_LENGTH = 128
+
+PILOT_CASES = (
+    PilotCase(
+        "capital",
+        "Visitors gather in Andorra la Vella.",
+        (PilotSpan(19, 35, "3041563"),),
+    ),
+    PilotCase(
+        "encamp",
+        "Encamp welcomes hikers in every season.",
+        (PilotSpan(0, 6, "3040686"),),
+    ),
+    PilotCase(
+        "canillo",
+        "Canillo lies beneath the northern peaks.",
+        (PilotSpan(0, 7, "3041204"),),
+    ),
+    PilotCase(
+        "escaldes-engordany",
+        "Escaldes-Engordany hosts a thermal spa.",
+        (PilotSpan(0, 18, "3338529"),),
+    ),
+    PilotCase(
+        "ordino",
+        "Ordino keeps its mountain paths quiet.",
+        (PilotSpan(0, 6, "3039678"),),
+    ),
+    PilotCase(
+        "la-massana",
+        "La Massana connects the western valleys.",
+        (PilotSpan(0, 10, "3040132"),),
+    ),
+    PilotCase(
+        "sant-julia-de-loria",
+        "Sant Julià de Lòria marks the southern route.",
+        (PilotSpan(0, 19, "3039163"),),
+    ),
+    PilotCase(
+        "route",
+        "The road links Encamp with Canillo.",
+        (PilotSpan(15, 21, "3040686"), PilotSpan(27, 34, "3041204")),
+    ),
+    PilotCase(
+        "catalan",
+        "Andorra la Vella és a prop d'Escaldes-Engordany.",
+        (PilotSpan(0, 16, "3041563"), PilotSpan(29, 47, "3338529")),
+    ),
+    PilotCase(
+        "spanish",
+        "Canillo y Encamp están en Andorra.",
+        (
+            PilotSpan(0, 7, "3041204"),
+            PilotSpan(10, 16, "3040686"),
+            PilotSpan(26, 33, "3041565"),
+        ),
+    ),
+    PilotCase(
+        "french",
+        "À Encamp, les rues sont calmes.",
+        (PilotSpan(2, 8, "3040686"),),
+    ),
+    PilotCase(
+        "no-target-place",
+        "A river crosses the market before winter arrives.",
+        (),
+    ),
+    PilotCase(
[…]
