### 2026-09-14 a4844f1
fix: enforce module purity from the architecture CLI
PURE_MODULES and find_impure_modules existed, and a pytest test called the
helper, but main() never did. The quality gauntlet runs the script, so the
gate printed "Architecture checks passed" while a module on that list
carried a torch import.

Probed before and after with a deliberate impurity in geoparser.evaluation:
exit 0 and silence beforehand, exit 1 naming `geoparser.evaluation -> torch`
afterwards.

The two regression tests assert the CLI's exit code rather than the
helper's return value, because the helper was never the part that wa
[…]
