### 2026-09-14 4425e53
feat: let the quality gauntlet run without network access
Adds --offline: uv commands stop resolving, the dependency stage asserts a
lockfile exists rather than checking it is current, and the smoke build
reuses the provisioned backend.

Every flag was checked against uv 0.11.16 rather than assumed. The tests
assert command tuples, so a flag that does not exist would still pass them.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
