### 2026-09-14 a2aab90
docs: document the pilot boundary and how to read its numbers
ADR 0003 records why the pilot is fixed, small and resource-bounded.

The guide adds what the raw report cannot say for itself. Exact resolution
accuracy on the committed run is 0.067, which reads as a broken resolver
and is not one: GeoNames stores each Andorran parish twice, once as a
populated place and once as the administrative division of the same name,
and five of the seven mismatches are exactly that pair. Counting those as
correct gives 0.400. One prediction is genuinely wrong -- Andorra resolves
to a radio station.

Both n
[…]
