### 2026-09-15 c494f2c
fix: point public repository metadata at the canonical repository
The MkDocs repository metadata, the citation record, and several
documentation links still named the NoeFlandre fork. Those URLs resolve,
so the staleness survived review: the docs built, the pages rendered, and
the "Edit this page" button quietly sent readers to a fork.

Point repo_name, repo_url, CITATION.cff's repository-code, and the
remaining blob links at dguzh/geoparser, and add a guard that fails when
any tracked file names a non-canonical owner for this repository. The
guard checks links to this repository only; the dgu
[…]
