### 2026-09-15 53cf4b5
fix: declare the spaCy plugin transformer pipelines load from
SpacyRecognizer(model_name="en_core_web_trf") raised ValueError [E002]
while the sm, md and lg models worked. Transformer pipelines start with a
`curated_transformer` component, and that factory ships in
spacy-curated-transformers rather than in spaCy; the package never
declared it, so a fully synced environment could not load any of them.

Declaring it makes the documented path work. The plugin caps itself at
Python 3.13 while this project supports 3.14, so the requirement carries
a marker and the recognizer now names the plugin wh
[…]
