+++ pyproject.toml
@@ -31,6 +31,14 @@ rich = "^15.0.0"
 sentence-transformers = "^5.1.2"
 shapely = "^2.1.2"
 spacy = "^3.8.7,!=3.8.14" # 3.8.14 has cp314 wheels on GitHub but not on PyPI (explosion/spaCy#13949)
+# spaCy's transformer pipelines -- en_core_web_trf and its siblings, which the
+# module guide recommends -- are built on the `curated_transformer` factory, and
+# that factory ships in this plugin rather than in spaCy itself. Without it
+# spacy.load() raises ValueError [E002] at load time, not at import, and the
+# message blames a missing custom component (#128). The plugin publishes no
+# release for Python 3.14, so the marker keeps 3.14 installable; SpacyRecognizer
+# explains the gap there instead of failing obscurely.
+spacy-curated-transformers = {version = "^0.3.1", python = "<3.14"}
 sqlalchemy = "^2.0.44"
 sqlmodel = "^0.0.39"
 torch = "^2.9.0"
