### 2026-09-15 4b8e653
docs: state the transformer plugin requirement in every trf example
`en_core_web_trf` appears in four documentation pages and in the demo
notebook, but only ever as "the more accurate model". Its first component
comes from `spacy-curated-transformers`, which spaCy does not install with
itself, so a reader who copies any of those examples meets
`Can't find factory for 'curated_transformer'` with nothing in the docs to
explain it (#128).

The caveat is written once, in `docs/_snippets/spacy-transformer-requirements.rst`,
and pulled into each page with `.. include::`. Four copies of the same thre
[…]
