### 2026-09-15 a409808
docs: pin the transformer plugin to the line spaCy 3.8 can use
The caveat added in the previous commit told readers to run
`pip install spacy-curated-transformers`. Unpinned, that resolves to the
plugin's 2.x line: 2.0.0 requires `spacy>=4.0.0.dev2`, and 2.1.2 requires
`thinc>=9.0.0.dev4`, while geoparser holds spaCy at ^3.8.7 and thinc at 8.x.
In a Python 3.13 environment with spaCy 3.8.16 the command does not merely
mislead, it fails: pip tries to build thinc 9 from source and the
cythonization errors out. The advice was worse than none.

`en_core_web_trf` 3.8.0 declares its own requirement,
[…]
