+++ tests/unit/test_docs.py
@@ -2,89 +2,153 @@
 Unit tests for the documentation sources.
 
 The transformer spaCy pipeline needs a plugin that neither spaCy nor Geoparser pulls in
-on its own, so every public example that reaches for it has to carry the same
-prerequisite caveat. These tests keep that true for examples added later.
+on its own, and the plugin's current major line is built for a later spaCy than the one
+Geoparser uses. Every public example that reaches for the model therefore has to carry
+the same prerequisite, with the same version bound. These tests find those examples
+themselves — across every tracked file, whatever its format — so an example added later
+cannot quietly go unqualified.
 """
 
+import re
+import subprocess
 from pathlib import Path
 
 import pytest
 
 REPO_ROOT = Path(__file__).resolve().parents[2]
 DOCS_DIR = REPO_ROOT / "docs"
+GUARD = Path(__file__).resolve()
 
 TRANSFORMER_MODEL = "en_core_web_trf"
-PLUGIN = "spacy-curated-transformers"
 ALTERNATIVE_MODEL = "en_core_web_lg"
 UNSUPPORTED_PYTHON = "3.14"
+
+# The bound `en_core_web_trf` 3.8.0 itself declares. Unpinned, pip installs the 2.x line,
+# which targets spaCy 4 and a thinc release spaCy 3.8 cannot use.
+PLUGIN_REQUIREMENT = "spacy-curated-transformers>=0.2.2,<1.0.0"
+
+# An install command that names the plugin without immediately constraining it.
+UNPINNED_INSTALL = re.compile(
+    r"""install\s+["']?spacy-curated-transformers(?![><=~!])"""
+)
 
 # The single place the caveat is written, and the line every page pulls it in with.
 CAVEAT_SNIPPET = DOCS_DIR / "_snippets" / "spacy-transformer-requirements.rst"
 CAVEAT_INCLUDE = ".. include:: /_snippets/spacy-transformer-requirements.rst"
 
-# Documentation that is not reStructuredText and so cannot use the shared include. It is
-# held to naming the plugin instead.
-OTHER_DOC_SOURCES = [REPO_ROOT / "demo" / "demo.ipynb"]
+
+def _tracked_files() -> list[Path]:
+    """Every file git tracks, which is every source and no generated artifact."""
+    try:
+        listing = subprocess.run(
+            ["git", "ls-files", "-z"],
+            cwd=REPO_ROOT,
+            capture_output=True,
+            check=True,
+            text=True,
+        ).stdout
+    except (OSError, subprocess.CalledProcessError) as error:  # pragma: no cover
+        pytest.skip(f"not a git checkout, cannot enumerate sources: {error}")
+    return [REPO_ROOT / name for name in listing.split("\0") if name]
 
 
 def _read(path: Path) -> str:
-    """Read a documentation source as text."""
-    return path.read_text(encoding="utf-8")
+    """Read a source as text, or return an empty string if it is not text at all."""
+    try:
+        return path.read_text(encoding="utf-8")
+    except (UnicodeDecodeError, OSError):
+        return ""
 
 
-def _rst_pages_mentioning_transformer_model() -> list[Path]:
-    """Every documentation page that presents the transformer model, snippet aside."""
+def _sources_mentioning_transformer_model() -> list[Path]:
+    """Every tracked source that presents the transformer model, this guard aside."""
     return sorted(
         path
-        for path in DOCS_DIR.rglob("*.rst")
-        if path != CAVEAT_SNIPPET and TRANSFORMER_MODEL in _read(path)
+        for path in _tracked_files()
+        if path.resolve() != GUARD and TRANSFORMER_MODEL in _read(path)
     )
+
+
+def _names(paths: list[Path]) -> list[str]:
[…]
