+++ demo/Dockerfile
@@ -9,8 +9,13 @@ RUN apt-get update && apt-get install -y \
     curl \
     && rm -rf /var/lib/apt/lists/*
 
-# Install geoparser and demo dependencies
-RUN pip install --no-cache-dir geoparser jupyter plotly
+# Install geoparser and demo dependencies. The notebook uses en_core_web_trf, whose
+# first component comes from the curated_transformer factory in spacy-curated-transformers
+# rather than from spaCy itself; without it the model cannot be loaded. The bound is the
+# one the model declares — unpinned, pip installs the 2.x line, which is built for a
+# later spaCy generation than the one geoparser uses.
+RUN pip install --no-cache-dir geoparser jupyter plotly \
+    "spacy-curated-transformers>=0.2.2,<1.0.0"
 
 # Download spaCy models
 RUN python -m spacy download en_core_web_sm && \
