### 2026-09-25 011cb84
Add spacy-curated-transformers so trf models load on first use
