### 2026-09-25 647cb55
Bump version to 0.6.1 in pyproject.toml
